% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{add_to_extract.ipumsi_extract}
\alias{add_to_extract.ipumsi_extract}
\title{Add values to an existing IPUMS International extract definition}
\usage{
\method{add_to_extract}{ipumsi_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{description}{Description of the extract.}

\item{samples}{Character vector of sample names to include in the extract.
Samples should be specified using IPUMS International
\href{https://international.ipums.org/international-action/samples/sample_ids}{sample ID values}.}

\item{variables}{Character vector of variable names or a list of
\code{ipums_variable} objects created by \code{\link[=new_variable]{new_variable()}}
containing specifications for all variables to include in the extract.

If a variable already exists in the extract, its specifications
will be added to those that already exist for that variable.}

\item{data_format}{Format for the output extract data file. One of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}.}

\item{...}{Ignored}
}
\value{
A modified \code{ipumsi_extract} object
}
\description{
Add new values or replace existing values in an IPUMS International extract
definition. All fields are optional, and if omitted, will be unchanged.
Supplying a value for fields that take a single value, such as
\code{description} and \code{data_format}, will replace the existing value with the
supplied value.

To remove existing values from an IPUMS International extract definition, use
\code{\link[=remove_from_extract.ipumsi_extract]{remove_from_extract()}}. To add values
contained in another extract definition, use \code{\link[=combine_extracts]{combine_extracts()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.

To modify variable-specific parameters for variables that already exist
in the extract, create a new variable specification with \code{\link[=new_variable]{new_variable()}}.
}
\examples{
extract <- define_extract_ipumsi(
  description = "Example IPUMS-I extract definition",
  samples = "eg2006a",
  variables = c("AGE", "SEX", "YEAR")
)

add_to_extract(extract, samples = "bw2011a")

extract2 <- add_to_extract(
  extract,
  samples = "bw2011a",
  variables = c("MARST", "RELATE")
)

# Add new variable-specific specifications as you would when defining
# an extract, even to modify variables that already exist in the definition
add_to_extract(
  extract,
  samples = "bw2011a",
  variables = list(
    new_variable("MARST", attached_characteristics = "father"),
    new_variable("SEX", case_selections = "2")
  )
)

# Values that only take a single value are replaced
add_to_extract(extract, description = "New description")$description

# You can also combine two separate extract requests together
combine_extracts(extract, extract2)
}
\seealso{
\code{\link[=remove_from_extract.ipumsi_extract]{remove_from_extract()}} to remove
values from an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_ipumsi]{define_extract_ipumsi()}} to create a new extract definition.
}
\keyword{internal}
