% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{add_to_extract}
\alias{add_to_extract}
\title{Add values to an IPUMS extract}
\usage{
add_to_extract(extract, ...)
}
\arguments{
\item{extract}{An object inheriting from \code{ipums_extract}}

\item{...}{Additional arguments specifying the extract fields and values to
add or modify. The available arguments correspond to the available
arguments in the extract definition function for the class of the extract
specified in \code{extract}.}
}
\value{

}
\description{
Add or modify values for specific fields in an existing extract. This
function is an S3 generic whose behavior will depend on the class (i.e.
collection) of the extract being modified.

In general, for a given collection, the available arguments to this
function are identical to those used when defining an extract for that
collection. For documentation on defining an extract, see
\code{\link{define_extract}}.

For collection-specific documentation, see:

\itemize{
  \item{\code{\link{add_to_extract.nhgis_extract}}}
  \item{\code{\link{add_to_extract.usa_extract}}}
}
}
