% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{add_to_extract.usa_extract}
\alias{add_to_extract.usa_extract}
\title{Add values to an existing USA extract}
\usage{
\method{add_to_extract}{usa_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  data_structure = NULL,
  rectangular_on = NULL,
  validate = TRUE,
  ...
)
}
\arguments{
\item{extract}{A \code{usa_extract} object. This can be a user-defined
extract (see \code{\link{define_extract_micro}()}) or an extract object
returned from another \code{ipumsr} API function.}

\item{samples}{Character vector of samples to add to the extract, if any.}

\item{variables}{Character vector of variables to add to the extract, if any.}

\item{validate}{Logical value indicating whether to check the modified
extract structure for validity. Defaults to \code{TRUE}}

\item{...}{Ignored}
}
\value{
A modified \code{usa_extract} object
}
\description{
Add new values to any extract fields of a USA extract.

To remove existing values from a USA extract, see
\code{\link[=remove_from_extract.usa_extract]{remove_from_extract}()}.
}
\examples{
my_extract <- define_extract_micro("usa", "Example", "us2013a", "YEAR")

add_to_extract(
  my_extract,
  description = "Revised extract",
  samples = "us2014a"
)
}
