% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{add_to_extract_micro}
\alias{add_to_extract_micro}
\alias{add_to_extract.micro_extract}
\title{Add values to an existing extract definition for an IPUMS microdata
collection}
\usage{
\method{add_to_extract}{micro_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  data_structure = NULL,
  rectangular_on = NULL,
  case_select_who = NULL,
  data_quality_flags = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{description}{Description of the extract.}

\item{samples}{Character vector of sample names to include in the extract.
Use \code{\link[=get_sample_info]{get_sample_info()}} to identify sample IDs for a given collection.}

\item{variables}{Character vector of variable names or a list of
\code{var_spec} objects created by \code{\link[=var_spec]{var_spec()}}
containing specifications for all variables to include in the extract.

If a variable already exists in the extract, its specifications
will be added to those that already exist for that variable.}

\item{data_format}{Format for the output extract data file. One of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}.}

\item{data_structure}{Data structure for the output the extract data.

\code{"rectangular"} provides person records with all requested household
information attached to respective household members.

\code{"hierarchical"} provides household records followed by person records.}

\item{rectangular_on}{If \code{data_structure} is \code{"rectangular"},
records on which to rectangularize. Currently only \code{"P"} (person records)
is supported.

Defaults to \code{"P"} if \code{data_structure} is \code{"rectangular"} and \code{NULL}
otherwise.}

\item{case_select_who}{Indication of how to interpret any case selections
included for variables in the extract.

\code{"individuals"} includes records for all individuals who match the
specified case selections.

\code{"households"} includes records for all members of each household that
contains an individual who matches the specified case selections.

Defaults to \code{"individuals"}. Use \code{\link[=var_spec]{var_spec()}} to add case selections for
specific variables.}

\item{data_quality_flags}{Set to \code{TRUE} to include data quality
flags for all applicable variables in the extract definition. This will
override the \code{data_quality_flags} specification for individual variables
in the definition.

Use \code{\link[=var_spec]{var_spec()}} to add data quality flags for specific variables.}

\item{...}{Ignored}
}
\value{
A modified \code{micro_extract} object
}
\description{
Add new values or replace existing values in an IPUMS microdata extract
definition.

All fields are optional, and if omitted, will be unchanged. Supplying a value
for fields that take a single value, such as \code{description} and \code{data_format},
will replace the existing value with the supplied value.

To remove existing values from an IPUMS microdata extract definition, use
\code{\link[=remove_from_extract.micro_extract]{remove_from_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.

To modify variable-specific parameters for variables that already exist
in the extract, create a new variable specification with \code{\link[=var_spec]{var_spec()}}.
}
\examples{
extract <- define_extract_usa(
  description = "2013 ACS Data",
  samples = "us2013a",
  variables = c("SEX", "AGE", "YEAR")
)

add_to_extract(extract, samples = "us2014a")

extract2 <- add_to_extract(
  extract,
  samples = "us2014a",
  variables = c("MARST", "BIRTHYR")
)

# Add new variable-specific specifications as you would when defining
# an extract, even to modify variables that already exist in the definition
add_to_extract(
  extract,
  samples = "us2014a",
  variables = list(
    var_spec("RACE", attached_characteristics = "mother"),
    var_spec("SEX", case_selections = "2")
  )
)

# Values that only take a single value are replaced
add_to_extract(extract, description = "New description")$description
}
\seealso{
\code{\link[=remove_from_extract.micro_extract]{remove_from_extract()}} to remove
values from an extract definition.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract]{define_extract_*()}} to create a new extract definition
from scratch
}
\keyword{internal}
