% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{add_value_labels}
\alias{add_value_labels}
\title{Add value labels to a dataset}
\usage{
add_value_labels(data, vnames, values)
}
\arguments{
\item{data}{The dataset to convert}

\item{vnames}{A vector of variable names}

\item{values}{A vector of nested data.frames with value and label columns
that describes what values should be assigned which label.
Currently is_missing_value column is ignored.}
}
\description{
Takes a dataset and a format library (eg from a variables control file)
and adds the value labels to the dataset. Because R doesn't handle labelled
values in the same way as other statistical packages, there are several options
for the format in which the value labels are stored. See Details for more information.
}
