% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{add_variable_labels}
\alias{add_variable_labels}
\title{Add variable labels to a dataset}
\usage{
add_variable_labels(data, vnames, vlabels)
}
\arguments{
\item{data}{The dataset to convert}

\item{vnames}{The names of the variables to be labelled}

\item{vlabels}{The labels to be added}
}
\description{
Takes a dataset and a format library (eg from a variables control file)
and adds the variable labels to the dataset. R isn't great about hanging onto
this kind of metainformation, so other functions may accidentally delete it.
}
