% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_metadata.R
\name{catalog_types}
\alias{catalog_types}
\title{List valid metadata types for an IPUMS data collection}
\usage{
catalog_types(collection)
}
\arguments{
\item{collection}{Character string indicating the IPUMS collection for which
to retrieve metadata.}
}
\value{
Character vector of valid endpoints
}
\description{
List the available metadata catalog endpoints for a particular IPUMS data
collection. For use with \code{\link[=get_metadata_catalog]{get_metadata_catalog()}}.
}
\examples{
catalog_types("nhgis")
}
\keyword{internal}
