% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{combine_extracts}
\alias{combine_extracts}
\title{Combine IPUMS extracts into a single extract definition}
\usage{
combine_extracts(...)
}
\arguments{
\item{...}{Arbitrary number of \code{ipums_extract} objects to be combined. All
extracts must belong to the same collection.}
}
\value{
An \code{ipums_extract} object of the same collection as the extracts
provided in \code{...} containing the combined extract definition.
}
\description{
Creates a single extract that includes all of the specifications included in
a set of \code{ipums_extract} objects.
}
\details{
Values that exist in more than one of the provided extracts will be
de-duplicated such that only one entry for the duplicated specification is
included in the final extract.

Some extract fields can only take a single value. In the event that the
extracts being combined each have different values for such a field, the
value found in the first extract provided to \code{...} is retained in the final
extract.

To add or remove values in extract fields, see \link{add_to_extract} or
\link{remove_from_extract}.
}
\examples{
ext1 <- define_extract_nhgis(
  datasets = "2011_2015_ACS5a",
  data_tables = "B00001",
  geog_levels = "county"
)

ext2 <- define_extract_nhgis(
  datasets = c("2011_2015_ACS5a", "2012_2016_ACS5a"),
  data_tables = c("B00002", "B01001"),
  geog_levels = "county"
)

combine_extracts(ext1, ext2)
}
