% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract_micro}
\alias{define_extract_micro}
\alias{define_extract_usa}
\alias{define_extract_cps}
\alias{define_extract_ipumsi}
\title{Define an extract request for an IPUMS microdata collection}
\usage{
define_extract_usa(
  description,
  samples,
  variables,
  data_format = "fixed_width",
  data_structure = "rectangular",
  rectangular_on = NULL
)

define_extract_cps(
  description,
  samples,
  variables,
  data_format = "fixed_width",
  data_structure = "rectangular",
  rectangular_on = NULL
)

define_extract_ipumsi(
  description,
  samples,
  variables,
  data_format = "fixed_width",
  data_structure = "rectangular",
  rectangular_on = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{samples}{Character vector of sample names to include in the extract.
Use \code{\link[=get_sample_info]{get_sample_info()}} to identify sample IDs for a given collection.}

\item{variables}{Character vector of variable names or a list of detailed
variable specifications for all variables in the extract request. Use
\code{\link[=var_spec]{var_spec()}} to create a \code{var_spec} object containing a detailed variable
specification. See examples.}

\item{data_format}{Format for the output extract data file. One of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}. Defaults to
\code{"fixed_width"}.}

\item{data_structure}{Data structure for the output the extract data.

\code{"rectangular"} provides person records with all requested household
information attached to respective household members.

\code{"hierarchical"} provides household records followed by person records.}

\item{rectangular_on}{If \code{data_structure} is \code{"rectangular"},
records on which to rectangularize. Currently only \code{"P"} (person records)
is supported.

Defaults to \code{"P"} if \code{data_structure} is \code{"rectangular"} and \code{NULL}
otherwise.}
}
\value{
An object of class \code{\link[=ipums_extract-class]{micro_extract}} containing
the extract definition.
}
\description{
Define the parameters of an IPUMS microdata extract request to be submitted
via the IPUMS API.

Currently supported microdata collections include:
\itemize{
\item \strong{IPUMS USA}: \code{define_extract_usa()}
\item \strong{IPUMS CPS}: \code{define_extract_cps()}
\item \strong{IPUMS International}: \code{define_extract_ipumsi()}
}

Use the \code{\link[=var_spec]{var_spec()}} helper to generate detailed variable-level
specifications for an extract request.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
usa_extract <- define_extract_usa(
  description = "2013-2014 ACS Data",
  samples = c("us2013a", "us2014a"),
  variables = c("SEX", "AGE", "YEAR")
)

usa_extract

# Use `var_spec()` to created detailed variable specifications:
usa_extract <- define_extract_usa(
  description = "Example USA extract definition",
  samples = c("us2013a", "us2014a"),
  variables = var_spec(
    "SEX",
    case_selections = "2",
    attached_characteristics = c("mother", "father")
  )
)

# For multiple variables, provide a list of `var_spec` objects and/or
# variable names.
cps_extract <- define_extract_cps(
  description = "Example CPS extract definition",
  samples = c("cps2020_02s", "cps2020_03s"),
  variables = list(
    var_spec("AGE", data_quality_flags = TRUE),
    var_spec("SEX", case_selections = "2"),
    "RACE"
  )
)

cps_extract

# To recycle specifications to many variables, it may be useful to
# create variables prior to defining the extract:
var_names <- c("AGE", "SEX")

var_spec <- purrr::map(
  var_names,
  ~ var_spec(.x, attached_characteristics = "mother")
)

ipumsi_extract <- define_extract_ipumsi(
  description = "Extract definition with predefined variables",
  samples = c("br2010a", "cl2017a"),
  variables = var_spec
)

# Extract specifications can be indexed by name
names(ipumsi_extract$samples)

names(ipumsi_extract$variables)

ipumsi_extract$variables$AGE

\dontrun{
# Use the extract definition to submit an extract request to the API
submit_extract(usa_extract)
}
}
\seealso{
\code{\link[=submit_extract]{submit_extract()}}, \code{\link[=download_extract]{download_extract()}}, and \code{\link[=get_extract_info]{get_extract_info()}} to
process and manage an extract request.

\code{\link[=save_extract_as_json]{save_extract_as_json()}} and \code{\link[=define_extract_from_json]{define_extract_from_json()}} to share an
extract definition.

\code{\link[=add_to_extract.micro_extract]{add_to_extract()}},
\code{\link[=remove_from_extract.micro_extract]{remove_from_extract()}}, and
\code{\link[=combine_extracts]{combine_extracts()}} to revise an extract definition.
}
