% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract_usa}
\alias{define_extract_usa}
\title{Define an IPUMS USA extract request}
\usage{
define_extract_usa(
  description,
  samples,
  variables,
  data_format = c("fixed_width", "csv", "stata", "spss", "sas9"),
  data_structure = "rectangular",
  rectangular_on = "P"
)
}
\arguments{
\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to include in the extract. Samples
should be specified using IPUMS USA
\href{https://usa.ipums.org/usa-action/samples/sample_ids}{sample ID values}.}

\item{variables}{Character vector of variables to include in the extract.}

\item{data_format}{The desired format of the extract data file. One of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}.}

\item{data_structure}{The desired structure of the extract data.

\code{"rectangular"} provides person records with all requested household
information attached to respective household members.

\code{"hierarchical"} provides household records followed by person records.

Currently, only \code{"rectangular"} extracts are supported.}

\item{rectangular_on}{Records on which to rectangularize. \code{"P"}
rectangularizes on person records. \code{"H"} provides household-only records.

Currently, only \code{"P"} is supported.}
}
\value{
An object of class \code{\link[=ipums_extract-class]{usa_extract}} containing
the extract definition.
}
\description{
Define an IPUMS USA extract request to be submitted via the IPUMS
API. An extract request contains the specifications required
to identify and obtain a particular set of data from the IPUMS USA system.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
my_extract <- define_extract_usa(
  description = "2013-2014 ACS Data",
  samples = c("us2013a", "us2014a"),
  variables = c("SEX", "AGE", "YEAR")
)

my_extract

\dontrun{
# Use the extract definition to submit an extract request to the API
submit_extract(my_extract)
}
}
\seealso{
\code{\link[=submit_extract]{submit_extract()}}, \code{\link[=download_extract]{download_extract()}}, and \code{\link[=get_extract_info]{get_extract_info()}} to
process and manage an extract request.

\code{\link[=save_extract_as_json]{save_extract_as_json()}} and \code{\link[=define_extract_from_json]{define_extract_from_json()}} to share an
extract definition.

\code{\link[=add_to_extract.usa_extract]{add_to_extract()}},
\code{\link[=remove_from_extract.usa_extract]{remove_from_extract()}}, and
\code{\link[=combine_extracts]{combine_extracts()}} to
revise an extract definition.
}
