% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_extract}
\alias{download_extract}
\title{Download a completed IPUMS data extract}
\usage{
download_extract(
  extract,
  download_dir = getwd(),
  overwrite = FALSE,
  wait = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY"),
  ...
)
}
\arguments{
\item{extract}{One of:
\itemize{
\item An \code{\link[=ipums_extract-class]{ipums_extract}} object
\item The data collection and extract number formatted as a string of the
form \code{"collection:number"} or as a vector of the form
\code{c("collection", number)}
\item The extract number for your default IPUMS collection. See
\code{\link[=set_ipums_default_collection]{set_ipums_default_collection()}}
}

Extract numbers do not need to be zero-padded. That is, use \code{1}, not
\code{"0001"}.

For a list of codes used to refer to each collection, see
\code{\link[=ipums_data_collections]{ipums_data_collections()}}.}

\item{download_dir}{Path to the directory where the files should be written.
Defaults to current working directory.}

\item{overwrite}{Logical indicating whether to overwrite any files that
already exist in \code{download_dir}. Defaults to \code{FALSE}.}

\item{wait}{Logical value indicating whether to wait for the completion of
the specified \code{extract} if it is not yet ready for download. Defaults to
\code{FALSE}. See \code{\link[=wait_for_extract]{wait_for_extract()}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable \code{"IPUMS_API_KEY"}.}

\item{...}{Further arguments passed to \code{\link[=wait_for_extract]{wait_for_extract()}}. Only used if
\code{wait = TRUE}.}
}
\value{
The path(s) to the files required to read the data
requested in the extract, invisibly.

For NHGIS, paths will be named with either \code{"data"} (for tabular data
files) or \code{"shape"} (for spatial data files) to
indicate the type of data the file contains.
}
\description{
Download an IPUMS data extract via the IPUMS API and write to disk.

For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
For NHGIS extracts, data files and GIS files (shapefiles) will be saved in
separate .zip archives. \code{download_extract()} will return a character vector
including the file paths to all downloaded files.

For microdata extracts, only the file path to the downloaded .xml DDI file
will be returned, as it is sufficient for reading the data provided in the
associated .gz data file.
}
\examples{
usa_extract <- define_extract_usa(
  description = "2013-2014 ACS Data",
  samples = c("us2013a", "us2014a"),
  variables = c("SEX", "AGE", "YEAR")
)

nhgis_extract <- define_extract_nhgis(
  description = "Example NHGIS extract",
  datasets = "1990_STF3",
  data_tables = "NP57",
  geog_levels = c("county", "tract")
)

\dontrun{
submitted_extract <- submit_extract(usa_extract)

# For microdata, the path to the DDI .xml codebook file is provided.
usa_xml_file <- download_extract(submitted_extract)

# Load with a `read_ipums_micro_*()` function
usa_data <- read_ipums_micro(usa_xml_file)

# You can also download previous extracts with their collection and number:
nhgis_files <- download_extract("nhgis:1")

# NHGIS extracts return a path to both the tabular and spatial data files,
# as applicable.
#
# Load NHGIS tabular data
nhgis_data <- read_nhgis(data = nhgis_files["data"])

# Load NHGIS spatial data
nhgis_geog <- read_ipums_sf(data = nhgis_files["shape"])

# If you have a default collection, you can use the extract number alone:
set_ipums_default_collection("usa")
download_extract(1)
}
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{set_ipums_api_key}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
