% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{dplyr_select_helpers}
\alias{dplyr_select_helpers}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{starts_with}
\alias{one_of}
\alias{everything}
\alias{current_vars}
\title{Select helpers from dplyr}
\description{
These are functions that are helpful for selecting variables
within dplyr's \code{\link[dplyr]{select}} functions. See
\code{\link[dplyr]{select_helpers}} for more details.
}
