% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{ds_spec}
\alias{ds_spec}
\alias{tst_spec}
\title{Create dataset and time series table specifications for IPUMS NHGIS extract
requests}
\usage{
ds_spec(name, data_tables, geog_levels, years = NULL, breakdown_values = NULL)

tst_spec(name, geog_levels, years = NULL)
}
\arguments{
\item{name}{Name of the dataset or time series table.}

\item{data_tables}{Vector of summary tables to retrieve for the given
dataset.}

\item{geog_levels}{Geographic levels (e.g. \code{"county"} or \code{"state"})
at which to obtain data for the given dataset or time series table.}

\item{years}{Years for which to obtain the data for the given dataset or time
series table.

For time series tables, all years are selected by default.

For datasets, use \code{"*"} to select all available years. Use
\code{\link[=get_metadata_nhgis]{get_metadata_nhgis()}} to determine if a dataset allows year selection.}

\item{breakdown_values}{\href{https://www.nhgis.org/frequently-asked-questions-faq#breakdowns}{Breakdown values}
to apply to the given dataset.}
}
\value{
A \code{ds_spec} or \code{tst_spec} object.
}
\description{
Provide specifications for individual datasets and time series
tables when defining an IPUMS NHGIS extract request.

Use \code{\link[=get_metadata_nhgis]{get_metadata_nhgis()}} to browse dataset and time series
table specification parameters.
}
\examples{
dataset <- ds_spec(
  "2013_2017_ACS5a",
  data_tables = c("B00001", "B01002"),
  geog_levels = "state"
)

tst <- tst_spec(
  "CW5",
  geog_levels = c("county", "tract"),
  years = "1990"
)

# Use variable specifications in an extract definition:
define_extract_nhgis(
  description = "Example extract",
  datasets = dataset,
  time_series_tables = tst
)
}
\keyword{internal}
