% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{extract_revisions}
\alias{extract_revisions}
\alias{add_to_extract}
\alias{remove_from_extract}
\title{Modify values in an existing extract definition}
\usage{
add_to_extract(extract, ...)

remove_from_extract(extract, ...)
}
\arguments{
\item{extract}{An object inheriting from \code{ipums_extract}}

\item{...}{Additional arguments specifying the extract fields and values to
modify. The available arguments correspond to the available arguments in
the extract definition function for the class of the extract specified in
\code{extract}.}
}
\value{
An object of the same class as \code{extract} containing the modified
  extract definition
}
\description{
Add or remove values for specific fields in an existing extract. These
functions are S3 generics whose behavior will depend on the class (i.e.
collection) of the extract being modified. Documentation for specific
collections can be found under their associated methods. For example:
\code{\link{add_to_extract.nhgis_extract}}.
}
\seealso{
\code{\link{add_to_extract.nhgis_extract}},
  \code{\link{add_to_extract.usa_extract}}
  \code{\link{remove_from_extract.nhgis_extract}},
  \code{\link{remove_from_extract.usa_extract}}
}
