% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{extract_tbl_to_list}
\alias{extract_tbl_to_list}
\title{Convert a tibble of extract definitions to a list}
\usage{
extract_tbl_to_list(extract_tbl, validate = TRUE)
}
\arguments{
\item{extract_tbl}{A \code{\link[tibble:tbl_df-class]{tibble}} (or
\code{\link[base:data.frame]{data.frame}}) that contains
the specifications for one or more \code{\link[=ipums_extract-class]{ipums_extract}}
objects.}

\item{validate}{Logical value indicating whether to
check that each of the output \code{ipums_extract}
objects contains a valid and complete extract
definition. Defaults to \code{TRUE}.}
}
\value{
A list of length equal to the number of extracts represented in
\code{extract_tbl}. Unique extracts can be identified by their extract
number, which is contained in the \code{number} column of \code{extract_tbl}.
}
\description{
Convert a \code{\link[tibble:tbl_df-class]{tibble}} containing the specifications for
one or more extract definitions to a list of
\code{\link[=ipums_extract-class]{ipums_extract}} objects.

Use \code{\link[=get_extract_info]{get_extract_info()}} with \code{table = TRUE} to obtain extract
information in \code{tibble} format.

For an overview of ipumsr API
functionality, see \code{vignette("ipums-api", package = "ipumsr")}.
}
\examples{
\dontrun{
# Get tibble of recent extracts
tbl_of_last_10_extracts <- get_extract_info("usa", table = TRUE)

# Filter down to extracts with "income" in the description
description_mentions_income <- grepl(
  "[Ii]ncome",
  tbl_of_last_10_extracts$description
)
income_extracts <- tbl_of_last_10_extracts[description_mentions_income, ]

# Convert tibble of extracts to list of extracts
income_extracts <- extract_tbl_to_list(income_extracts)

# Now it's easier to operate on those elements as extract objects:
revised_income_extract <- add_to_extract(
  income_extracts[[1]],
  samples = "us2018a"
)

submitted_revised_income_extract <- submit_extract(revised_income_extract)
}

}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{get_extract_info}()},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
