% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helpers.R
\name{get_extract_history}
\alias{get_extract_history}
\title{Helper function to retrieve extract history for a specified
IPUMS data collection}
\usage{
get_extract_history(collection)
}
\arguments{
\item{collection}{The IPUMS data collection for the extract.}
}
\value{
A table of extract definitions
}
\description{
Helper function to retrieve extract history for a specified
IPUMS data collection
}
\examples{
my_extracts <- get_extract_history("cps")

\dontrun{
# Retrieve your extract history as a data frame
my_extracts <- get_extract_history("cps")

# Now you can filter for extract contents/features
statefip_extracts <- filter(my_extracts, str_detect(variables, "STATEFIP"))
}
}
