% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.r
\name{ip_list_files}
\alias{ip_list_files}
\alias{ip_list_data_files}
\alias{ip_list_shape_files}
\alias{ip_list_raster_files}
\title{List files available for analysis in an IPUMS extract zip file}
\usage{
ip_list_files(file, types = NULL, data_layer = NULL, shape_layer = NULL,
  raster_layer = NULL)

ip_list_data_files(file, data_layer = NULL)

ip_list_shape_files(file, shape_layer = NULL)

ip_list_raster_files(file, raster_layer = NULL)
}
\arguments{
\item{file}{An IPUMS extract zip file}

\item{types}{One or more of "data", "shape", or "raster" indicating
what type of files to look for.}

\item{data_layer}{A regex filter for data files to look for}

\item{shape_layer}{A regex filter for shape files to look for}

\item{raster_layer}{A regex filter for raster files to look for}
}
\description{
Find which files can be loaded from an IPUMS extract zip file.
}
