% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_read.r
\name{ip_read_data}
\alias{ip_read_data}
\title{Read data from an IPUMS extract}
\usage{
ip_read_data(ddi, vars = NULL, n_max = Inf, data_file = NULL,
  verbose = TRUE)
}
\arguments{
\item{ddi}{Either a filepath to a DDI xml file downloaded from
the website, or a ip_ddi object parsed by \code{\link{ip_read_ddi}}}

\item{vars}{A vector of variable names using \code{\link[dplyr]{select}}-style
convetions. For hierarchical data, the rectype id variable will be added even
if it is not specified.}

\item{n_max}{The maximum number of records to load.}

\item{data_file}{Specify a directory to look for the data file.
If left empty, it will look in the same directory as the DDI file.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\description{
Reads a dataset downloaded from the IPUMS extract system.
}
\details{
For IPUMS projects with microdata, it relies on a downloaded
DDI codebook and a fixed-width file. Loads the data with
value labels (using \code{\link[haven]{labelled}} format)
and variable labels.
}
\examples{
\dontrun{
data <- ip_read_data("cps_00001.xml")
}
}
\seealso{
ip_read_nhgis ip_read_terra_raster ip_read_terra_area
}
