% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_read.r
\name{ip_read_terra_area}
\alias{ip_read_terra_area}
\title{Read data from an IPUMS Terra area extract}
\usage{
ip_read_terra_area(data_file, ddi_file = NULL, shape_file = NULL,
  data_layer = NULL, shape_layer = NULL, verbose = TRUE)
}
\arguments{
\item{data_file}{Path to the data file, which can either be the .zip file directly
downloaded from the IPUMS Terra website, or to the csv unzipped from the download.}

\item{ddi_file}{(Optional) If the download is unzipped, path to the .xml file which
provides usage and ciation information for extract.}

\item{shape_file}{(Optional) If the download is unzipped, path to the .zip or .shp file
representing the the shape file. If only the data table is needed, can be set to FALSE
to indicate not to load the shape file.}

\item{data_layer}{If data_file is an extract with multiple csvs, a regular expression indicating
which .csv data layer to load.}

\item{shape_layer}{If data_file is an extract with multiple shape files, a regular expression
indicating which shape layer to load.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\value{
A \code{\link[raster]{raster}} object
}
\description{
Reads a area-level dataset downloaded from the IPUMS Terra extract system.
}
\examples{
\dontrun{
data <- ip_read_terra_area("2552_bundle.zip", "LCDECIDOPZM2013.tiff")
}
}
\seealso{
ip_read_terra_raster ip_read_data ip_read_nhgis
}
