% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_api_download_extract}
\alias{ipums_api_download_extract}
\alias{ipums_api_download_extract.ipums_extract}
\alias{ipums_api_download_extract.data.frame}
\alias{ipums_api_download_extract.character}
\title{Download an extract}
\usage{
ipums_api_download_extract(object, ...)

\method{ipums_api_download_extract}{ipums_extract}(
  extract,
  download_dir,
  overwrite = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

\method{ipums_api_download_extract}{data.frame}(
  extract_df,
  download_dir,
  overwrite = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

\method{ipums_api_download_extract}{character}(
  collection,
  extract_number,
  download_dir,
  overwrite = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{object}{Either an object of class \code{ipums_extract}, or the name of
an IPUMS data collection as a length one character vector.}

\item{extract}{An extract object created by \code{\link{ipums_api_build_extract}}.}

\item{download_dir}{In what folder should the downloaded files be saved?}

\item{overwrite}{Logical indicating whether to overwrite files that already
exist. Defaults to \code{FALSE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}

\item{extract_df}{A data.frame in which each row defines one extract.}

\item{collection}{The IPUMS data collection for the extract.}

\item{extract_number}{The extract number.}
}
\value{
A character vector of paths to the downloaded .xml DDI files. If the
  first argument is an \code{ipums_extract} or data collection, the character
  vector will have length one. If the first argument is a \code{data.frame},
  the character vector will have length equal to the number of rows of the
  \code{data.frame}.
}
\description{
Download an extract
}
\section{Methods (by class)}{
\itemize{
\item \code{ipums_extract}: Download files associated with an
\code{ipums_extract} object

\item \code{data.frame}: Download files associated with IPUMS
extract definitions stored in rows of a data.frame.

\item \code{character}: Download files associated with an
extract, given a data collection and extract number
}}

