% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_api_extract_downloadable}
\alias{ipums_api_extract_downloadable}
\alias{ipums_api_extract_downloadable.ipums_extract}
\alias{ipums_api_extract_downloadable.character}
\alias{ipums_api_extract_downloadable.data.frame}
\title{Is the extract ready to download?}
\usage{
ipums_api_extract_downloadable(object, ...)

\method{ipums_api_extract_downloadable}{ipums_extract}(extract, api_key = Sys.getenv("IPUMS_API_KEY"))

\method{ipums_api_extract_downloadable}{character}(
  collection,
  extract_number,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

\method{ipums_api_extract_downloadable}{data.frame}(extract_df)
}
\arguments{
\item{object}{Either an object of class \code{ipums_extract}, the name of an IPUMS
data collection as a length one character vector, or a data.frame with an
extract definition in each row.}

\item{extract}{An extract object created by \code{\link{ipums_api_build_extract}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}

\item{collection}{The IPUMS data collection for the extract.}

\item{extract_number}{The extract number.}

\item{extract_df}{A data.frame in which each row contains the definition of
an IPUMS extract.}
}
\value{
A logical vector. If the first argument is an object of class
  \code{ipums_extract} or the name of a data collection, the function will
  return a length-one logical vector. If the first
  argument is a \code{data.frame}, the function will return a logical vector
  with length equal to the number of rows of the \code{data.frame}.
}
\description{
This function checks whether the given extract is ready to download,
returning TRUE for extracts that are ready and FALSE for those that are not.
}
\section{Methods (by class)}{
\itemize{
\item \code{ipums_extract}: Is the extract defined by the
given \code{ipums_extract} object ready to download?

\item \code{character}: Is the extract defined by the
given data collection and extract number ready to download?

\item \code{data.frame}: Are the extracts defined in the
rows of the given data.frame ready to download?
}}

