% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_api_get_extract}
\alias{ipums_api_get_extract}
\alias{ipums_api_get_extract.ipums_extract}
\alias{ipums_api_get_extract.character}
\alias{ipums_api_get_extract.data.frame}
\title{Get information about a submitted extract}
\usage{
ipums_api_get_extract(object, ...)

\method{ipums_api_get_extract}{ipums_extract}(extract, api_key = Sys.getenv("IPUMS_API_KEY"))

\method{ipums_api_get_extract}{character}(
  collection,
  extract_number,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

\method{ipums_api_get_extract}{data.frame}(extract_df, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{object}{Either an object of class \code{ipums_extract}, or the name of
an IPUMS data collection as a length one character vector.}

\item{extract}{An extract object created by \code{\link{ipums_api_build_extract}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}

\item{collection}{The IPUMS data collection for the extract.}

\item{extract_number}{The extract number.}

\item{extract_df}{A data.frame in which each row defines one extract.}
}
\value{
If the first argument is an \code{ipums_extract} object or a data
  collection, the function returns an \code{ipums_extract} object. If the
  first argument is a \code{data.frame} of extract definitions, the function
  returns a \code{\link[tibble]{tbl_df}}.
}
\description{
Get information about a submitted extract
}
\section{Methods (by class)}{
\itemize{
\item \code{ipums_extract}: Get information about a submitted
extract, given an \code{ipums_extract} object.

\item \code{character}: Get information about a submitted
extract, given a data collection and extract number.

\item \code{data.frame}: Get information about submitted extracts,
given a data.frame in which each row defines one extract.
}}

