% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_api_get_recent_extracts}
\alias{ipums_api_get_recent_extracts}
\alias{ipums_api_get_recent_extracts_list}
\alias{ipums_api_get_recent_extracts_tbl}
\title{Get information on recent extracts}
\usage{
ipums_api_get_recent_extracts_list(
  collection,
  how_many = 10,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

ipums_api_get_recent_extracts_tbl(
  collection,
  how_many = 10,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{collection}{The IPUMS data collection for the extract.}

\item{how_many}{Number of recent extracts for which you'd like information.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
For \code{ipums_api_get_recent_extracts_list()}, a list of
  \code{ipums_extract} objects. For \code{ipums_api_get_recent_extracts_tbl()},
  a \code{\link[tibble]{tbl-df}} with information on one extract in each row.
}
\description{
Get information on up to ten recent extracts for a given IPUMS collection
via the IPUMS API, returned either as a list or tibble.
}
