% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_api_submit_extract}
\alias{ipums_api_submit_extract}
\title{Submit an extract request via the IPUMS API}
\usage{
ipums_api_submit_extract(extract, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{extract}{An extract object created by \code{\link{ipums_api_build_extract}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
An object of class \code{ipums_extract} containing the extract
  definition and newly-assigned extract number of the submitted extract.
}
\description{
Submit an extract request via the IPUMS API
}
