% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_api_wait_for_extract}
\alias{ipums_api_wait_for_extract}
\alias{ipums_api_wait_for_extract.ipums_extract}
\alias{ipums_api_wait_for_extract.character}
\title{Wait for extract to finish}
\usage{
ipums_api_wait_for_extract(object, ...)

\method{ipums_api_wait_for_extract}{ipums_extract}(
  extract,
  initial_delay_seconds = 10,
  max_delay_seconds = 300,
  timeout_seconds = 10800,
  verbose = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

\method{ipums_api_wait_for_extract}{character}(
  collection,
  number,
  initial_delay_seconds = 10,
  max_delay_seconds = 300,
  timeout_seconds = 10800,
  verbose = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{object}{Either an object of class \code{ipums_extract}, or the name of
an IPUMS data collection as a length one character vector.}

\item{extract}{An \code{ipums_extract} object created with
\code{\link{ipums_api_build_extract}} and submitted with
\code{\link{ipums_api_submit_extract}}.}

\item{initial_delay_seconds}{How many seconds to wait before first status
check.}

\item{max_delay_seconds}{Maximum seconds to wait between status checks. The
function doubles the wait time after each check, but will cap the wait
wait time at this maximum value (300 seconds, or 5 minutes, by default).}

\item{timeout_seconds}{Maximum total number of seconds to continue waiting
for the extract before throwing an error. Defaults to 10,800 seconds (three
hours).}

\item{verbose}{If \code{TRUE}, the default, messages will be printed at the
beginning of each wait interval with the current wait time, each time the
status of the extract is checked, and when the extract is ready to
download. Setting this argument to \code{FALSE} will silence these
messages.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}

\item{collection}{The IPUMS data collection for the extract.}
}
\value{
An object of class \code{ipums_extract} containing the extract
  definition and the URLs from which to download extract files.
}
\description{
Wait for an extract to finish by periodically checking its status and
returning when the extract is ready to download.
}
\section{Methods (by class)}{
\itemize{
\item \code{ipums_extract}: Wait until the extract defined in the
given \code{ipums_extract} object is complete and ready to download.

\item \code{character}: Wait until the extract defined by the
supplied data collection and extract number is complete and ready to
download.
}}

