% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{read_nhgis_codebook}
\alias{read_nhgis_codebook}
\alias{read_ipums_codebook}
\title{Read metadata from an NHGIS extract codebook file}
\usage{
read_nhgis_codebook(cb_file, data_layer = NULL, raw = FALSE)

read_ipums_codebook(cb_file, data_layer = NULL)
}
\arguments{
\item{cb_file}{Path to the codebook file to be loaded. This can be a .zip
archive as provided by the extract system or \code{\link[=download_extract]{download_extract()}},
a directory containing the codebook, or the codebook .txt file itself.}

\item{data_layer}{dplyr \code{\link[dplyr]{select}}-style notation for
uniquely identifying the data layer to load. Required for reading from
.zip files for extracts with multiple files.}

\item{raw}{If \code{TRUE}, read lines of the provided \code{cb_file} instead of
summarizing variable information in an \code{ipums_ddi} object. Defaults to
\code{FALSE}.}
}
\value{
If \code{raw = FALSE}, an \code{ipums_ddi} object with information on the
variables contained in the data for the extract associated with the given
\code{cb_file}.

If \code{raw = TRUE}, a character vector with one element for each
line of the given \code{cb_file}.
}
\description{
NHGIS extracts include a .txt codebook file with metadata about
the contents of the extract. This loads the information contained in this
file into a structured format.
}
\details{
This function includes functionality that previously was contained in
\code{read_ipums_codebook()}. \code{read_ipums_codebook()} supported codebooks
for both NHGIS and IPUMS Terra extracts. IPUMS Terra functionality is no
longer being maintained, and \code{read_ipums_codebook()} has been deprecated.
}
\examples{
# Example NHGIS extract
nhgis_file <- ipums_example("nhgis0707_csv.zip")
codebook <- read_nhgis_codebook(nhgis_file)

# Summary of variables included in the extract:
codebook$var_info

}
