% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_info.r
\name{ipums_conditions}
\alias{ipums_conditions}
\title{Get IPUMS citation and conditions}
\usage{
ipums_conditions(object)
}
\arguments{
\item{object}{A DDI object (loaded with \code{\link{read_ipums_ddi}})}
}
\description{
Gets information about citation and conditions from a DDI.
}
