% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{ipums_ddi-class}
\alias{ipums_ddi-class}
\alias{ipums_ddi}
\title{\code{ipums_ddi} class}
\description{
The \code{ipums_ddi} class provides a data structure for storing the information
contained in the DDI (.xml) files provided with IPUMS microdata extracts.
This is primarily used when loading IPUMS data, but can also be used
to explore metadata for the data in an extract.

The DDI contains metadata about the extract files themselves, including
file name, file path, and extract date as well as information about variables
present in the data, including variable names, descriptions, data types,
implied decimals, and positions in the fixed-width files.

This information is used when loading IPUMS data to correctly parse the
fixed width file format and attach additional variable metadata to the
data object once loaded into R.

Variable metadata for NHGIS extracts can also be stored in
\code{ipums_ddi} objects, even though these metadata are stored in "codebook"
(.txt) files, not .xml files.
\subsection{Creating an \code{ipums_ddi} object}{
\itemize{
\item To create an \code{ipums_ddi} object from an IPUMS microdata extract, use
\code{\link[=read_ipums_ddi]{read_ipums_ddi()}}.
\item To create an \code{ipums_ddi} object from an IPUMS NHGIS extract, use
\code{\link[=read_nhgis_codebook]{read_nhgis_codebook()}}
}
}

\subsection{Loading data}{
\itemize{
\item To load the data associated with an \code{ipums_ddi} object, use
\code{\link[=read_ipums_micro]{read_ipums_micro()}}, \code{\link[=read_ipums_micro_chunked]{read_ipums_micro_chunked()}}, or
\code{\link[=read_ipums_micro_yield]{read_ipums_micro_yield()}}
}
}

\subsection{View metadata}{
\itemize{
\item Use \code{\link[=ipums_var_info]{ipums_var_info()}} to explore variable-level metadata for the variables
included in a dataset.
\item Use \code{\link[=ipums_file_info]{ipums_file_info()}} to explore file-level metadata for an extract.
}
}
}
\keyword{internal}
