% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{ipums_example}
\alias{ipums_example}
\title{Get path to IPUMS example datasets}
\usage{
ipums_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, all available example files will be
listed.}
}
\value{
The path to a specific example file or a vector of all
available files.
}
\description{
Construct file path to example extracts included with ipumsr. These data
are used in package examples and can be used to experiment with ipumsr
functionality.
}
\examples{
# List all available example files
ipums_example()

# Get path to a specific example file
file <- ipums_example("cps_00157.xml")

read_ipums_micro(file)
}
