% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ipums_extract-class}
\alias{ipums_extract-class}
\alias{ipums_extract}
\title{\code{ipums_extract} class}
\description{
The \code{ipums_extract} class provides a data structure for storing the
definition and status of a submitted or unsubmitted IPUMS data extract
request for the purpose of interacting with the IPUMS extract API.

\code{ipums_extract} is a superclass encompassing all of the collection-specific
extract classes. Currently supported collections are:
\itemize{
\item IPUMS USA
\item IPUMS CPS
\item IPUMS NHGIS
}

All objects with class \code{ipums_extract} will also have a collection-specific
subclass (e.g. \code{usa_extract}, \code{cps_extract}) to accommodate
collection-specific differences in extract options and content. All
of these classes share the similarities described below.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\section{Properties}{

Objects of class \code{ipums_extract} have:
\itemize{
\item A \code{class} attribute of the form
\code{c("{collection}_extract", "ipums_extract")}. For instance,
\code{c("cps_extract", "ipums_extract")}
\item A base type of \code{"list"}
\item A \code{names} attribute that is a character vector the same length as the
underlying list
}
}

\section{Creating an extract}{

\itemize{
\item Create an \code{ipums_extract} object from scratch with the appropriate
\link{define_extract} function. These functions take the form
\verb{define_extract_\{collection\}}
\item Use \code{\link[=get_extract_info]{get_extract_info()}} to retrieve information about previously-submitted
extract requests
\item Create an \code{ipums_extract} object from a JSON-formatted definition with
\code{\link[=define_extract_from_json]{define_extract_from_json()}}
}
}

\section{Submitting an extract}{

\itemize{
\item Use \code{\link[=submit_extract]{submit_extract()}} to submit an extract request for processing through
the IPUMS extract API
\item \code{\link[=wait_for_extract]{wait_for_extract()}} will periodically check the status of a submitted
extract request until it is ready to download
\item Use \code{\link[=is_extract_ready]{is_extract_ready()}} to manually check whether a submitted extract
request is ready to download
}
}

\section{Downloading an extract}{

\itemize{
\item Download the data contained in a completed extract with
\code{\link[=download_extract]{download_extract()}}
}
}

\section{Revising an extract definition}{

\itemize{
\item Modify values in an existing extract definition with \code{\link[=add_to_extract]{add_to_extract()}} and
\code{\link[=remove_from_extract]{remove_from_extract()}}
\item Combine definitions from multiple extracts into a single extract with
\code{\link[=combine_extracts]{combine_extracts()}}
}
}

\section{Saving an extract}{

\itemize{
\item Save an extract to a JSON-formatted file with \code{\link[=save_extract_as_json]{save_extract_as_json()}}
}
}

