% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_pillar_print.r
\name{ipums_labelled}
\alias{ipums_labelled}
\title{Create the ipums_labelled class}
\usage{
ipums_labelled(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Other parameters passed to the methods}
}
\description{
To get nice printing in the pillar package we had to add an extra class
to the labelled class. In normal use, you shouldn't need to worry about
this function at all.
}
\keyword{internal}
