% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-ipums-read.R
\name{ipums_list_data}
\alias{ipums_list_data}
\alias{ipums_list_shape}
\alias{ipums_list_raster}
\title{List files available for analysis in an IPUMS extract.}
\usage{
ipums_list_data(file, data_layer = NULL)

ipums_list_shape(file, shape_layer = NULL)

ipums_list_raster(file, raster_layer = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These function are special cases of \code{\link{ipums_list_files}}.

Please use that function instead.
}
\keyword{internal}
