% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.r
\name{ipums_list_files}
\alias{ipums_list_files}
\title{List files available for analysis in an IPUMS extract}
\usage{
ipums_list_files(
  file,
  file_select = NULL,
  types = NULL,
  data_layer = deprecated(),
  shape_layer = deprecated(),
  raster_layer = deprecated()
)
}
\arguments{
\item{file}{Path to a .zip archive or directory containing the IPUMS extract
to be examined.}

\item{file_select}{If \code{file} contains multiple files, an expression
indicating the file names to be included in the output. Files can be
selected by providing a character vector specifying the file name,
\code{\link{dplyr_select_style}} conventions, or by index position. Only files
that match the provided expression will be displayed.}

\item{types}{One or more of \code{"data"}, \code{"shape"}, or \code{"raster"} indicating
the type of files to include in the output. The use of \code{"raster"} has been
deprecated and will be removed in a future release.}

\item{data_layer, shape_layer, raster_layer}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{file_select} instead.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}} showing the types and names of the
available files.
}
\description{
Identify the files that can be loaded from an IPUMS extract. On Windows,
this is generally a zip file (which you can optionally unzip). On
macOS, they are generally unzipped for you, so there will be a
directory.
}
\examples{
nhgis_file <- ipums_example("nhgis0972_csv.zip")
ipums_list_files(nhgis_file) # Only one extract available

}
