% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{ipumsimport_example}
\alias{ipumsimport_example}
\title{Get path to ipumsimport example}
\usage{
ipumsimport_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
Get access to example extracts from this package's `inst/extdata`
directory.
}
\examples{
ipumsimport_example()
ipumsimport_example("cps00006.xml")
}
