% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipumsr.r
\docType{package}
\name{ipumsr-package}
\alias{ipumsr}
\alias{ipumsr-package}
\title{ipumsr: Read 'IPUMS' Extract Files}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An easy way to import census, survey and geographic data provided by 'IPUMS' into R plus tools to help use the associated metadata to make analysis easier. 'IPUMS' data describing 1.4 billion individuals drawn from over 750 censuses and surveys is available free of charge from our website <https://ipums.org>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.ipums.org}
  \item \url{https://github.com/ipums/ipumsr}
  \item Report bugs at \url{https://github.com/ipums/ipumsr/issues}
}

}
\author{
\strong{Maintainer}: Derek Burk \email{ipums+cran@umn.edu}

Authors:
\itemize{
  \item Greg Freedman Ellis
}

Other contributors:
\itemize{
  \item Joe Grover [contributor]
  \item Minnesota Population Center \email{ipums@umn.edu} [copyright holder]
}

}
\keyword{internal}
