% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_join.R
\name{join_failures}
\alias{join_failures}
\title{Report on observations dropped during a join}
\usage{
join_failures(join_results)
}
\arguments{
\item{join_results}{A data frame that has just been created by an
\link[=ipums_shape_left_join]{ipums shape join}.}
}
\value{
A list of data frames, where the first element (\code{shape}) includes
the observations dropped from the shapefile and the second (\code{data})
includes the
observations dropped from the data file.
}
\description{
Helper to display observations that were not matched when joining tabular
and spatial data.
}
\keyword{internal}
