% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_helpers.r
\name{lbl_na_if}
\alias{lbl_na_if}
\title{Convert labelled data values to NA}
\usage{
lbl_na_if(x, .predicate)
}
\arguments{
\item{x}{A \code{\link[haven:labelled]{labelled}} vector}

\item{.predicate}{A function taking \code{.val} and \code{.lbl} arguments that
returns a logical value. Can be provided as an anonymous function or
formula. See examples.}
}
\value{
A \code{\link[haven:labelled]{labelled}} vector
}
\description{
Convert data values in a \code{\link[haven:labelled]{labelled}} vector
to \code{NA} based on the value labels associated with that vector. Ignores
values that do not have a label.
}
\details{
The predicate function uses the terms \code{.val} and \code{.lbl} to refer to the
values and labels, respectively. Use \code{.val} when you want to convert to
\code{NA} conditional on the \emph{values} in the data. Use \code{.lbl} when you want
to convert to \code{NA} conditional on the \emph{labels} used to refer to those values.
}
\examples{
x <- haven::labelled(
  c(10, 10, 11, 20, 30, 99, 30, 10),
  c(Yes = 10, `Yes - Logically Assigned` = 11, No = 20, Maybe = 30, NIU = 99)
)

# Anonymous function notation
lbl_na_if(x, function(.val, .lbl) .val >= 90)
lbl_na_if(x, \(.val, .lbl) .val >= 90 | .lbl \%in\% c("Maybe"))

# purrr-style notation
lbl_na_if(x, ~.lbl \%in\% c("Maybe"))

# Referring to function by name
na_function <- function(.val, .lbl) .val >= 90
lbl_na_if(x, na_function)
}
\seealso{
Other lbl_helpers: 
\code{\link{lbl_add}()},
\code{\link{lbl_clean}()},
\code{\link{lbl_define}()},
\code{\link{lbl}()},
\code{\link{zap_ipums_attributes}()}
}
\concept{lbl_helpers}
