% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-ipums-read.R
\name{make_ddi_from_scratch}
\alias{make_ddi_from_scratch}
\title{Create an \code{ipums_ddi} object}
\usage{
make_ddi_from_scratch(
  file_name = NULL,
  file_path = NULL,
  file_type = NULL,
  ipums_project = NULL,
  extract_date = NULL,
  extract_notes = NULL,
  rectypes = NULL,
  rectype_idvar = NULL,
  rectypes_keyvars = NULL,
  var_info = NULL,
  conditions = NULL,
  citation = NULL,
  file_encoding = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Constructor to make a new \code{\link{ipums_ddi}} object.

This function has been deprecated because it is a low-level function that is
unlikely to be useful externally. Use \code{\link[=read_ipums_ddi]{read_ipums_ddi()}} (for microdata
projects) or \code{\link[=read_nhgis_codebook]{read_nhgis_codebook()}} (for NHGIS) to create an
\code{ipums_ddi} object.
}
\keyword{internal}
