% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{read_ddi}
\alias{read_ddi}
\title{Read metadata about an IPUMS extract from a DDI (.xml) file}
\usage{
read_ddi(ddi_file, data_layer = NULL)
}
\arguments{
\item{ddi_file}{Filepath to DDI xml file}

\item{data_layer}{If dadi_file is an extract with multiple DDIs, a regular expression indicating
which .xml data layer to load.}
}
\description{
Reads the metadata about an IPUMS extract from a DDI file into R.
Includes information about variable and value labels, terms of
usage for the data and positions for the fixed-width file.
}
\examples{
\dontrun{
metadata <- read_ddi("cps_00001.xml")
}
}
\seealso{
Other ipums_metadata: \code{\link{read_ipums_codebook}}
}
