% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-ipums-read.R
\name{read_ipums_codebook}
\alias{read_ipums_codebook}
\title{Read metadata from an IPUMS Terra extract codebook file}
\usage{
read_ipums_codebook(cb_file, data_layer = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Support for IPUMS Terra has been discontinued. \code{read_ipums_codebook()} has
been deprecated and will be removed in a future release.

To read an NHGIS codebook, use \code{\link{read_nhgis_codebook}}.
}
\keyword{internal}
