% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.r
\name{read_ipums_output}
\alias{read_ipums_output}
\title{Read a dataset from the output folder using the YAML}
\usage{
read_ipums_output(yaml, data_file, vars = everything(), n_max = Inf)
}
\arguments{
\item{yaml}{Either a character indicating a filepath to the YAML, or
a dataset from a YAML already loaded using \code{\link{read_var_info_yaml}}.}

\item{data_file}{Filepath to fixed-width file.}

\item{vars}{Dplyr \code{\link[dplyr]{select}}-style group of variables to be loaded.}

\item{n_max}{An integer which specifies the maximum number of rows to read
from the dataset.}
}
\description{
Reads a dataset from the output folder for analysis. Not as fast as
quick.freq.pl and other unix tools, but if you select only a few
variables, can go pretty fast (even on your local computer).
}
\examples{
\dontrun{
data <- read_ipums_output("ses2013c_sestat.yml",
                          "ses2013c_highered.dat"
                          c(RECTYPE, PERSONID, ND2MED, ND2NED),
                          )
}
}
