% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_read.r
\name{read_ipums_sf}
\alias{read_ipums_sf}
\alias{read_ipums_sp}
\title{Read boundary (GIS) files from an IPUMS extract}
\usage{
read_ipums_sf(
  shape_file,
  shape_layer = NULL,
  vars = NULL,
  encoding = NULL,
  bind_multiple = TRUE,
  add_layer_var = NULL,
  ...
)

read_ipums_sp(
  shape_file,
  shape_layer = NULL,
  vars = NULL,
  encoding = NULL,
  bind_multiple = TRUE,
  add_layer_var = NULL,
  ...
)
}
\arguments{
\item{shape_file}{Path to a single .shp file, a .zip archive from an IPUMS
extract, or a directory containing at least one .shp file.}

\item{shape_layer}{For .zip extracts with multiple files, the name of the
shape files to load. Accepts a character vector specifying the file name,
an integer specifying the index of the file to load, or
\code{\link{dplyr_select_style}} conventions. If multiple files are
selected, \code{bind_multiple} must be equal to \code{TRUE}.}

\item{vars}{Names of variables to include in the output data.
By default, includes all variables found in the provided \code{shape_file}.}

\item{encoding}{The text encoding to use when reading the shape file. Typically
the defaults should read the data correctly, but for some extracts you may
need to set them manually.
If \code{NULL}, will attempt to guess the correct encoding using the associated
.cpg file. If none is available, it will default to \code{"latin1"}.
NHGIS files default to \code{"latin1"}. Terra files default to \code{"UTF-8"}.}

\item{bind_multiple}{If \code{TRUE} and \code{shape_file} contains multiple .shp files,
row-bind the files into a single output. Note that some data sources may
not be able to be combined (e.g. \code{SpatialPolygonsDataFrame} and
\code{SpatialPointsDataFrame}).}

\item{add_layer_var}{If \code{TRUE}, add a variable to the output data indicating
the file that each row originates from. By default, this column will be
named \code{"layer"}. If \code{"layer"} already exists in the data, a unique name
will be created to avoid name conflicts. The column name will always be
prefixed with \code{"layer"}. Defaults to \code{TRUE} if \code{bind_multiple = TRUE} and
\code{FALSE} otherwise.}

\item{...}{Additional arguments passed to \code{\link[readr:read_sf]{sf::read_sf()}}
or \code{\link[rgdal:readOGR]{rgdal::readOGR()}}. Note that some arguments are
fixed.}
}
\value{
For \code{read_ipums_sf}, an \code{sf} object. For \code{read_ipums_sp},
a \code{SpatialPolygonsDataFrame} or \code{SpatialPointsDataFrame} object.
}
\description{
Read spatial data from an IPUMS extract into R.
Data can be loaded using \href{https://r-spatial.github.io/sf/}{\code{sf}} or
\href{https://cran.r-project.org/web/packages/sp/index.html}{\code{sp}}.
}
\examples{
shape_file <- ipums_example("nhgis0707_shape_small.zip")
# If sf package is availble, can load as sf object
if (require(sf)) {
  sf_data <- read_ipums_sf(shape_file)
}

# If sp package is available, can load as SpatialPolygonsDataFrame
if (require(sp) && require(rgdal)) {
  sp_data <- read_ipums_sp(shape_file)
}

}
\seealso{
Other ipums_read: 
\code{\link{read_ipums_micro_chunked}()},
\code{\link{read_ipums_micro_yield}()},
\code{\link{read_ipums_micro}()},
\code{\link{read_nhgis}()},
\code{\link{read_terra_area}()},
\code{\link{read_terra_micro}()},
\code{\link{read_terra_raster}()}
}
\concept{ipums_read}
