% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgis_read.r, R/shape_read.r
\name{read_nhgis_sp}
\alias{read_nhgis_sp}
\alias{read_ipums_sp}
\title{Read boundary (GIS) files from an IPUMS extract}
\usage{
read_nhgis_sp(
  shape_file,
  shape_layer = NULL,
  vars = NULL,
  encoding = NULL,
  bind_multiple = FALSE,
  add_layer_var = NULL,
  ...
)

read_ipums_sp(
  shape_file,
  shape_layer = NULL,
  vars = NULL,
  encoding = NULL,
  bind_multiple = TRUE,
  add_layer_var = NULL,
  ...
)
}
\arguments{
\item{shape_file}{Path to a single .shp file, a .zip archive from an IPUMS
extract, or a directory containing at least one .shp file.}

\item{shape_layer}{If \code{shape_file} contains multiple files, an expression
identifying the files to load. Accepts a character string specifying the
file name, \code{\link{dplyr_select_style}} conventions, or index positions. If
multiple files are selected, \code{bind_multiple} must be equal to \code{TRUE}.}

\item{vars}{Names of variables to include in the output data.
By default, includes all variables found in the provided \code{shape_file}.}

\item{encoding}{The text encoding to use when reading the shape file. Typically
the defaults should read the data correctly, but for some extracts you may
need to set them manually.
If \code{NULL}, will attempt to guess the correct encoding using the associated
.cpg file. If none is available, it will default to \code{"latin1"}.
NHGIS files default to \code{"latin1"}. Terra files default to \code{"UTF-8"}.}

\item{bind_multiple}{If \code{TRUE} and \code{shape_file} contains multiple .shp files,
row-bind the files into a single output. Note that some data sources may
not be able to be combined (e.g. \code{SpatialPolygonsDataFrame} and
\code{SpatialPointsDataFrame}).}

\item{add_layer_var}{If \code{TRUE}, add a variable to the output data indicating
the file that each row originates from. By default, this column will be
named \code{"layer"}. If \code{"layer"} already exists in the data, a unique name
will be created to avoid name conflicts. The column name will always be
prefixed with \code{"layer"}. Defaults to \code{TRUE} if \code{bind_multiple = TRUE} and
\code{FALSE} otherwise.}

\item{...}{Additional arguments passed to
\code{\link[rgdal:readOGR]{rgdal::readOGR()}}. Note that some arguments are fixed.}
}
\value{
A \code{SpatialPolygonsDataFrame} or \code{SpatialPointsDataFrame} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Read spatial data from an IPUMS extract into R using the
\href{https://cran.r-project.org/web/packages/sp/index.html}{\code{sp}} package.

This function has been deprecated because of the upcoming retirement
of the \code{rgdal} package. For more information, click
\href{https://r-spatial.org/r/2022/04/12/evolution.html}{here}.

Please use \code{\link[=read_ipums_sf]{read_ipums_sf()}} to load spatial data from IPUMS. To convert
to a SpatialPolygonsDataFrame, use \code{\link[sf:coerce-methods]{sf::as_Spatial()}}.

\code{read_nhgis_sp} is an alias for \code{read_ipums_sp}, but defaults to using
\code{bind_multiple = FALSE}.
}
\examples{
shape_file <- ipums_example("nhgis0707_shape_small.zip")

# If sp package is available, can load as SpatialPolygonsDataFrame
if (require(sp) && require(rgdal)) {
  sp_data <- read_ipums_sp(shape_file)
}

}
\seealso{
Other ipums_read: 
\code{\link{read_ipums_micro_chunked}()},
\code{\link{read_ipums_micro_yield}()},
\code{\link{read_ipums_micro}()},
\code{\link{read_nhgis_sf}()},
\code{\link{read_nhgis}()},
\code{\link{read_terra_area}()},
\code{\link{read_terra_micro}()},
\code{\link{read_terra_raster}()}
}
\concept{ipums_read}
\keyword{internal}
