% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgis_read.r
\name{read_nhgis}
\alias{read_nhgis}
\alias{read_nhgis_sf}
\alias{read_nhgis_sp}
\title{Read data from an NHGIS extract}
\usage{
read_nhgis(
  data_file,
  data_layer = NULL,
  var_attrs = c("val_labels", "var_label", "var_desc"),
  show_conditions = TRUE,
  ...
)

read_nhgis_sf(
  shape_file,
  shape_layer = NULL,
  vars = NULL,
  encoding = NULL,
  bind_multiple = FALSE,
  add_layer_var = NULL,
  ...
)

read_nhgis_sp(
  shape_file,
  shape_layer = NULL,
  vars = NULL,
  encoding = NULL,
  bind_multiple = FALSE,
  add_layer_var = NULL,
  ...
)
}
\arguments{
\item{data_file}{Path to the data file, a .zip archive from an IPUMS
extract, or a directory containing the data file.}

\item{data_layer}{For .zip extracts with multiple datasets, the name of the
data to load. Accepts a character vector specifying the file name, or
\code{\link{dplyr_select_style}} conventions. Data layer must uniquely identify
a dataset.}

\item{var_attrs}{Variable attributes to add from the codebook, defaults to
adding all (val_labels, var_label and var_desc). See
\code{\link{set_ipums_var_attributes}} for more details.}

\item{show_conditions}{If \code{TRUE}, print IPUMS conditions to console when
loading data. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{read_csv()}} or
\code{\link[readr:read_fwf]{read_fwf()}}.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}} of the data found in \code{data_file}.
}
\description{
Load a dataset downloaded from the NHGIS extract system.
}
\examples{
csv_file <- ipums_example("nhgis0707_csv.zip")
data_only <- read_nhgis(csv_file)

}
\seealso{
Other ipums_read: 
\code{\link{read_ipums_micro_chunked}()},
\code{\link{read_ipums_micro_yield}()},
\code{\link{read_ipums_micro}()},
\code{\link{read_ipums_sf}()},
\code{\link{read_terra_area}()},
\code{\link{read_terra_micro}()},
\code{\link{read_terra_raster}()}
}
\concept{ipums_read}
