% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgis_read.r
\name{read_nhgis_sf}
\alias{read_nhgis_sf}
\alias{read_nhgis_sp}
\title{Read data from an NHGIS extract with both spatial and tabular data}
\usage{
read_nhgis_sf(
  data_file,
  shape_file,
  data_layer = NULL,
  shape_layer = data_layer,
  shape_encoding = "latin1",
  verbose = TRUE,
  var_attrs = c("val_labels", "var_label", "var_desc")
)

read_nhgis_sp(
  data_file,
  shape_file,
  data_layer = NULL,
  shape_layer = data_layer,
  shape_encoding = "latin1",
  verbose = TRUE,
  var_attrs = c("val_labels", "var_label", "var_desc")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Please load spatial and tabular data separately using \code{\link[=read_ipums_sf]{read_ipums_sf()}} and
\code{\link[=read_nhgis]{read_nhgis()}}, respectively. Splitting the use of these functions allows
for more control over the data loading and joining process.

To join spatial and tabular data, use an
\link[=ipums_shape_left_join]{ipums_shape_*_join} function.

To convert a \code{SpatialPolygonsDataFrame} or other \code{sp} object to an \code{sf}
object, use \code{\link[sf:coerce-methods]{sf::as_Spatial()}}.
}
\keyword{internal}
