% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.r
\name{read_var_info_yaml}
\alias{read_var_info_yaml}
\title{Read a the YAML metadata on variables from an output folder.}
\usage{
read_var_info_yaml(file)
}
\arguments{
\item{file}{filepath to a YAML (.yml) file in the output folder.}
}
\description{
Reads in the YAML metadata. Though usually called from within
\code{\link{read_ipums_output}}, this function can be useful
for learning about the variables before loading the full dataset.
}
