% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{reconstruct_api_response}
\alias{reconstruct_api_response}
\title{Reattach specifications not returned in API response to NHGIS extracts}
\usage{
reconstruct_api_response(extract, response_extract)
}
\description{
Reattach specifications not returned in API response to NHGIS extracts
}
\details{
In current version of NHGIS API (v1), not all specifications are returned in
the API response after extract submission.

The following are not provided in the API response:
\itemize{
\item shapefiles
\item breakdown_and_data_type_layout
\item time_series_table_layout
\item geographic_extents
\item status
}

Upon completion, the API still provides the requested data. However, without
these specifications it is not possible to reconstruct an extract based
solely on the API response.

Future versions of the API may address this issue at which point
extract_list_from_json() can likely handle extract reconstruction alone.
}
