% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_factor}
\alias{zap_labels}
\alias{is.labelled}
\alias{\%<-\%}
\alias{problems}
\alias{spec}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{everything}
\alias{deprecated}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{haven}{\code{\link[haven]{as_factor}}, \code{\link[haven:labelled]{is.labelled}}, \code{\link[haven]{zap_labels}}}

  \item{lifecycle}{\code{\link[lifecycle]{deprecated}}}

  \item{readr}{\code{\link[readr]{problems}}, \code{\link[readr]{spec}}}

  \item{tidyselect}{\code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{starts_with}}}

  \item{zeallot}{\code{\link[zeallot:operator]{\%<-\%}}}
}}

