% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{remove_from_extract.cps_extract}
\alias{remove_from_extract.cps_extract}
\title{Remove values from an existing IPUMS CPS extract definition}
\usage{
\method{remove_from_extract}{cps_extract}(extract, samples = NULL, variables = NULL, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{samples}{Character vector of samples to remove from the extract.}

\item{variables}{Character vector of variables to remove from the extract.}

\item{...}{Ignored}
}
\value{
A modified \code{cps_extract} object
}
\description{
Remove existing values from an IPUMS CPS extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS CPS extract definition, see
\code{\link[=add_to_extract.cps_extract]{add_to_extract()}}. When replacing values,
it is best to first add new values using
\code{add_to_extract()} before removing values with
\code{remove_from_extract()}. This limits the possibility of producing a
temporarily invalid extract specification.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
\examples{
cps_extract <- define_extract_cps(
  description = "CPS example",
  samples = c("cps2019_03s", "cps2020_03s"),
  variables = c("AGE", "SEX", "YEAR")
)

remove_from_extract(
  cps_extract,
  samples = "cps2020_03s",
  variables = c("AGE", "SEX")
)

# To replace values, use add_to_extract() first to avoid invalid
# extract definitions:
revised_extract <- add_to_extract(cps_extract, variables = "MARST")
remove_from_extract(revised_extract, variables = c("AGE", "SEX", "YEAR"))
}
\seealso{
\code{\link[=add_to_extract.cps_extract]{add_to_extract()}} to add values
to an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_cps]{define_extract_cps()}} to create a new extract definition.
}
\keyword{internal}
