% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{remove_from_extract.ipumsi_extract}
\alias{remove_from_extract.ipumsi_extract}
\title{Remove values from an existing IPUMS International extract definition}
\usage{
\method{remove_from_extract}{ipumsi_extract}(extract, samples = NULL, variables = NULL, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{samples}{Character vector of sample names to remove from the extract.}

\item{variables}{Names of the variables to remove from the extract. All
variable-specific fields for the indicated variables will also be removed.}

\item{...}{Ignored}
}
\value{
A modified \code{ipumsi_extract} object
}
\description{
Remove existing values from an IPUMS International extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS International extract definition, see
\code{\link[=add_to_extract.ipumsi_extract]{add_to_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.

To retain a variable while modifying its particular specifications, first
remove that variable, then add a new specification using \code{add_to_extract()}.
}
\examples{
ipumsi_extract <- define_extract_ipumsi(
  description = "IPUMS-I example",
  samples = c("bj2002a", "bj2013a"),
  variables = list(
    new_variable("AGE", data_quality_flags = TRUE),
    new_variable("SEX", case_selections = "2"),
    "YEAR"
  )
)

remove_from_extract(
  ipumsi_extract,
  samples = "bj2013a",
  variables = c("AGE", "YEAR")
)
}
\seealso{
\code{\link[=add_to_extract.ipumsi_extract]{add_to_extract()}} to add values
to an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_ipumsi]{define_extract_ipumsi()}} to create a new extract definition.
}
\keyword{internal}
