% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{remove_from_extract.micro_extract}
\alias{remove_from_extract.micro_extract}
\title{Remove values from an existing extract definition for an IPUMS microdata
project}
\usage{
\method{remove_from_extract}{micro_extract}(extract, samples = NULL, variables = NULL, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{samples}{Character vector of sample names to remove from the extract
definition.}

\item{variables}{Names of the variables to remove from the extract
definition. All variable-specific fields for the indicated variables
will also be removed.}

\item{...}{Ignored}
}
\value{
A modified \code{micro_extract} object
}
\description{
Remove existing values from an IPUMS microdata extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS microdata extract definition, see
\code{\link[=add_to_extract.micro_extract]{add_to_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.

To retain a variable while modifying its particular specifications, first
remove that variable, then add a new specification using \code{add_to_extract()}.
}
\examples{
usa_extract <- define_extract_usa(
  description = "USA example",
  samples = c("us2013a", "us2014a"),
  variables = list(
    var_spec("AGE", data_quality_flags = TRUE),
    var_spec("SEX", case_selections = "1"),
    "RACE"
  )
)

remove_from_extract(
  usa_extract,
  samples = "us2014a",
  variables = c("AGE", "RACE")
)
}
\seealso{
\code{\link[=add_to_extract.micro_extract]{add_to_extract()}} to add values
to an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract]{define_extract_*()}} to create a new extract definition
from scratch.
}
\keyword{internal}
