% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{remove_from_extract.nhgis_extract}
\alias{remove_from_extract.nhgis_extract}
\title{Remove values from an existing NHGIS extract definition}
\usage{
\method{remove_from_extract}{nhgis_extract}(
  extract,
  datasets = NULL,
  time_series_tables = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{datasets}{Names of the datasets
to remove from the extract definition. All \code{data_tables}, \code{geog_levels},
\code{years}, and \code{breakdown_values} associated with the specified
\code{datasets} will also be removed.}

\item{time_series_tables}{Names of the time series tables
to remove from the extract definition. All \code{geog_levels} and \code{years}
associated  with the specified \code{time_series_tables} will also be removed.}

\item{geographic_extents}{Geographic extents to remove from the extract
definition.}

\item{shapefiles}{\href{https://www.nhgis.org/gis-files}{Shapefiles} to remove
from the extract definition.}

\item{...}{Ignored}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Remove existing values from an IPUMS NHGIS extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS NHGIS extract definition, use
\code{\link[=add_to_extract.nhgis_extract]{add_to_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
Any extract fields that are rendered irrelevant after modifying the extract
will be automatically removed. (For instance, if all \code{time_series_tables}
are removed from an extract, \code{tst_layout} will also be
removed.) Thus, it is not necessary to explicitly remove these values.

If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.

To retain a dataset or time series table while modifying its particular
specifications, first remove that dataset or time series table, then add a
new specification using \code{add_to_extract()}.
}
\examples{
extract <- define_extract_nhgis(
  datasets = ds_spec(
    "1990_STF1",
    data_tables = c("NP1", "NP2", "NP3"),
    geog_levels = "county"
  ),
  time_series_tables = list(
    tst_spec("CW3", c("state", "county")),
    tst_spec("CW5", c("state", "county"))
  )
)

remove_from_extract(
  extract,
  time_series_tables = c("CW3", "CW5")
)

remove_from_extract(
  extract,
  datasets = "1990_STF1"
)
}
\seealso{
\code{\link[=add_to_extract.nhgis_extract]{add_to_extract()}} to add values
to an extract definition.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_nhgis]{define_extract_nhgis()}} to create a new extract definition.
}
\keyword{internal}
