% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{remove_from_extract.usa_extract}
\alias{remove_from_extract.usa_extract}
\title{Remove values from an existing IPUMS USA extract definition}
\usage{
\method{remove_from_extract}{usa_extract}(extract, samples = NULL, variables = NULL, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{samples}{Character vector of samples to remove from the extract.}

\item{variables}{Character vector of variables to remove from the extract.}

\item{...}{Ignored}
}
\value{
A modified \code{usa_extract} object
}
\description{
Remove existing values from an IPUMS USA extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS USA extract definition, see
\code{\link[=add_to_extract.usa_extract]{add_to_extract()}}. When replacing values,
it is best to first add new values using
\code{add_to_extract()} before removing values with
\code{remove_from_extract()}. This limits the possibility of producing a
temporarily invalid extract specification.

For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
\examples{
usa_extract <- define_extract_usa(
  description = "USA example",
  samples = c("us2013a", "us2014a"),
  variables = c("AGE", "SEX", "YEAR")
)

remove_from_extract(
  usa_extract,
  samples = "us2014a",
  variables = c("AGE", "SEX")
)

# To replace values, use add_to_extract() first to avoid invalid
# extract definitions:
revised_extract <- add_to_extract(usa_extract, samples = "us2015a")
remove_from_extract(revised_extract, samples = c("us2013a", "us2014a"))
}
\keyword{internal}
