% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{revise_extract_micro}
\alias{revise_extract_micro}
\alias{add_to_extract.usa_extract}
\alias{remove_from_extract.usa_extract}
\title{Modify values in an existing microdata extract definition}
\usage{
\method{add_to_extract}{usa_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  data_structure = NULL,
  rectangular_on = NULL,
  validate = TRUE
)

\method{remove_from_extract}{usa_extract}(
  extract,
  samples = NULL,
  variables = NULL,
  validate = TRUE,
  ...
)
}
\arguments{
\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to include in the extract. Samples
should be specified using the
\href{https://usa.ipums.org/usa-action/samples/sample_ids}{Sample ID values}.}

\item{variables}{Character vector of variables to include in the extract.}

\item{data_format}{The desired format of the extract data file (one of
"fixed_width", "csv", "stata", "spss", or "sas9").}

\item{data_structure}{Currently, this must be "rectangular", which is also
the default. In the future, the API will also support "hierarchical"
extracts.}

\item{rectangular_on}{Currently, this must be "P", indicating that the
extract will be rectangularized on person records. In the future, the API
will also support household-only extracts (\code{rectangular_on = "H"}).}
}
\value{
A modified \code{usa_extract} object
}
\description{
Either add or remove values in the extract fields for an existing
\code{usa_extract} object.
}
