% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{revise_extract_nhgis}
\alias{revise_extract_nhgis}
\alias{add_to_extract.nhgis_extract}
\alias{remove_from_extract.nhgis_extract}
\title{Modify values in an existing NHGIS extract definition}
\usage{
\method{add_to_extract}{nhgis_extract}(
  extract,
  description = NULL,
  datasets = NULL,
  ds_tables = NULL,
  ds_geog_levels = NULL,
  ds_years = NULL,
  ds_breakdown_values = NULL,
  geographic_extents = NULL,
  breakdown_and_data_type_layout = NULL,
  time_series_tables = NULL,
  tst_geog_levels = NULL,
  tst_layout = NULL,
  shapefiles = NULL,
  data_format = NULL,
  validate = TRUE
)

\method{remove_from_extract}{nhgis_extract}(
  extract,
  datasets = NULL,
  ds_tables = NULL,
  ds_geog_levels = NULL,
  ds_years = NULL,
  ds_breakdown_values = NULL,
  geographic_extents = NULL,
  time_series_tables = NULL,
  tst_geog_levels = NULL,
  shapefiles = NULL,
  validate = TRUE,
  ...
)
}
\arguments{
\item{description}{Description of the extract.}

\item{datasets}{Character vector of datasets to include in the extract,
if any. For more information on NHGIS datasets, click
\href{https://www.nhgis.org/overview-nhgis-datasets}{here}.}

\item{ds_tables}{Character vector or list of summary tables to include in the
extract. This is a subfield of \code{datasets}: see details for available
syntax options. Required if any datasets are included in the extract.}

\item{ds_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
requested summary tables. This is a subfield of \code{datasets}: see
details for available syntax options. Required if any datasets are
included in the extract.}

\item{ds_years}{Character or integer vector or list of years for which to
obtain the data contained in the requested summary tables. Use \code{"*"}
to select all available years for the specified dataset. This is a subfield
of \code{datasets}: see details for available syntax options.  Not all
datasets allow year selection; see \code{\link{get_nhgis_metadata}} to
determine if a dataset allows year selection.}

\item{ds_breakdown_values}{Character vector or list of selected breakdown
values to apply to the requested summary tables. If more than one breakdown
value is requested, \code{breakdown_and_data_type_layout} must also be
specified. This is a subfield of \code{datasets}: see details for available
syntax options.}

\item{geographic_extents}{Character vector of geographic instances to use as
extents for all datasets included in the extract. Use \code{"*"}
to select all available extents. Required when any dataset in the extract
includes a \code{ds_geog_levels} value that requires extent selection.
See \code{\link{get_nhgis_metadata}} to determine whether a requested
\code{ds_geog_level} requires extent selection. Currently, NHGIS supports
extent selection only for blocks and block groups.}

\item{breakdown_and_data_type_layout}{Character indicating the desired layout
of any datasets that have multiple data types or breakdown values. One of
\code{"single_file"} (the default), which keeps all data types and
breakdown values in one file, or \code{"separate_files"}, which splits each
data type or breakdown value into its own file. Required if any datasets
included in the extract consist of multiple data types (for instance,
estimates and margins of error) or have multiple breakdown values
specified. See \code{\link{get_nhgis_metadata}} to determine whether a
requested dataset has multiple data types.}

\item{time_series_tables}{Character vector of time series tables to include
in the extract, if any. For more information on NHGIS time series tables,
click \href{https://www.nhgis.org/time-series-tables}{here}.}

\item{tst_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
provided times series tables. This is a subfield of
\code{time_series_tables}: see details for available syntax options.
Required if any time series tables are included in the extract.}

\item{tst_layout}{The desired layout of all time series tables included in
the extract. One of \code{"time_by_column_layout"} (the default),
\code{"time_by_row_layout"}, or \code{"time_by_file_layout"}. Required when
any time series tables are specified.}

\item{shapefiles}{Character vector of shapefiles to include in the extract,
if any. For more information on NHGIS shapefiles, click
\href{https://www.nhgis.org/gis-files}{here}.}

\item{data_format}{The desired format of the extract data file. One of
\code{"csv_no_header"}, \code{"csv_header"}, or \code{"fixed_width"}.
The default \code{"csv_no_header"} provides only a minimal header in the
first row, while \code{"csv_header"} adds a second, more descriptive header
row. Required when any datasets or time series tables are included in the
extract.}

\item{validate}{Logical value indicating whether to check the modified
extract structure for validity. Defaults to \code{TRUE}}

\item{...}{Not used}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Either add or remove values in the extract fields for an existing
\code{nhgis_extract} object.
}
\details{
NHGIS extracts may contain multiple datasets or time series tables. Each
dataset or time series table is associated with several subfields that apply
only to that particular dataset or time series table. Dataset subfields are
prefixed with \code{ds_}, while time series table subfields are prefixed with
\code{tst_}.

In general, the subfield arguments abide by the same syntax options when
revising an extract as they do when defining an extract (see
\code{\link{define_extract_nhgis}} for details on the core extract definition
syntax). However, there are some exceptions:

In \code{add_to_extract}, if no values are provided to the \code{datasets}
or \code{time_series_tables} arguments, the values passed to any of their
associated subfield arguments will be evaluated relative to all the
datasets or time series tables that \emph{already exist in the extract}. On
the other hand, if any values are passed to the \code{datasets} or
\code{time_series_tables} arguments, the values pass to any of their
associated subfield arguments will be evaluated relative only to
the \emph{specified} datasets or time series tables. For example, if new
datasets are added, then the values provided in \code{ds_tables} will apply
to the new datasets only. Existing datasets can therefore be modified either
by supplying their name to the \code{datasets} argument or by including the
desired additions to their subfields as a named list in the subfield
argument.

In \code{remove_from_extract}, all values passed to either \code{datasets}
or \code{time_series_tables} will be removed from the extract in their
entirety (i.e. along with their currently associated subfield values) before
considering the values passed to any subfield arguments. Subfield arguments
will then be evaluated relative to all the datasets or time series tables
that still exist in the extract.

Because of these intricacies, if you're working with an extract with multiple
datasets or time series tables, the most explicit way to specify subfield
arguments is to use the named list syntax. This ensures that subfield
values are matched to the appropriate datasets or time series tables.

See the examples for demonstrations of this functionality.

For extract fields that take a single value, \code{add_to_extract} will
replace the existing value with the new value provided for that field.

Note that it is possible to produce invalid extracts using
\code{remove_from_extract} (for instance, an extract that includes a
time series table without associated geography levels). This can occur if
you intend to replace the existing values for a required extract field.
If your goal is not simply to add or remove values, but replace values in
an extract, it is recommended that you first use \code{add_to_extract}, and
then use \code{remove_from_extract}, as this will avoid temporarily producing
an invalid extract. Alternatively, you can set \code{validate = FALSE} in
\code{remove_from_extract} to prevent extract validation while you
make the replacement.
}
