% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{ripums_example}
\alias{ripums_example}
\title{Get path to ripumstest examples}
\usage{
ripums_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
Get access to example extracts. Some extracts (such as the full
nhgis shape file and vignette examples) are too big for inclusion
on CRAN and so have been included in a spearate package called
'ripumstest'.
}
\details{
The 'ripumstest' package can be installed using the command:
\code{devtools::install_github('mnpopcenter/ripumstest')}
}
\examples{
ripums_example()
ripums_example("cps_00006.xml")
}
