% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{save_extract_as_json}
\alias{save_extract_as_json}
\title{Save an \code{\link[=ipums_extract-class]{ipums_extract}} to disk as JSON}
\usage{
save_extract_as_json(extract, file)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{file}{File path to which to write the JSON-formatted extract
definition.}
}
\value{
The file path where the extract definition was written, invisibly
}
\description{
Save an \code{\link[=ipums_extract-class]{ipums_extract}} to a JSON-formatted file.
For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
Note that the output JSON file will only contain the extract
parameters required to submit a new extract request. Ancillary fields,
like \code{number}, \code{status}, etc. will not be included.
}
\examples{
my_extract <- define_extract_usa("Example", "us2013a", "YEAR")

extract_json_path <- file.path(tempdir(), "usa_extract.json")
save_extract_as_json(my_extract, file = extract_json_path)

copy_of_my_extract <- define_extract_from_json(extract_json_path)

identical(my_extract, copy_of_my_extract)
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{set_ipums_api_key}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
