% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_datasets}
\alias{validate_datasets}
\title{Validate NHGIS dataset specifications}
\usage{
validate_datasets(extract)
}
\arguments{
\item{extract}{An extract object created with
\code{\link{define_extract_nhgis}} or
returned from another ipumsr API function.}
}
\description{
Validate NHGIS dataset specifications
}
