% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_ipums_extract}
\alias{validate_ipums_extract}
\title{Validate structure of an IPUMS extract}
\usage{
validate_ipums_extract(x)
}
\arguments{
\item{x}{An extract object created with
\code{\link{define_extract_nhgis}}, \code{\link{define_extract_micro}} or
returned from another ipumsr API function.}
}
\description{
Validate structure of an IPUMS extract
}
