% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{validate_time_series_tables}
\alias{validate_time_series_tables}
\title{Validate NHGIS time_series_tables specifications}
\usage{
validate_time_series_tables(extract)
}
\arguments{
\item{extract}{An extract object created with
\code{\link{define_extract_nhgis}} or
returned from another ipumsr API function.}
}
\description{
Validate NHGIS time_series_tables specifications
}
