% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_extract.nhgis_extract}
\alias{add_to_extract.nhgis_extract}
\title{Add values to an existing NHGIS extract}
\usage{
\method{add_to_extract}{nhgis_extract}(
  extract,
  description = NULL,
  datasets = NULL,
  ds_tables = NULL,
  ds_geog_levels = NULL,
  ds_years = NULL,
  ds_breakdown_values = NULL,
  geographic_extents = NULL,
  breakdown_and_data_type_layout = NULL,
  time_series_tables = NULL,
  tst_geog_levels = NULL,
  tst_layout = NULL,
  shapefiles = NULL,
  data_format = NULL,
  ...
)
}
\arguments{
\item{extract}{A \code{nhgis_extract} object. This can be a user-defined
extract (see \code{\link{define_extract_nhgis}()}) or an extract object
returned from another \code{ipumsr} API function.}

\item{description}{Description of the extract.}

\item{datasets}{Character vector of datasets to add or modify in the extract.
Dataset names that do not already exist in the extract will be added along
with the values passed to any dataset subfield arguments. Dataset names
that already exist in the extract will not be added, but their subfields
will be modified based on the values provided to any dataset subfield
arguments. See details.}

\item{ds_tables}{Character vector or list of summary tables to add to the
extract. This is a subfield of \code{datasets}: values provided to this argument
will be applied to the datasets included in \code{datasets} (see details for
available syntax options). Required if any new datasets are being added to
the extract.}

\item{ds_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in
the requested summary tables. This is a subfield of \code{datasets}: values
provided to this argument will be applied to the datasets included in
\code{datasets} (see details for available syntax options). Required if
any new datasets are being added to the extract.}

\item{ds_years}{Character or integer vector or list of years for which to
obtain the data contained in the requested summary tables. Use "*" to
select all available years for the specified dataset. This is a subfield of
\code{datasets}: values provided to this argument will be applied to the datasets
included in \code{datasets} (see details for available syntax options).
Not all datasets allow year selection; see \code{\link{get_nhgis_metadata}}
to determine if a dataset allows year selection.}

\item{ds_breakdown_values}{Character vector or list of selected breakdown
values to apply to the requested summary tables. If more than one breakdown
value is requested, \code{breakdown_and_data_type_layout} must also be specified.
This is a subfield of \code{datasets}: values provided to this argument will be
applied to the datasets included in \code{datasets} (see details for
available syntax options).}

\item{geographic_extents}{Character vector of geographic instances to use as
extents for all datasets included in the extract. Use \code{"*"}
to select all available extents. Required when any dataset in the extract
includes a \code{ds_geog_levels} value that requires extent selection.
See \code{\link{get_nhgis_metadata}()} to determine whether a requested
\code{ds_geog_level} requires extent selection. Currently, NHGIS supports
extent selection only for blocks and block groups.}

\item{breakdown_and_data_type_layout}{The desired layout
of any datasets that have multiple data types or breakdown values. One of
\code{"single_file"} (the default), which keeps all data types and
breakdown values in one file, or \code{"separate_files"}, which splits each
data type or breakdown value into its own file. Required if any datasets
included in the extract consist of multiple data types (for instance,
estimates and margins of error) or have multiple breakdown values
specified. See \code{\link{get_nhgis_metadata}()} to determine whether a
requested dataset has multiple data types.}

\item{time_series_tables}{Character vector of time series tables to add or
modify in the extract. Time series table names that do not already exist in
the extract will be added along with the values passed to any time series
table subfield arguments. Time series table names that already exist in the
extract will not be added, but their subfields will be modified based on
the values provided to any time series table subfield arguments.}

\item{tst_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
provided times series tables. This is a subfield of
\code{time_series_tables}: values provided to this argument will be applied
to the time series tables included in \code{time_series_tables} (see details
for available syntax options). Required if any new time series tables are
being added to the extract.}

\item{tst_layout}{The desired layout of all time series tables included in
the extract. One of \code{"time_by_column_layout"} (the default),
\code{"time_by_row_layout"}, or \code{"time_by_file_layout"}. Required when
any time series tables are specified.}

\item{shapefiles}{Character vector of shapefiles to add to the extract, if
any. For more information on NHGIS shapefiles, click here.}

\item{data_format}{The desired format of the extract data file. One of
\code{"csv_no_header"}, \code{"csv_header"}, or \code{"fixed_width"}.
The default \code{"csv_no_header"} provides only a minimal header in the
first row, while \code{"csv_header"} adds a second, more descriptive header
row. Required when any datasets or time series tables are included in the
extract.}

\item{...}{Ignored}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Add new values to any extract fields of an NHGIS extract. This can include
adding new datasets and/or time series tables along with associated
subfields, modifying dataset and/or time series table subfields, or
updating extract-wide parameters.

To remove existing values from an NHGIS extract, click
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract}()}.

In general, adding to an extract follows the same syntax conventions as used
in \code{\link{define_extract_nhgis}()}. See details for more information
on how values passed to dataset and time series table subfields are
interpreted.
}
\details{
NHGIS extracts may contain multiple datasets or time series tables. Each
dataset or time series table is associated with several subfields that apply
only to that particular dataset or time series table. Dataset subfields are
prefixed with \code{ds_}, while time series table subfields are prefixed with
\code{tst_}.

There are three ways the values passed to these arguments can be provided:

\itemize{
  \item{If values are passed as a \strong{vector}, they will be
  applied across all datasets or time series tables specified in the
  \code{datasets} or \code{time_series_tables} arguments}
  \item{If values are passed as an \strong{unnamed list}, they will be
  matched to datasets or time series tables by index. That
  is, the first element will be associated with the first dataset included
  in \code{datasets}, the second element with the second dataset in
  \code{datasets}, etc.}
  \item{If values are passed as a \strong{named list}, they
  will be matched to datasets or time series tables by name. Names should
  correspond to datasets or time series tables that exist in the extract
  definition (unrecognized names will be ignored).}
}

If no values are provided to the \code{datasets}
or \code{time_series_tables} arguments, all the existing datasets and/or
time series tables will be considered to be eligible for modification. That
is, the values passed to any dataset or time series table subfield arguments
will be evaluated relative to all of the datasets or time series tables that
exist in the extract.

If any values are passed to the \code{datasets} or
\code{time_series_tables} arguments, the values passed to any of their
associated subfield arguments will be evaluated relative only to
the \emph{specified} datasets or time series tables.

For example, if new datasets are added, then the values provided in
\code{ds_tables} will apply
to the new datasets only. Existing datasets can be modified either
by supplying their name to the \code{datasets} argument or by including the
desired additions to their subfields as a named list in the subfield
argument.

For extract fields that take a single value, \code{add_to_extract} will
replace the existing value with the new value provided for that field.
It is not necessary to first remove this value using
\code{remove_from_extract}.
}
\examples{
extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  ds_tables = c("NP1", "NP2"),
  ds_geog_levels = "county"
)

# Modify existing datasets in extract:
extract <- add_to_extract(
  extract,
  ds_tables = "NP3"
)

# Add new dataset. New subfields will only be attached to datasets specified
# in the datasets argument:
extract <- add_to_extract(
  extract,
  datasets = "1980_STF1",
  ds_tables = "NT1A",
  ds_geog_levels = c("county", "state")
)

# Modify existing datasets.
# Vectors recycle to all datasets, lists match by index or name.
add_to_extract(
  extract,
  ds_tables = list(`1990_STF1` = "NP4", `1980_STF1` = "NT1B"),
  ds_geog_levels = "nation"
)
}
