% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_extract.nhgis_extract}
\alias{add_to_extract.nhgis_extract}
\title{Add values to an existing IPUMS NHGIS extract definition}
\usage{
\method{add_to_extract}{nhgis_extract}(
  extract,
  description = NULL,
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  breakdown_and_data_type_layout = NULL,
  tst_layout = NULL,
  data_format = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{description}{Description of the extract.}

\item{datasets}{\href{https://www.nhgis.org/overview-nhgis-datasets}{Datasets}
to include in the extract request.}

\item{data_tables}{Summary tables to retrieve for each of the requested
\code{datasets}. If no \code{datasets} are provided, applies to those that
already exist in the extract definition. Required if any new \code{datasets}
are specified.

See Details section for syntax options when an extract definition has
multiple \code{datasets}.}

\item{time_series_tables}{\href{https://www.nhgis.org/time-series-tables}{Time series tables}
to include in the extract request.}

\item{geog_levels}{Geographic levels (for example, \code{"county"} or \code{"state"})
at which to obtain data for the requested \code{datasets} and
\code{time_series_tables}. If no \code{datasets} and/or \code{time_series_tables} are
provided, applies to those that already
exist in the extract definition. Required if any new \code{datasets} or
\code{time_series_tables} are specified.

See Details section for syntax options when an extract definition has
multiple \code{datasets} or \code{time_series_tables}.}

\item{years}{Years for which to obtain the data contained in the requested
\code{datasets}. If no \code{datasets} are provided, applies to those that
already exist in the extract definition.

Use \code{"*"} to select all available years for a given dataset.
Use \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a dataset allows year selection.

See Details section for syntax options when an extract definition has
multiple \code{datasets}.}

\item{breakdown_values}{\href{https://www.nhgis.org/frequently-asked-questions-faq#breakdowns}{Breakdown values}
to to apply to the requested \code{datasets}. If no \code{datasets} are provided,
applies to those that already exist in the extract definition.

See Details section for syntax options when an extract definition has
multiple \code{datasets}.}

\item{geographic_extents}{Vector of geographic extents to use for
all of the \code{datasets} in the extract definition (for instance, to obtain
data within a particular state). Use \code{"*"} to select all available extents.

Required when the extract definition includes any \code{geog_levels} that
require extent selection. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a
geographic level requires extent selection. At the time of writing, NHGIS
supports extent selection only for blocks and block groups.}

\item{shapefiles}{\href{https://www.nhgis.org/gis-files}{Shapefiles} to include
in the extract request.}

\item{breakdown_and_data_type_layout}{The desired layout
of any \code{datasets} that have multiple data types or breakdown values.
\itemize{
\item \code{"single_file"} (default) keeps all data types and breakdown values in
one file
\item \code{"separate_files"} splits each data type or breakdown value into its
own file
}

Required if any \code{datasets} included in the extract definition consist of
multiple data types (for instance, estimates and margins of error) or have
multiple breakdown values specified. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to
determine whether a requested dataset has multiple data types.}

\item{tst_layout}{The desired layout of all \code{time_series_tables} included in
the extract definition. One of:
\itemize{
\item \code{"time_by_column_layout"} (wide format, default): rows correspond to
geographic units, columns correspond to different times in the time
series
\item \code{"time_by_row_layout"} (long format): rows correspond to a single
geographic unit at a single point in time
\item \code{"time_by_file_layout"}: data for different times are provided in
separate files
}

Required when an extract definition includes any \code{time_series_tables}.}

\item{data_format}{The desired format of the extract data file. One of:
\itemize{
\item \code{"csv_no_header"} (default) includes only a minimal header in the first
row
\item \code{"csv_header"} includes a second, more descriptive header row.
(By default this header is removed when reading data with \code{\link[=read_nhgis]{read_nhgis()}},
but it can be retained.)
\item \code{"fixed_width"} provides data in a fixed width format
}

Required when an extract definition includes any \code{datasets} or
\code{time_series_tables}.}

\item{...}{Ignored}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Add new values to an IPUMS NHGIS extract definition.
All fields are optional, and if omitted, will be unchanged.
Supplying a value for fields that take a single value, such as
\code{description} and \code{data_format}, will replace the existing value with
the supplied value.

In general, adding to an extract follows the same syntax conventions as used
in \code{\link[=define_extract_nhgis]{define_extract_nhgis()}}. See Details section for more
information on how values passed to dataset and time series table subfields
are interpreted.

To remove existing values from an IPUMS NHGIS extract definition, use
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract()}}. To add values
contained in another extract definition, use \code{\link[=combine_extracts]{combine_extracts()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
For extract definitions that include any \code{datasets},
corresponding \code{data_tables} and \code{geog_levels} must be specified. For
extract definitions that include any \code{time_series_tables}, corresponding
\code{geog_levels} must be specified.

NHGIS extract definitions may contain multiple \code{datasets} or
\code{time_series_tables.} Each dataset or time series table is associated with
several subfields that apply only to that particular dataset or time series
table.

The following arguments are applied to an extract definition's \code{datasets}:
\itemize{
\item \code{data_tables}
\item \code{years}
\item \code{breakdown_values}
}

The following arguments are applied to an extract definition's \code{datasets}
\emph{and} \code{time_series_tables}:
\itemize{
\item \code{geog_levels}
}

To add the values of these arguments to \emph{multiple} \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a vector.
In this case, if any values are
provided in the \code{datasets} or \code{time_series_tables} arguments, subfield values
will only be applied to those \code{datasets} and/or \code{time_series_tables}.
Otherwise, they will be applied to all \code{datasets} and/or
\code{time_series_tables} that exist in the extract definition.

To add the values of these arguments to \emph{specific} \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a named
list. In this case, names should correspond to the \code{datasets} and/or
\code{time_series_tables} and their associated values will only be added to
the indicated dataset or time series table. Values can be added for new
\code{datasets} and \code{time_series_tables} or those that already exist in the
extract definition.

See examples for demonstrations of this syntax.

Note that while it is possible to match elements to \code{datasets} and/or
\code{time_series_tables} by index by providing an unnamed list, we recommend
using names to explicitly specify the values that should be added to each
of the request's \code{datasets} and \code{time_series_tables}.

For extract fields that take a single value, \code{add_to_extract()} will
replace the existing value with the new value provided for that field.
It is not necessary to first remove this value using
\code{remove_from_extract()}.

If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
\examples{
extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  geog_levels = "county"
)

# Modify existing datasets in extract:
extract2 <- add_to_extract(
  extract,
  data_tables = "NP3"
)

extract2

# Add a new dataset or time series table. Fields that apply to each dataset
# or time series table are only applied to the newly added dataset or
# time series table:
extract <- add_to_extract(
  extract,
  datasets = "1980_STF1",
  data_tables = "NT1A",
  geog_levels = c("county", "state")
)

add_to_extract(
  extract,
  time_series_tables = "A00",
  geog_levels = "state"
)

# To modify existing datasets, use a named list to specify which datasets
# should be attached to which `data_tables` or `geog_levels`.
# Vector arguments are attached to all datasets.
add_to_extract(
  extract,
  data_tables = list(
    "1990_STF1" = "NP4",
    "1980_STF1" = "NT1B"
  ),
  geog_levels = "nation"
)

# Values that can only take a single value are replaced
add_to_extract(extract, data_format = "fixed_width")$data_format

# You can also combine two extract requests together
combine_extracts(extract, extract2)
}
\seealso{
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract()}} to remove
values from an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_nhgis]{define_extract_nhgis()}} to create a new extract definition.
}
\keyword{internal}
