% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{add_to_extract.nhgis_extract}
\alias{add_to_extract.nhgis_extract}
\title{Add values to an existing IPUMS NHGIS extract definition}
\usage{
\method{add_to_extract}{nhgis_extract}(
  extract,
  description = NULL,
  datasets = NULL,
  time_series_tables = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  breakdown_and_data_type_layout = NULL,
  tst_layout = NULL,
  data_format = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{description}{Description of the extract.}

\item{datasets}{List of \code{ipums_dataset} objects created by \code{\link[=new_dataset]{new_dataset()}}
containing the specifications
for the \href{https://www.nhgis.org/overview-nhgis-datasets}{datasets}
to include in the extract request. See examples.

If a dataset already exists in the extract, its new specifications
will be added to those that already exist for that dataset.}

\item{time_series_tables}{List of \code{nhgis_tst} objects created by \code{\link[=new_tst]{new_tst()}}
containing the specifications for the
\href{https://www.nhgis.org/time-series-tables}{time series tables}
to include in the extract request.

If a time series table already exists in the extract, its new
specifications will be added to those that already exist for that time
series table.}

\item{geographic_extents}{Vector of geographic extents to use for
all of the \code{datasets} in the extract definition (for instance, to obtain
data within a particular state). Use \code{"*"} to select all available extents.

Required when any of the \code{datasets} included in the extract definition
include \code{geog_levels} that require extent selection. See
\code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a geographic level requires extent
selection. At the time of writing, NHGIS supports extent selection only
for blocks and block groups.}

\item{shapefiles}{\href{https://www.nhgis.org/gis-files}{Shapefiles} to include
in the extract request.}

\item{breakdown_and_data_type_layout}{The desired layout
of any \code{datasets} that have multiple data types or breakdown values.
\itemize{
\item \code{"single_file"} (default) keeps all data types and breakdown values in
one file
\item \code{"separate_files"} splits each data type or breakdown value into its
own file
}

Required if any \code{datasets} included in the extract definition consist of
multiple data types (for instance, estimates and margins of error) or have
multiple breakdown values specified. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to
determine whether a requested dataset has multiple data types.}

\item{tst_layout}{The desired layout of all \code{time_series_tables} included in
the extract definition.
\itemize{
\item \code{"time_by_column_layout"} (wide format, default): rows correspond to
geographic units, columns correspond to different times in the time
series
\item \code{"time_by_row_layout"} (long format): rows correspond to a single
geographic unit at a single point in time
\item \code{"time_by_file_layout"}: data for different times are provided in
separate files
}

Required when an extract definition includes any \code{time_series_tables}.}

\item{data_format}{The desired format of the extract data file.
\itemize{
\item \code{"csv_no_header"} (default) includes only a minimal header in the first
row
\item \code{"csv_header"} includes a second, more descriptive header row.
\item \code{"fixed_width"} provides data in a fixed width format
}

Note that \code{\link[=read_nhgis]{read_nhgis()}} removes the additional header row in
\code{"csv_header"} files by default.

Required when an extract definition includes any \code{datasets} or
\code{time_series_tables}.}

\item{...}{Ignored}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Add new values to an IPUMS NHGIS extract definition.
All fields are optional, and if omitted, will be unchanged.
Supplying a value for fields that take a single value, such as
\code{description} and \code{data_format}, will replace the existing value with
the supplied value.

In general, adding to an extract follows the same syntax conventions as used
in \code{\link[=define_extract_nhgis]{define_extract_nhgis()}}. Use \code{new_dataset()} and \code{new_tst()} to create
dataset and time series table specifications.

To remove existing values from an IPUMS NHGIS extract definition, use
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
For extract fields that take a single value, \code{add_to_extract()} will
replace the existing value with the new value provided for that field.
It is not necessary to first remove this value using
\code{remove_from_extract()}.

If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
\examples{
extract <- define_extract_nhgis(
  datasets = new_dataset("1990_STF1", c("NP1", "NP2"), "county")
)

# Add a new dataset or time series table to the extract
add_to_extract(
  extract,
  datasets = new_dataset("1990_STF2a", "NPA1", "county")
)

add_to_extract(
  extract,
  time_series_tables = new_tst("A00", "state")
)

# If a dataset/time series table name already exists in the definition
# its specification will be modified by adding the new specification to
# the existing one
add_to_extract(
  extract,
  datasets = new_dataset("1990_STF1", "NP4", "nation")
)

# Values that can only take a single value are replaced
add_to_extract(extract, data_format = "fixed_width")$data_format
}
\seealso{
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract()}} to remove
values from an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=define_extract_nhgis]{define_extract_nhgis()}} to create a new extract definition.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.
}
\keyword{internal}
