% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{add_to_extract}
\alias{add_to_extract}
\title{Add values to an existing IPUMS extract definition}
\usage{
add_to_extract(extract, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{...}{Additional arguments specifying the extract fields and values to
add to the extract definition.

All arguments available in a collection's
\code{\link[=define_extract]{define_extract_*()}} function can be
passed to \code{add_to_extract()}.}
}
\value{
An object of the same class as \code{extract} containing the modified
extract definition
}
\description{
Add or replace values in an existing \code{ipums_extract} object.
This function is an S3 generic whose behavior will depend on the
subclass (i.e. collection) of the extract being modified.
\itemize{
\item To add to an \strong{IPUMS Microdata} extract definition, click
\link[=add_to_extract.micro_extract]{here}. This includes:
\itemize{
\item IPUMS USA
\item IPUMS CPS
\item IPUMS International
}
\item To add to an \strong{IPUMS NHGIS} extract definition, click
\link[=add_to_extract.nhgis_extract]{here}
}

In general, for a given collection, the arguments to \code{add_to_extract()}
are identical to those used when defining an extract for that collection. For
more information about defining an extract, click \link[=define_extract]{here}.

To remove existing values from an extract definition, use
\code{\link[=remove_from_extract]{remove_from_extract()}}. To add values
contained in another extract definition, use \code{\link[=combine_extracts]{combine_extracts()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
usa_extract <- define_extract_usa(
  description = "2013 ACS Data",
  samples = "us2013a",
  variables = c("SEX", "AGE", "YEAR")
)

nhgis_extract <- define_extract_nhgis(
  datasets = new_dataset(
    "1990_STF1",
    data_tables = c("NP1", "NP2"),
    geog_levels = "county"
  )
)

# Add new samples and variables
add_to_extract(
  usa_extract,
  samples = c("us2014a", "us2015a"),
  variables = new_variable("MARST", data_quality_flags = TRUE)
)

# Update existing variables
add_to_extract(
  usa_extract,
  variables = new_variable("SEX", case_selections = "1")
)

# Add a new dataset or time series table
add_to_extract(
  nhgis_extract,
  datasets = new_dataset(
    "1980_STF1",
    data_tables = "NT1A",
    geog_levels = c("county", "state")
  )
)

# Update existing datasets/time series tables
add_to_extract(
  nhgis_extract,
  datasets = new_dataset("1990_STF1", c("NP1", "NP2"), "state")
)

# Values that can only take a single value are replaced
add_to_extract(nhgis_extract, data_format = "fixed_width")$data_format
}
\seealso{
\code{\link[=remove_from_extract]{remove_from_extract()}} to remove values from an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract]{define_extract_*()}} to create a new extract request from
scratch.
}
