% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{add_to_extract}
\alias{add_to_extract}
\title{Add values to an existing IPUMS extract}
\usage{
add_to_extract(extract, ...)
}
\arguments{
\item{extract}{An object inheriting from \code{ipums_extract}}

\item{...}{Additional arguments specifying the extract fields and values to
add to the extract. The available arguments correspond to the available
arguments in the extract definition function for the class of the extract
specified in \code{extract}.}
}
\value{
An object of the same class as \code{extract} containing the modified
  extract definition
}
\description{
Add or update values for specific fields in an existing \code{ipums_extract}
object. This function is an S3 generic whose behavior will depend on the
class (i.e. collection) of the extract being modified.

In general, for a given collection, the arguments to \code{add_to_extract}
are identical to those used when defining an extract for that collection. For
more about defining an extract, click \link[=define_extract]{here}.

For collection-specific documentation, see the links below:

\itemize{
  \item To add to an \strong{NHGIS} extract, click
    \link[=add_to_extract.nhgis_extract]{here}
  \item To add to a \strong{microdata} extract, click
    \link[=add_to_extract.usa_extract]{here}
}

To remove existing values from an extract, see
\code{\link{remove_from_extract}}.
}
