% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_extract}
\alias{add_to_extract}
\title{Add values to an existing IPUMS extract definition}
\usage{
add_to_extract(extract, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{...}{Additional arguments specifying the extract fields and values to
add to the extract definition.

All arguments available in a collection's \link{define_extract} function can be
passed to \code{add_to_extract()}.}
}
\value{
An object of the same class as \code{extract} containing the modified
extract definition
}
\description{
Add or replace values in an existing \code{ipums_extract} object.
This function is an S3 generic whose behavior will depend on the
subclass (i.e. collection) of the extract being modified.
\itemize{
\item To add to an \strong{IPUMS USA} extract definition, click
\link[=add_to_extract.usa_extract]{here}
\item To add to a \strong{IPUMS CPS} extract definition, click
\link[=add_to_extract.cps_extract]{here}
\item To add to an \strong{IPUMS NHGIS} extract definition, click
\link[=add_to_extract.nhgis_extract]{here}
}

In general, for a given collection, the arguments to \code{add_to_extract()}
are identical to those used when defining an extract for that collection. For
more information about defining an extract, click \link[=define_extract]{here}.

To remove existing values from an extract definition, see
\code{\link[=remove_from_extract]{remove_from_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
usa_extract <- define_extract_usa(
  description = "2013 ACS Data",
  samples = "us2013a",
  variables = c("SEX", "AGE", "YEAR")
)

add_to_extract(usa_extract, samples = "us2014a")
add_to_extract(usa_extract, variables = c("MARST", "INCTOT"))

nhgis_extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  geog_levels = "county"
)

# Add a new dataset/time series table
add_to_extract(
  nhgis_extract,
  datasets = "1980_STF1",
  data_tables = "NT1A",
  geog_levels = c("county", "state")
)

# Add to existing datasets/time series tables
add_to_extract(nhgis_extract, geog_levels = "state")

# Values that can only take a single value are replaced
add_to_extract(nhgis_extract, data_format = "fixed_width")$data_format
}
\seealso{
\code{\link[=remove_from_extract]{remove_from_extract()}} to remove values from an extract definition.

\code{\link[=combine_extracts]{combine_extracts()}} to combine multiple extract definitions.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_usa]{define_extract_usa()}}, \code{\link[=define_extract_cps]{define_extract_cps()}}, or \code{\link[=define_extract_nhgis]{define_extract_nhgis()}}
to create a new extract definition.
}
