% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_extract}
\alias{add_to_extract}
\title{Add values to an existing IPUMS extract}
\usage{
add_to_extract(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  data_structure = NULL,
  rectangular_on = NULL,
  validate = TRUE
)
}
\arguments{
\item{extract}{An extract object created with \code{\link{define_extract_micro}} or
returned from another ipumsr API function.}

\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to add to the extract, if any.}

\item{variables}{Character vector of variables to add to the extract, if any.}

\item{data_format}{The desired format of the extract data file (one of
"fixed_width", "csv", "stata", "spss", or "sas9").}

\item{data_structure}{Currently, this must be "rectangular", which is also
the default. In the future, the API will also support "hierarchical"
extracts.}

\item{rectangular_on}{Currently, this must be "P", indicating that the
extract will be rectangularized on person records. In the future, the API
will also support household-only extracts (\code{rectangular_on = "H"}).}

\item{validate}{Logical value indicating whether to check the modified
extract structure for validity. Defaults to \code{TRUE}.}
}
\value{
A modified \code{ipums_extract} object
}
\description{
Add new values to any extract fields of an \code{ipums_extract} object.

To remove existing values from an extract, see
\code{\link{remove_from_extract}}.
}
\examples{
my_extract <- define_extract_micro("usa", "Example", "us2013a", "YEAR")

add_to_extract(
  my_extract,
  description = "Revised extract",
  samples = "us2014a"
)
}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
