% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_extract.usa_extract}
\alias{add_to_extract.usa_extract}
\title{Add values to an existing IPUMS USA extract definition}
\usage{
\method{add_to_extract}{usa_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to add to the extract.
Values should correspond to IPUMS USA
\href{https://usa.ipums.org/usa-action/samples/sample_ids}{sample ID values}}

\item{variables}{Character vector of variables to add to the extract.}

\item{data_format}{The desired format of the extract data file. One of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}.}

\item{...}{Ignored}
}
\value{
A modified \code{usa_extract} object
}
\description{
Add new values or replace existing values in an IPUMS USA extract definition.
All fields are optional, and if omitted, will be unchanged. Supplying a value
for fields that take a single value, such as \code{description} and \code{data_format},
will replace the existing value with the supplied value.

To remove existing values from an IPUMS USA extract definition, see
\code{\link[=remove_from_extract.usa_extract]{remove_from_extract()}}.

For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
