% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{add_to_extract.usa_extract}
\alias{add_to_extract.usa_extract}
\title{Add values to an existing USA extract}
\usage{
\method{add_to_extract}{usa_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  data_structure = NULL,
  rectangular_on = NULL,
  validate = TRUE
)
}
\arguments{
\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to add to the extract, if any.}

\item{variables}{Character vector of variables to add to the extract, if any.}

\item{data_format}{The desired format of the extract data file (one of
"fixed_width", "csv", "stata", "spss", or "sas9").}

\item{data_structure}{Currently, this must be "rectangular", which is also
the default. In the future, the API will also support "hierarchical"
extracts.}

\item{rectangular_on}{Currently, this must be "P", indicating that the
extract will be rectangularized on person records. In the future, the API
will also support household-only extracts (\code{rectangular_on = "H"}).}

\item{validate}{Logical value indicating whether to check the modified
extract structure for validity. Defaults to \code{TRUE}}
}
\value{
A modified \code{usa_extract} object
}
\description{
Add new values to any extract fields of a USA extract.

To remove existing values from an extract, see
\code{\link{remove_from_extract.usa_extract}}.
}
\examples{
my_extract <- define_extract_micro("usa", "Example", "us2013a", "YEAR")

add_to_extract(
  my_extract,
  description = "Revised extract",
  samples = "us2014a"
)
}
