% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{add_to_extract.usa_extract}
\alias{add_to_extract.usa_extract}
\title{Add values to an existing IPUMS USA extract}
\usage{
\method{add_to_extract}{usa_extract}(
  extract,
  description = NULL,
  samples = NULL,
  variables = NULL,
  data_format = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to add to the extract.
Values should correspond to \link{USA sample ID values}
(https://usa.ipums.org/usa-action/samples/sample_ids)}

\item{variables}{Character vector of variables to add to the extract.}

\item{data_format}{The desired format of the extract data file (one of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}).}

\item{...}{Ignored}
}
\value{
A modified \code{usa_extract} object
}
\description{
Add new values or replace existing values in an IPUMS USA extract. All
fields are optional, and if omitted, will be unchanged. Supplying a value
for fields that take a single value, such as \code{description} and \code{data_format},
will replace the existing value with the supplied value.

To remove existing values from an IPUMS USA extract, see
\code{\link[=remove_from_extract.usa_extract]{remove_from_extract()}}.

For an overview of ipumsr microdata API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract, this function will reset the definition to an
unsubmitted state.
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract.cps_extract}()},
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{define_extract}},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract.cps_extract}()},
\code{\link{remove_from_extract.usa_extract}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
