% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{combine_extracts}
\alias{combine_extracts}
\title{Combine multiple IPUMS extract definitions into one}
\usage{
combine_extracts(...)
}
\arguments{
\item{...}{Arbitrary number of \code{ipums_extract} objects to be combined. All
extracts must belong to the same collection.}
}
\value{
An \code{ipums_extract} object of the same collection as the extracts
provided in \code{...} containing the combined extract definition.
}
\description{
Create a single extract definition that includes all of the
specifications included in a set of \code{ipums_extract} objects of the same
collection.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
Values that exist in more than one of the provided extract definitions will
be de-duplicated such that only one entry for the duplicated specification is
included in the final extract.

Some extract fields can only take a single value. In the event that the
definitions being combined each have different values for such a field, the
value found in the first extract definition provided to \code{...} is retained in
the final extract.

To modify values in a single extract definition, see
\code{\link[=add_to_extract]{add_to_extract()}} or
\code{\link[=remove_from_extract]{remove_from_extract()}}.
}
\examples{
ext1 <- define_extract_nhgis(
  datasets = ds_spec(
    "2011_2015_ACS5a",
    data_tables = "B00001",
    geog_levels = "county"
  )
)

ext2 <- define_extract_nhgis(
  datasets = list(
    ds_spec("2011_2015_ACS5a", c("B00002", "B01001"), "county"),
    ds_spec("2012_2016_ACS5a", c("B00002", "B01001"), "county")
  )
)

combine_extracts(ext1, ext2)
}
\seealso{
\code{\link[=add_to_extract]{add_to_extract()}} and \code{\link[=remove_from_extract]{remove_from_extract()}} to revise an extract
definition.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract]{define_extract_*()}} to create a new extract definition
from scratch.
}
