% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract}
\alias{define_extract}
\title{Define an IPUMS extract object}
\description{
Define an extract request object for a specific IPUMS collection to be
to be submitted via the IPUMS extract API for that collection. For an
overview of \code{ipumsr} API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.

Currently, ipumsr supports extract definitions for NHGIS and USA collections.

For collection-specific documentation, see the links below:

\itemize{
  \item{\code{\link{define_extract_nhgis}}}
  \item{\code{\link{define_extract_micro}}}
}
}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{define_extract_nhgis}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
