% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract}
\alias{define_extract}
\title{Define an IPUMS extract object}
\description{
Specify the parameters for a new IPUMS extract request object to be
submitted via the IPUMS extract API.
}
\section{Supported collections}{

Currently, \code{ipumsr} supports extract definitions for the following
collections:

\itemize{
  \item{IPUMS NHGIS}
  \item{IPUMS USA}
}

For information on defining NHGIS extracts, see
\code{\link{define_extract_nhgis}()}

For information defining extracts for IPUMS microdata collections, see
\code{\link{define_extract_micro}()}
}

\section{Value}{

These functions produce an object inheriting from class \code{ipums_extract}
with a subclass based on the collection corresponding to the extract. For
instance, an IPUMS NHGIS extract created with
\code{\link{define_extract_nhgis}} will return an object of classes
\code{nhgis_extract} and \code{ipums_extract}. These objects are compatible
with the rest of the API functionality provided by \code{ipumsr}.

For an overview of \code{ipumsr} API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}

\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
