% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract}
\alias{define_extract}
\title{Define an IPUMS extract object}
\description{
Specify the parameters for a new IPUMS extract request object to be
submitted via the IPUMS extract API.
}
\section{Supported collections}{

Currently, ipumsr supports extract definitions for the following
collections:
\itemize{
\item \strong{IPUMS NHGIS}: \code{\link[=define_extract_nhgis]{define_extract_nhgis()}}
\item \strong{IPUMS USA}: \code{\link[=define_extract_usa]{define_extract_usa()}}
\item \strong{IPUMS CPS}: \code{\link[=define_extract_cps]{define_extract_cps()}}
}
}

\section{Value}{

These functions produce an \code{\link[=ipums_extract-class]{ipums_extract}} object
with a subclass based on the collection corresponding to the extract request.
For instance, an IPUMS NHGIS extract created with
\code{\link[=define_extract_nhgis]{define_extract_nhgis()}} will return an object of classes
\code{nhgis_extract} and \code{ipums_extract}. These objects are compatible
with the rest of the API functionality provided by ipumsr.

For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}

