% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract}
\alias{define_extract}
\title{Define an IPUMS extract object}
\description{
Specify the parameters for a new IPUMS extract request object to be
submitted via the IPUMS API.

Currently, \code{ipumsr} supports extract definitions for the following
collections:

\itemize{
  \item{IPUMS NHGIS}
  \item{IPUMS USA}
}

For information on how to define an extract for a specific collection, use
the links below:

\itemize{
  \item{\code{\link{define_extract_nhgis}}}
  \item{\code{\link{define_extract_micro}}}
}

For an overview of \code{ipumsr} API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\seealso{
Other ipums_api_micro: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()}

Other ipums_api_nhgis: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{get_nhgis_metadata}()}
}
\concept{ipums_api_micro}
\concept{ipums_api_nhgis}
