% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract}
\alias{define_extract}
\title{Define an extract request object}
\usage{
define_extract(
  collection,
  description,
  samples,
  variables,
  data_format = c("fixed_width", "csv", "stata", "spss", "sas9"),
  data_structure = c("rectangular", "hierarchical"),
  rectangular_on = c("P", "H")
)
}
\arguments{
\item{collection}{The IPUMS data collection for the extract.}

\item{description}{Description of the extract.}

\item{samples}{Character vector of samples to include in the extract. Samples
should be specified using the
\href{https://usa.ipums.org/usa-action/samples/sample_ids}{Sample ID values}.}

\item{variables}{Character vector of variables to include in the extract.}

\item{data_format}{The desired format of the extract data file (one of
"fixed_width", "csv", "stata", "spss", or "sas9").}

\item{data_structure}{One of "rectangular" or "hierarchical" (defaults to
"rectangular")}

\item{rectangular_on}{For rectangular \code{data_structure}, should the data
be rectangular on persons ("P") or households ("H")? Defaults to "P". If
\code{data_structure} is hierarchical, this argument is ignored and set to
missing automatically.}
}
\value{
An object of class \code{ipums_extract} containing the extract
  definition.
}
\description{
Define an extract request object to be submitted via the IPUMS API. For an
overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\examples{
my_extract <- define_extract("usa", "Example", "us2013a", "YEAR")

my_hierarchical_extract <- define_extract(
  collection = "usa",
  description = "A hierarchical example",
  samples = c("us2013a", "us2014a"),
  variables = c("YEAR", "AGE", "SEX", "RACE"),
  data_format = "csv",
  data_structure = "hierarchical"
)

}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
