% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract_from_json}
\alias{define_extract_from_json}
\title{Create an extract object from a JSON-formatted definition}
\usage{
define_extract_from_json(extract_json, collection)
}
\arguments{
\item{extract_json}{A JSON string, or the path to file containing JSON.}

\item{collection}{The IPUMS data collection for the extract.}
}
\value{
An object of class "ipums_extract".
}
\description{
Create an object inheriting from class "ipums_extract" based on an extract
definition formatted as JSON. For an overview of ipumsr API functionality,
see \code{vignette("ipums-api", package = "ipumsr")}.
}
\examples{
my_extract <- define_extract_micro("usa", "Example", "us2013a", "YEAR")

extract_json_path <- file.path(tempdir(), "usa_extract.json")
save_extract_as_json(my_extract, file = extract_json_path)

copy_of_my_extract <- define_extract_from_json(extract_json_path, "usa")

identical(my_extract, copy_of_my_extract)

}
\seealso{
Other ipums_api_micro: 
\code{\link{define_extract_micro}()},
\code{\link{define_extract}}

Other ipums_api_nhgis: 
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract}},
\code{\link{get_nhgis_metadata}()}
}
\concept{ipums_api_micro}
\concept{ipums_api_nhgis}
