% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract_nhgis}
\alias{define_extract_nhgis}
\title{Define an NHGIS extract request}
\usage{
define_extract_nhgis(
  description = "",
  dataset = NULL,
  data_tables = NULL,
  ds_geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  time_series_table = NULL,
  ts_geog_levels = NULL,
  shapefiles = NULL,
  data_format = NULL,
  breakdown_and_data_type_layout = NULL,
  time_series_table_layout = NULL,
  geographic_extents = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{dataset}{Character name of dataset to include in the extract, if any.}

\item{data_tables}{Character vector of data tables associated with the
specified dataset to include in the extract. Required if a dataset is
specified.}

\item{ds_geog_levels}{Character vector of geographic levels associated with
the specified dataset to include in the extract. Required if a dataset is
specified.}

\item{years}{Character or integer vector of years associated with the
specified dataset to include in the extract. Use "*" to select all
available years for the specified dataset.}

\item{breakdown_values}{Character vector of selected breakdown values to
associated with the specified dataset to include in the extract. If more
than one breakdown value is requested, `breakdown_and_data_type_layout`
must also be specified.}

\item{time_series_table}{Character name of time series table to include in
the extract, if any.}

\item{ts_geog_levels}{Character vector of geographic levels associated with
the specified time series table to include in the extract. Required if a
time series table is specified.}

\item{shapefiles}{Character vector of shapefiles to include in the extract,
if any.}

\item{data_format}{The desired format of the extract data file (one of
\code{"csv_no_header"}, \code{"csv_header"}, or \code{"fixed_width"}).
\code{"csv_no_header"} provides a minimal header in the first row, while
\code{"csv_header"} adds a second, more descriptive header row.
Required when any dataset or time_series_table is selected.}

\item{breakdown_and_data_type_layout}{Character indicating the layout of the
dataset when multiple data types or breakdown values are specified.
\code{"separate_files"} splits up each data type or breakdown combo into
its own file. \code{"single_file"} keeps all datatypes and breakdown combos
in one file. Required when a dataset has multiple breakdowns or data types.}

\item{time_series_table_layout}{Character indicating the layout of the
specified time series table. One of \code{"time_by_column_layout"},
\code{"time_by_row_layout"}, or \code{"time_by_file_layout"}. Required when
a time series table is specified.}

\item{geographic_extents}{A list of geographic instances to use as extents
for the dataset specified in the request. To select all extents, use
\code{"*"}. Required when a geographic level on a dataset is specified
where \code{has_geog_extent_selection} is true.}
}
\value{
An object of class \code{ipums_extract} containing the extract
  definition.
}
\description{
Define an extract request object to be submitted via the IPUMS NHGIS
extract API.
}
\examples{
my_extract <- define_extract_nhgis(
  description = "test nhgis extract 1",
  dataset = "1990_STF3",
  data_tables = "NP57",
  ds_geog_levels = "tract",
  data_format = "csv_no_header"
)

}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract_micro}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
