% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract_nhgis}
\alias{define_extract_nhgis}
\title{Define an NHGIS extract request}
\usage{
define_extract_nhgis(
  description = "",
  datasets = NULL,
  ds_tables = NULL,
  ds_geog_levels = NULL,
  ds_years = NULL,
  ds_breakdown_values = NULL,
  geographic_extents = NULL,
  breakdown_and_data_type_layout = NULL,
  time_series_tables = NULL,
  tst_geog_levels = NULL,
  tst_layout = NULL,
  shapefiles = NULL,
  data_format = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{datasets}{Character vector of datasets to include in the extract,
if any. For more information on NHGIS datasets, click
\href{https://www.nhgis.org/overview-nhgis-datasets}{here}.}

\item{ds_tables}{Character vector or list of summary tables to include for
the provided datasets. This is a subfield of \code{datasets} (see details
for available
syntax options). Required if any datasets are included in the extract.}

\item{ds_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
requested summary tables. This is a subfield of \code{datasets} (see
details for available syntax options). Required if any datasets are
included in the extract.}

\item{ds_years}{Character or integer vector or list of years for which to
obtain the data contained in the requested summary tables. Use \code{"*"}
to select all available years for the specified dataset. This is a subfield
of \code{datasets} (see details for available syntax options). Not all
datasets allow year selection; see \code{\link{get_nhgis_metadata}()} to
determine if a dataset allows year selection.}

\item{ds_breakdown_values}{Character vector or list of selected breakdown
values to apply to the requested summary tables. If more than one breakdown
value is requested, \code{breakdown_and_data_type_layout} must also be
specified. This is a subfield of \code{datasets} (see details for available
syntax options).}

\item{geographic_extents}{Character vector of geographic instances to use as
extents for all datasets included in the extract. Use \code{"*"}
to select all available extents. Required when any dataset in the extract
includes a \code{ds_geog_levels} value that requires extent selection.
See \code{\link{get_nhgis_metadata}()} to determine whether a requested
\code{ds_geog_level} requires extent selection. Currently, NHGIS supports
extent selection only for blocks and block groups.}

\item{breakdown_and_data_type_layout}{The desired layout
of any datasets that have multiple data types or breakdown values. One of
\code{"single_file"} (the default), which keeps all data types and
breakdown values in one file, or \code{"separate_files"}, which splits each
data type or breakdown value into its own file. Required if any datasets
included in the extract consist of multiple data types (for instance,
estimates and margins of error) or have multiple breakdown values
specified. See \code{\link{get_nhgis_metadata}()} to determine whether a
requested dataset has multiple data types.}

\item{time_series_tables}{Character vector of time series tables to include
in the extract, if any. For more information on NHGIS time series tables,
click \href{https://www.nhgis.org/time-series-tables}{here}.}

\item{tst_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
provided times series tables. This is a subfield of
\code{time_series_tables} (see details for available syntax options).
Required if any time series tables are included in the extract.}

\item{tst_layout}{The desired layout of all time series tables included in
the extract. One of \code{"time_by_column_layout"} (the default),
\code{"time_by_row_layout"}, or \code{"time_by_file_layout"}. Required when
any time series tables are specified.}

\item{shapefiles}{Character vector of shapefiles to include in the extract,
if any. For more information on NHGIS shapefiles, click
\href{https://www.nhgis.org/gis-files}{here}.}

\item{data_format}{The desired format of the extract data file. One of
\code{"csv_no_header"}, \code{"csv_header"}, or \code{"fixed_width"}.
The default \code{"csv_no_header"} provides only a minimal header in the
first row, while \code{"csv_header"} adds a second, more descriptive header
row. Required when any datasets or time series tables are included in the
extract.}
}
\value{
An object of class \code{nhgis_extract} containing the extract
  definition.
}
\description{
Define an extract request object to be submitted via the IPUMS NHGIS
extract API.

For more information on NHGIS data, click
\href{https://www.nhgis.org/data-availability}{here}. For the NHGIS FAQ,
click \href{https://www.nhgis.org/frequently-asked-questions-faq}{here}.

For general information on the IPUMS NHGIS Extract API, click
\href{https://developer.ipums.org/docs/workflows/create_extracts/nhgis_data/}{here}.
Note that some of the terminology used in the API has been altered
for use in the \code{ipumsr} package.
}
\details{
NHGIS extracts may contain multiple datasets or time series tables. Each
dataset or time series table is associated with several subfields that apply
only to that particular dataset or time series table. Dataset subfields are
prefixed with \code{ds_}, while time series table subfields are prefixed with
\code{tst_}.

There are three ways the values passed to these arguments can be provided:

\itemize{
  \item{If values are passed as a \strong{vector}, they will be
  applied to all of the datasets or time series tables in the extract
  definition.}
  \item{If values are passed as an \strong{unnamed list}, they will be
  matched to the datasets or time series tables by index. That
  is, the first element will be associated with the first dataset in the
  extract definition, the second element with the second dataset in the
  extract, etc.}
  \item{If values are passed as a \strong{named list}, they
  will be matched to the datasets or time series tables by name. Names should
  correspond to datasets or time series tables that exist in the extract
  definition (unrecognized names will be ignored).}
}

Users can find more information on the values that can be passed to these
subfield arguments for a given dataset or time series table by using
\code{\link{get_nhgis_metadata}()}.
}
\examples{
define_extract_nhgis(
  description = "Test NHGIS extract",
  datasets = "1990_STF3",
  ds_tables = "NP57",
  ds_geog_levels = c("county", "tract"),
)

# For extracts with multiple datasets or time series tables, subfield
# arguments are recycled to all datasets:
define_extract_nhgis(
  description = "Extract with multiple datasets",
  datasets = c("2014_2018_ACS5a", "2015_2019_ACS5a"),
  ds_tables = "B01001",
  ds_geog_levels = c("state", "county")
)

# To attach specific subfield values to each dataset or time series table,
# pass a list to the subfield argument.
# With an unnamed list, items are matched by index:
define_extract_nhgis(
  description = "Extract with multiple time series tables",
  time_series_tables = c("CW3", "CW5"),
  tst_geog_levels = list("state", "county")
)

# With a named list, items are matched by name:
define_extract_nhgis(
  description = "Extract with multiple time series tables",
  time_series_tables = c("CW3", "CW5"),
  tst_geog_levels = list(CW5 = "county", CW3 = "state")
)

}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_usa}()},
\code{\link{define_extract}},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
