% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract_nhgis}
\alias{define_extract_nhgis}
\title{Define an IPUMS NHGIS extract request}
\usage{
define_extract_nhgis(
  description = "",
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  breakdown_and_data_type_layout = NULL,
  tst_layout = NULL,
  data_format = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{datasets}{\href{https://www.nhgis.org/overview-nhgis-datasets}{Datasets}
to include in the extract request.}

\item{data_tables}{Summary tables to retrieve for each of the requested
\code{datasets}. Required for all \code{datasets} in the extract definition.

See Details section for syntax options when an extract definition has
multiple \code{datasets}.}

\item{time_series_tables}{\href{https://www.nhgis.org/time-series-tables}{Time series tables}
to include in the extract request.}

\item{geog_levels}{Geographic levels (e.g. \code{"county"} or \code{"state"})
at which to obtain data for the requested \code{datasets} or
\code{time_series_tables}. Required for all \code{datasets} and \code{time_series_tables}
in the extract definition.

See Details section for syntax options when an extract
definition has multiple \code{datasets} and/or \code{time_series_tables}.}

\item{years}{Years for which to obtain the data contained in the requested
\code{datasets} or \code{time_series_tables}. Use \code{"*"} to select all available
years for a given dataset or time series table.

Use \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a dataset allows year selection.

See Details section for syntax options when an extract definition has
multiple \code{datasets}.}

\item{breakdown_values}{\href{https://www.nhgis.org/frequently-asked-questions-faq#breakdowns}{Breakdown values}
to apply to the requested \code{datasets}.

See Details section for syntax options when an extract definition has
multiple \code{datasets}.}

\item{geographic_extents}{Vector of geographic extents to use for
all of the \code{datasets} in the extract definition (for instance, to obtain
data within a particular state). Use \code{"*"} to select all available extents.

Required when the extract definition includes any \code{geog_levels} that
require extent selection. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a
geographic level requires extent selection. At the time of writing, NHGIS
supports extent selection only for blocks and block groups.}

\item{shapefiles}{\href{https://www.nhgis.org/gis-files}{Shapefiles} to include
in the extract request.}

\item{breakdown_and_data_type_layout}{The desired layout
of any \code{datasets} that have multiple data types or breakdown values.
\itemize{
\item \code{"single_file"} (default) keeps all data types and breakdown values in
one file
\item \code{"separate_files"} splits each data type or breakdown value into its
own file
}

Required if any \code{datasets} included in the extract definition consist of
multiple data types (for instance, estimates and margins of error) or have
multiple breakdown values specified. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to
determine whether a requested dataset has multiple data types.}

\item{tst_layout}{The desired layout of all \code{time_series_tables} included in
the extract definition. One of:
\itemize{
\item \code{"time_by_column_layout"} (wide format, default): rows correspond to
geographic units, columns correspond to different times in the time
series
\item \code{"time_by_row_layout"} (long format): rows correspond to a single
geographic unit at a single point in time
\item \code{"time_by_file_layout"}: data for different times are provided in
separate files
}

Required when an extract definition includes any \code{time_series_tables}.}

\item{data_format}{The desired format of the extract data file. One of:
\itemize{
\item \code{"csv_no_header"} (default) includes only a minimal header in the first
row
\item \code{"csv_header"} includes a second, more descriptive header row.
(By default this header is removed when reading data with \code{\link[=read_nhgis]{read_nhgis()}},
but it can be retained.)
\item \code{"fixed_width"} provides data in a fixed width format
}

Required when an extract definition includes any \code{datasets} or
\code{time_series_tables}.}
}
\value{
An object of class \code{\link[=ipums_extract-class]{nhgis_extract}} containing
the extract definition.
}
\description{
Define an IPUMS NHGIS request to be submitted via the IPUMS
API. An extract request contains the specifications required
to obtain a particular set of data from the IPUMS NHGIS system.

Use \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to browse and identify data sources for use
in NHGIS extract definitions.

See Details section for information about creating valid extract
definitions. For general information, see the NHGIS
\href{https://www.nhgis.org/data-availability}{data source overview} and the
\href{https://www.nhgis.org/frequently-asked-questions-faq}{FAQ}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
An NHGIS extract definition must include at least one dataset, time series
table, or shapefile. For extract definitions that include any \code{datasets},
corresponding \code{data_tables} and \code{geog_levels} must be specified. For
extract definitions that include any \code{time_series_tables}, corresponding
\code{geog_levels} must be specified.

NHGIS extract definitions may contain multiple \code{datasets} or
\code{time_series_tables.} Each dataset or time series table is associated with
several subfields that apply only to that particular dataset or time series
table.

The following arguments are applied to the requested \code{datasets}:
\itemize{
\item \code{data_tables}
\item \code{breakdown_values}
}

The following arguments are applied to the requested \code{datasets} \emph{and}
\code{time_series_tables}:
\itemize{
\item \code{geog_levels}
\item \code{years}
}

To apply the values of these arguments to \emph{all} of the \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a vector:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{define_extract_nhgis(
  datasets = c("2016_2020_ACS5a", "2017_2021_ACS5a"),
  data_tables = "B01001",
  geog_levels = c("county", "state")
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 2016_2020_ACS5a
#>   Tables: B01001
#>   Geog Levels: county, state
#> 
#> Dataset: 2017_2021_ACS5a
#>   Tables: B01001
#>   Geog Levels: county, state
}\if{html}{\out{</div>}}

To apply the values of these arguments to \emph{specific} \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a named
list. In this case, names should correspond to the \code{datasets} and/or
\code{time_series_tables} and their associated values will only be applied to
the indicated dataset or time series table:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{define_extract_nhgis(
  datasets = "2016_2020_ACS5a",
  data_tables = "B01001",
  time_series_tables = "A00",
  geog_levels = list(
    "2016_2020_ACS5a" = "county",
    "A00" = "state"
  )
)
#> Unsubmitted IPUMS NHGIS extract 
#> Description: 
#> 
#> Dataset: 2016_2020_ACS5a
#>   Tables: B01001
#>   Geog Levels: county
#> 
#> Time Series Table: A00
#>   Geog Levels: state
}\if{html}{\out{</div>}}

Note that while it is possible to match elements to \code{datasets} and/or
\code{time_series_tables} by index by providing an unnamed list, we recommend
using names to explicitly specify the values that should be applied to each
of the request's \code{datasets} and \code{time_series_tables}.

Alternatively, you can create multiple extract definitions and combine
them with \code{\link[=combine_extracts]{combine_extracts()}}.
}
\examples{
# Extract definition for tables from an NHGIS dataset
my_extract <- define_extract_nhgis(
  description = "Example NHGIS extract",
  datasets = "1990_STF3",
  data_tables = "NP57",
  geog_levels = c("county", "tract")
)

my_extract

# Extract definition for an NHGIS time series table
define_extract_nhgis(
  description = "Example NHGIS extract",
  time_series_tables = "CL8",
  geog_levels = "county"
)

# Use get_nhgis_metadata() to identify data source names
\dontrun{
get_nhgis_metadata("time_series_tables")
}

# Fields that are attached to specific datasets/time series tables
# are recycled to all datasets/time series tables in an extract by default
define_extract_nhgis(
  description = "Extract definition with multiple datasets",
  datasets = c("2014_2018_ACS5a", "2015_2019_ACS5a"),
  data_tables = "B01001",
  geog_levels = c("state", "county")
)

# To instead attach specific values to each dataset or time series table,
# use a named list instead of a vector:
define_extract_nhgis(
  description = "Extract with multiple time series tables",
  time_series_tables = c("CW3", "CW5"),
  geog_levels = list(CW5 = "county", CW3 = "state")
)

# For extracts with datasets and time series tables,
# geog_levels are matched to both sources:
define_extract_nhgis(
  description = "Extract with datasets and time series tables",
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  time_series_tables = "CL6",
  geog_levels = "state"
)

define_extract_nhgis(
  description = "Extract with datasets and time series tables",
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  time_series_tables = "CL6",
  geog_levels = list("1990_STF1" = "county", "CL6" = "state")
)

\dontrun{
# Use the extract definition to submit an extract request to the API
submit_extract(my_extract)
}
}
\seealso{
\code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to find data to include in an extract definition.

\code{\link[=submit_extract]{submit_extract()}}, \code{\link[=download_extract]{download_extract()}}, and \code{\link[=get_extract_info]{get_extract_info()}} to
process and manage an extract request.

\code{\link[=save_extract_as_json]{save_extract_as_json()}} and \code{\link[=define_extract_from_json]{define_extract_from_json()}} to share an
extract definition.

\code{\link[=add_to_extract.nhgis_extract]{add_to_extract()}},
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract()}} and
\code{\link[=combine_extracts]{combine_extracts()}} to
revise an extract definition.
}
