% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract_nhgis}
\alias{define_extract_nhgis}
\title{Define an NHGIS extract request}
\usage{
define_extract_nhgis(
  description = "",
  datasets = NULL,
  ds_tables = NULL,
  ds_geog_levels = NULL,
  ds_years = NULL,
  ds_breakdown_values = NULL,
  geographic_extents = NULL,
  breakdown_and_data_type_layout = NULL,
  time_series_tables = NULL,
  tst_geog_levels = NULL,
  tst_layout = NULL,
  shapefiles = NULL,
  data_format = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{datasets}{Character vector of datasets to include in the extract,
if any. For more information on NHGIS datasets, click
\href{https://www.nhgis.org/overview-nhgis-datasets}{here}.}

\item{ds_tables}{Character vector or list of summary tables to include in the
extract. If passed as a list, elements will be matched to the extract's
datasets by index (see details). Required if any datasets are included in
the extract.}

\item{ds_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
requested summary tables. If passed as a list, elements will be matched to
the extract's datasets by index (see details). Required if any datasets are
included in the extract.}

\item{ds_years}{Character or integer vector or list of years for which to
obtain the data contained in the requested summary tables. Use \code{"*"}
to select all available years for the specified dataset. If passed as a
list, elements will be matched to the extract's datasets by index (see
details). Not all datasets allow year selection; see
\code{\link{get_nhgis_metadata}} to determine if a dataset allows year
selection.}

\item{ds_breakdown_values}{Character vector or list of selected breakdown
values to apply to the requested summary tables. If more than one breakdown
value is requested, \code{breakdown_and_data_type_layout} must also be
specified. If passed as a list, elements will be matched to the extract's
datasets by index (see details).}

\item{geographic_extents}{A list of geographic instances to use as extents
for all datasets included in the extract. Use \code{"*"}
to select all available extents. Required when any dataset in the extract
includes a \code{ds_geog_levels} value that requires extent selection.
See \code{\link{get_nhgis_metadata}} to determine whether a requested
\code{ds_geog_level} requires extent selection. Currently, NHGIS supports
extent selection only for blocks and block groups.}

\item{breakdown_and_data_type_layout}{Character indicating the desired layout
of any datasets that have multiple data types or breakdown values.
The default \code{"single_file"} keeps all data types and breakdown values
in one file. \code{"separate_files"} splits each
data type or breakdown value into its own file. Required if any datasets
included in the extract consist of multiple data types (for instance,
estimates and margins of error) or have multiple breakdown values
specified. See \code{\link{get_nhgis_metadata}} to determine whether a
requested dataset has multiple data types.}

\item{time_series_tables}{Character vector of time series tables to include
in the extract, if any. For more information on NHGIS time series tables,
click \href{https://www.nhgis.org/time-series-tables}{here}.}

\item{tst_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") for which to obtain the data contained in the
provided times series tables. If passed as a list, elements will be matched to
the extract's time series tables by index (see details). Required if any
time series tables are included in the extract.}

\item{tst_layout}{Character indicating the layout of all
time series tables included in the extract. One of
\code{"time_by_column_layout"} (the default), \code{"time_by_row_layout"},
or \code{"time_by_file_layout"}. Required when any time series tables are
specified.}

\item{shapefiles}{Character vector of shapefiles to include in the extract,
if any. For more information on NHGIS shapefiles, click
\href{https://www.nhgis.org/gis-files}{here}.}

\item{data_format}{The desired format of the extract data file. One of
\code{"csv_no_header"}, \code{"csv_header"}, or \code{"fixed_width"}.
The default \code{"csv_no_header"} provides only a minimal header in the
first row, while \code{"csv_header"} adds a second, more descriptive header
row. Required when any datasets or time series tables are included in the
extract.}
}
\value{
An object of class \code{nhgis_extract} containing the extract
  definition.
}
\description{
Define an extract request object to be submitted via the IPUMS NHGIS
extract API.
}
\details{
An extract can contain multiple datasets, each of which may or may not
contain the same values for the associated dataset subfields
(\code{ds_tables}, \code{ds_geog_levels}, \code{ds_years}, and
\code{ds_breakdown_values}). When specifying an extract that contains multiple
datasets, you can either choose to recycle these subfield specifications to
all datasets in the extract, or to specify specific values for the subfields
for each dataset.

To recycle subfield specifications to all datasets, pass
a vector to the subfield argument. To provide unique subfield values for each
dataset in the extract, instead pass a list of values to the subfield
argument. In this case, the values for that subfield will be linked to the
datasets in index order (for instance, if 2 datasets are specified, the
first element in the list of \code{ds_tables} would be linked to the first
dataset, and the second element in the list of \code{ds_tables} would be
linked to the second dataset). The same holds true for time series tables.

If a subfield argument is passed as a list, the list must be the same length
as the number of datasets (for all arguments with prefix \code{ds_}) or
time series tables (for all arguments with prefix \code{tst_}) specified in
the extract definition. Note that individual \emph{elements} of these lists
can be of arbitrary length.
}
\examples{
define_extract_nhgis(
  description = "Test NHGIS extract",
  datasets = "1990_STF3",
  ds_tables = "NP57",
  ds_geog_levels = c("county", "tract"),
)

# For extracts with multiple datasets or time series tables, subfield
# arguments are recycled to all datasets:
define_extract_nhgis(
  description = "Extract with multiple datasets",
  datasets = c("2014_2018_ACS5a", "2015_2019_ACS5a"),
  ds_tables = "B01001",
  ds_geog_levels = c("state", "county")
)

# To attach specific subfield values to each dataset or time series table,
# pass a list to the subfield argument:
define_extract_nhgis(
  description = "Extract with multiple time series tables",
  time_series_tables = c("CW3", "CW5"),
  tst_geog_levels = list("state", "county")
)

}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract_micro}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
