% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{define_extract_nhgis}
\alias{define_extract_nhgis}
\title{Define an IPUMS NHGIS extract request}
\usage{
define_extract_nhgis(
  description = "",
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  breakdown_and_data_type_layout = NULL,
  tst_layout = NULL,
  data_format = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{datasets}{Vector of
\href{https://www.nhgis.org/overview-nhgis-datasets}{datasets} to include in
the extract request.}

\item{data_tables}{Summary tables to retrieve for each of the requested
\code{datasets}.

If a vector, values are applied to all \code{datasets} in the
extract definition. If a named list, values are applied
to the \code{datasets} indicated by name in the list. See details.

Required if an extract definition includes any \code{datasets}.}

\item{time_series_tables}{Vector of
\href{https://www.nhgis.org/time-series-tables}{time series tables}
to include in the extract request.}

\item{geog_levels}{Geographic levels (for example, \code{"county"} or \code{"state"})
at which to obtain data for the requested \code{datasets} or
\code{time_series_tables}.

If a vector, values are applied to all \code{datasets} and \code{time_series_tables}
in the extract definition. If a named list, values are applied
to the \code{datasets} and/or \code{time_series_tables} indicated by name in the
list. See details.

Required if an extract definition includes any \code{datasets} or
\code{time_series_tables}.}

\item{years}{Years for which to obtain the data contained in the requested
\code{datasets}. Use \code{"*"} to select all available years for a given dataset.

If a vector, values are applied to all \code{datasets}
in the extract definition. If a named list, values are applied
to the \code{datasets} indicated by name in the list. See details.

Use \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a dataset allows year selection.}

\item{breakdown_values}{\href{https://www.nhgis.org/frequently-asked-questions-faq#breakdowns}{breakdown values}
to apply to the requested \code{data_tables}.

If a vector, values are applied to all \code{datasets}
in the extract definition. If a named list, values are applied
to the \code{datasets} indicated by name in the list. See details.

If more than one breakdown value is requested,
\code{breakdown_and_data_type_layout} must also be specified.}

\item{geographic_extents}{Vector of geographic extents to use for
all of the \code{datasets} in the extract definition (for instance, to obtain
data within a particular state). Use \code{"*"} to select all available extents.

Required when the extract definition includes any \code{geog_levels} that
require extent selection. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to determine if a
geographic level requires extent selection.

Currently, the NHGIS supports extent selection only for blocks and block
groups.}

\item{shapefiles}{Vector of
\href{https://www.nhgis.org/gis-files}{shapefiles} to include in the extract
request.}

\item{breakdown_and_data_type_layout}{The desired layout
of any \code{datasets} that have multiple data types or breakdown values.
\itemize{
\item \code{"single_file"} (default) keeps all data types and breakdown values in
one file.
\item \code{"separate_files"} splits each data type or breakdown value into its
own file.
}

Required if any \code{datasets} included in the extract definition consist of
multiple data types (for instance, estimates and margins of error) or have
multiple breakdown values specified. See \code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}} to
determine whether a requested dataset has multiple data types.}

\item{tst_layout}{The desired layout of all \code{time_series_tables} included in
the extract definition. One of:
\itemize{
\item \code{"time_by_column_layout"} (wide format, default): rows correspond to geographic
units, columns correspond to different times in the time series
\item \code{"time_by_row_layout"} (long format): rows correspond to a single
geographic unit at a single point in time
\item \code{"time_by_file_layout"}: data for different times are provided in
separate files
}

Required when an extract definition includes any \code{time_series_tables}.}

\item{data_format}{The desired format of the extract data file. One of:
\itemize{
\item \code{"csv_no_header"} (default) includes only a minimal header in the first
row.
\item \code{"csv_header"} includes a second, more descriptive header row. Note that
this header is removed by default when reading data with \code{\link[=read_nhgis]{read_nhgis()}}.
\item \code{"fixed_width"} provides data in a fixed width format.
}

Required when an extract definition includes any \code{datasets} or
\code{time_series_tables}.}
}
\value{
An object of class \code{\link[=ipums_extract-class]{nhgis_extract}} containing
the extract definition.
}
\description{
Define an IPUMS NHGIS request to be submitted via the IPUMS
extract API. An extract request contains the specifications required
to identify and obtain a particular set of data from the IPUMS NHGIS system.

See details for more information about creating valid IPUMS NHGIS extract
definitions. Codes for particular data sources can be found using
\code{\link[=get_nhgis_metadata]{get_nhgis_metadata()}}.

More information about IPUMS NHGIS can be found at the following links:
\itemize{
\item \href{https://www.nhgis.org/data-availability}{NHGIS data}
\item \href{https://www.nhgis.org/frequently-asked-questions-faq}{NHGIS FAQ}
\item \href{https://developer.ipums.org/docs/workflows/create_extracts/nhgis_data/}{NHGIS API}
}
}
\details{
An NHGIS extract definition must include at least one dataset, time series
table, or shapefile. For extract definitions that include any \code{datasets},
corresponding \code{data_tables} and \code{geog_levels} must be specified. For
extract definitions that include any \code{time_series_tables}, corresponding
\code{geog_levels} must be specified.

NHGIS extract definitions may contain multiple \code{datasets} or
\code{time_series_tables.} Each dataset or time series table is associated with
several subfields that apply only to that particular dataset or time series
table.

The following arguments are applied to the requested \code{datasets}:
\itemize{
\item \code{data_tables}
\item \code{years}
\item \code{breakdown_values}
}

The following arguments are applied to the requested \code{datasets} \emph{and}
\code{time_series_tables}:
\itemize{
\item \code{geog_levels}
}

To apply the values of these arguments to \emph{all} of the \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a vector.

To apply the values of these arguments to \emph{specific} \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a named
list. In this case, names should correspond to the \code{datasets} and/or
\code{time_series_tables} and their associated values will only be applied to
the indicated dataset or time series table.

Note that while it is possible to match elements to \code{datasets} and/or
\code{time_series_tables} by index by providing an unnamed list, we recommend
using names to explicitly specify the values that should be applied to each
of the request's \code{datasets} and \code{time_series_tables}.

Alternatively, extract definitions for individual datasets and
time series tables can be combined using \code{\link[=combine_extracts]{combine_extracts()}}.
}
\examples{
# Extract definition for tables from an NHGIS dataset
my_extract <- define_extract_nhgis(
  description = "Example NHGIS extract",
  datasets = "1990_STF3",
  data_tables = "NP57",
  geog_levels = c("county", "tract")
)

my_extract

# Extract definition for an NHGIS time series table
define_extract_nhgis(
  description = "Example NHGIS extract",
  time_series_tables = "CL8",
  geog_levels = "county"
)

# Use get_nhgis_metadata() to identify data source names
\dontrun{
get_nhgis_metadata("time_series_tables")
}

# Fields that are attached to specific datasets/time series tables
# are recycled to all datasets/time series tables in an extract by default
define_extract_nhgis(
  description = "Extract definition with multiple datasets",
  datasets = c("2014_2018_ACS5a", "2015_2019_ACS5a"),
  data_tables = "B01001",
  geog_levels = c("state", "county")
)

# To instead attach specific values to each dataset or time series table,
# use a named list instead of a vector:
define_extract_nhgis(
  description = "Extract with multiple time series tables",
  time_series_tables = c("CW3", "CW5"),
  geog_levels = list(CW5 = "county", CW3 = "state")
)

# For extracts with datasets and time series tables,
# geog_levels are matched to both sources:
define_extract_nhgis(
  description = "Extract with datasets and time series tables",
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  time_series_tables = "CL6",
  geog_levels = "state"
)

define_extract_nhgis(
  description = "Extract with datasets and time series tables",
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2"),
  time_series_tables = "CL6",
  geog_levels = list("1990_STF1" = "county", "CL6" = "state")
)

\dontrun{
# Use the extract definition to submit an extract request to the API
submit_extract(my_extract)
}
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_usa}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{set_ipums_api_key}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
