% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{define_extract_usa}
\alias{define_extract_usa}
\title{Define an IPUMS USA extract request}
\usage{
define_extract_usa(
  description,
  samples,
  variables,
  data_format = "fixed_width",
  data_structure = "rectangular",
  rectangular_on = NULL
)
}
\arguments{
\item{description}{Description of the extract.}

\item{samples}{Character vector of sample names to include in the extract.
Samples should be specified using IPUMS USA
\href{https://usa.ipums.org/usa-action/samples/sample_ids}{sample ID values}.}

\item{variables}{Character vector of variable names or a list of
\code{ipums_variable} objects created by \code{\link[=new_variable]{new_variable()}}
containing specifications for all variables to include in the extract.}

\item{data_format}{Format for the output extract data file. One of
\code{"fixed_width"}, \code{"csv"}, \code{"stata"}, \code{"spss"}, or \code{"sas9"}. Defaults to
\code{"fixed_width"}.}

\item{data_structure}{Data structure for the output the extract data.

\code{"rectangular"} provides person records with all requested household
information attached to respective household members.

\code{"hierarchical"} provides household records followed by person records.}

\item{rectangular_on}{If \code{data_structure} is \code{"rectangular"},
records on which to rectangularize. Currently only \code{"P"} (person records)
is supported.

Defaults to \code{"P"} if \code{data_structure} is \code{"rectangular"} and \code{NULL}
otherwise.}
}
\value{
An object of class \code{\link[=ipums_extract-class]{usa_extract}} containing
the extract definition.
}
\description{
Define the parameters of an IPUMS USA extract request to be submitted via the
IPUMS API.

Use the \code{\link[=new_variable]{new_variable()}} helper to generate detailed variable-level
specifications for an extract.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
usa_extract <- define_extract_usa(
  description = "2013-2014 ACS Data",
  samples = c("us2013a", "us2014a"),
  variables = c("SEX", "AGE", "YEAR")
)

usa_extract

# Use `new_variable()` to created detailed variable specifications:
usa_extract <- define_extract_usa(
  description = "Example USA extract definition",
  samples = c("us2013a", "us2014a"),
  variables = new_variable(
    "SEX",
    case_selections = "2",
    attached_characteristics = c("mother", "father")
  )
)

# For multiple variables, provide a list of `ipums_variable` objects and/or
# variable names:
usa_extract <- define_extract_usa(
  description = "Example USA extract definition",
  samples = c("us2013a", "us2014a"),
  variables = list(
    new_variable("AGE", data_quality_flags = TRUE),
    new_variable("SEX", case_selections = "2"),
    "RACE"
  )
)

# To recycle specifications to many variables, it may be useful to
# create variable specifications prior to defining the extract:
var_names <- c("AGE", "SEX")

var_spec <- purrr::map(
  var_names,
  ~ new_variable(.x, data_quality_flags = TRUE)
)

define_extract_usa(
  description = "Extract definition with predefined variables",
  samples = c("us2013a", "us2014a"),
  variables = var_spec
)

# Extract specifications can be indexed in the resulting `ipums_extract`
# object:
names(usa_extract$samples)

names(usa_extract$variables)

usa_extract$variables$AGE

\dontrun{
# Use the extract definition to submit an extract request to the API
submit_extract(usa_extract)
}
}
\seealso{
\code{\link[=submit_extract]{submit_extract()}}, \code{\link[=download_extract]{download_extract()}}, and \code{\link[=get_extract_info]{get_extract_info()}} to
process and manage an extract request.

\code{\link[=save_extract_as_json]{save_extract_as_json()}} and \code{\link[=define_extract_from_json]{define_extract_from_json()}} to share an
extract definition.

\code{\link[=add_to_extract.usa_extract]{add_to_extract()}},
\code{\link[=remove_from_extract.usa_extract]{remove_from_extract()}}, and
\code{\link[=combine_extracts]{combine_extracts()}} to
revise an extract definition.
}
