% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_extract}
\alias{download_extract}
\title{Download an IPUMS data extract}
\usage{
download_extract(
  extract,
  download_dir = getwd(),
  overwrite = FALSE,
  wait = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY"),
  ...
)
}
\arguments{
\item{extract}{One of:
\itemize{
\item An \code{\link[=ipums_extract-class]{ipums_extract}} object
\item The data collection and extract number formatted as a single string of the
form \code{"collection:number"}
\item The data collection and extract number formatted as a vector of the form
\code{c("collection", number)}
\item The extract number for the collection specified by the
\code{IPUMS_DEFAULT_COLLECTION} environment variable. See
\code{\link[=set_ipums_default_collection]{set_ipums_default_collection()}}
}

The extract number does not need to be zero-padded (e.g., use \code{"usa:1"}
or \code{c("usa", "1")}, not \code{"usa:00001"} or \code{c("usa", "00001")}).
See below for examples of each form.

For a list of codes used to refer to each collection, see
\code{\link[=ipums_data_collections]{ipums_data_collections()}}.}

\item{download_dir}{In what folder should the downloaded files be saved?
Defaults to current working directory.}

\item{overwrite}{Logical indicating whether to overwrite files that already
exist. Defaults to \code{FALSE}.}

\item{wait}{Logical value indicating whether to wait for the completion of
the specified \code{extract} if it is not yet ready for download. Defaults to
\code{FALSE}. See \code{\link[=wait_for_extract]{wait_for_extract()}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable \code{"IPUMS_API_KEY"}.}

\item{...}{Further arguments passed to \code{\link[=wait_for_extract]{wait_for_extract()}}. Only used if
\code{wait = TRUE}.}
}
\value{
Invisibly returns the path(s) to the files required to read the data
requested in the extract.
}
\description{
Download an IPUMS data extract via the IPUMS API.

For an overview of ipumsr
microdata API functionality, see \code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
For NHGIS extracts, data files and GIS files (shapefiles) will be saved in
separate zip archives. \code{download_extract()} will return a character vector
including the file paths to all downloaded files.

For microdata extracts, only the file path to the downloaded .xml DDI file
will be returned, as it is sufficient for loading the data provided in the
associated .gz data file.
}
\examples{
my_extract <- define_extract_usa("Example", "us2013a", "YEAR")

\dontrun{
submitted_extract <- submit_extract(my_extract)

# Download extract by supplying an ipums_extract object:
path_to_ddi_file <- download_extract(submitted_extract)

# By supplying the data collection and extract number, as a string:
path_to_ddi_file <- download_extract("usa:1")
# Note that there is no space before or after the colon, and no zero-padding
# of the extract number.

# By supplying the data collection and extract number, as a vector:
path_to_ddi_file <- download_extract(c("usa", "1"))

# If you have a default collection, you can use the extract number alone:
set_ipums_default_collection("usa")
download_extract(1)
}

}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract.cps_extract}()},
\code{\link{add_to_extract.usa_extract}()},
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{define_extract}},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract.cps_extract}()},
\code{\link{remove_from_extract.usa_extract}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
