% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_download_extract.R
\name{download_extract}
\alias{download_extract}
\title{Download an IPUMS data extract}
\usage{
download_extract(
  extract,
  download_dir = getwd(),
  overwrite = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{extract}{One of:
  \itemize{
    \item{An object inheriting from class \code{ipums_extract}}
    \item{The data collection and extract number formatted as a single
          string of the form \code{"collection:number"}}
    \item{The data collection and extract number formatted as a vector of the
          form \code{c("collection", "number")}}
  }
The extract number does not need to be zero-padded (e.g., use \code{"usa:1"}
or \code{c("usa", "1")}, not \code{"usa:00001"} or \code{c("usa", "00001")}).
See Examples section below for examples of each form.}

\item{download_dir}{In what folder should the downloaded files be saved?
Defaults to current working directory.}

\item{overwrite}{Logical indicating whether to overwrite files that already
exist. Defaults to \code{FALSE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
Invisibly returns the path(s) to the files required to read the data
  requested in the extract.
}
\description{
Download a completed IPUMS data extract via the IPUMS extract API.
}
\details{
For NHGIS extracts, data files and GIS files will be saved in separate zip
archives. \code{download_extract()} will return a character vector including
the file paths to all downloaded files.

For microdata extracts, only the file path to the downloaded .xml DDI file
will be returned, as it is sufficient for loading the data provided in the
associated .gz data file.

For an overview of ipumsr
API functionality, see \code{vignette("ipums-api", package = "ipumsr")}.
}
\examples{
my_extract <- define_extract_micro("usa", "Example", "us2013a", "YEAR")

\dontrun{
submitted_extract <- submit_extract(my_extract)

# Download extract by supplying extract object:
path_to_ddi_file <- download_extract(submitted_extract)

# By supplying the data collection and extract number, as a string:
path_to_ddi_file <- download_extract("usa:1")
# Note that there is no space before or after the colon, and no zero-padding
# of the extract number.

# By supplying the data collection and extract number, as a vector:
path_to_ddi_file <- download_extract(c("usa", "1"))
}

}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract}},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
