% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_extract}
\alias{download_extract}
\title{Download an IPUMS data extract}
\usage{
download_extract(
  extract,
  download_dir = getwd(),
  overwrite = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{extract}{Either an object of class \code{ipums_extract}; the data
collection and extract number of an existing extract formatted as a single
string (\code{"collection: number"}) or a length two vector
(\code{c("collection", "number")}); or a \code{data.frame} in which each
row contains an extract definition..}

\item{download_dir}{In what folder should the downloaded files be saved?
Defaults to current working directory.}

\item{overwrite}{Logical indicating whether to overwrite files that already
exist. Defaults to \code{FALSE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
Invisibly, the path to the downloaded .xml DDI file.
}
\description{
Download an IPUMS data extract
}
\seealso{
Other ipums_api: 
\code{\link{define_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
