% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{extract_list_to_tbl}
\alias{extract_list_to_tbl}
\title{Convert a list of extract definitions to a tibble}
\usage{
extract_list_to_tbl(extract_list)
}
\arguments{
\item{extract_list}{A list of \code{ipums_extract} objects.}
}
\value{
A \code{\link[tibble]{tbl_df}} with number of rows equal to the
  length of \code{extract_list}, in which each rows contains the definition
  of one extract.
}
\description{
Convert a list of \code{ipums_extract} objects to a
\code{\link[tibble]{tbl_df}} in which each row contains the definition of one
extract.
}
\seealso{
Other ipums_api: 
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
