% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_extract_info.R
\name{extract_tbl_to_list}
\alias{extract_tbl_to_list}
\alias{extract_list_to_tbl}
\title{Convert recent extract definitions from tibble to list format}
\usage{
extract_tbl_to_list(extract_tbl, validate = TRUE)

extract_list_to_tbl(extract_list)
}
\arguments{
\item{extract_tbl}{A \code{\link[tibble:tbl_df-class]{tibble}} (or
\code{\link[base:data.frame]{data.frame}}) containing the specifications for one
or more \code{\link[=ipums_extract-class]{ipums_extract}} objects.

Use \code{\link[=get_extract_history]{get_extract_history()}} with \code{as_table = TRUE} to produce such an object.}

\item{validate}{Logical value indicating whether to
check that each of the output \code{ipums_extract}
objects contains a valid and complete extract
definition. Defaults to \code{TRUE}.}

\item{extract_list}{A list of \code{\link[=ipums_extract-class]{ipums_extract}} objects
or a single \code{ipums_extract} object.}
}
\value{
For \code{\link[=extract_tbl_to_list]{extract_tbl_to_list()}}, a list of length equal to the number of
extracts represented in \code{extract_tbl}.

For \code{\link[=extract_list_to_tbl]{extract_list_to_tbl()}}, a \code{\link[tibble:tbl_df-class]{tibble}}
representing the specifications for each of the extract requests
represented in \code{extract_list}.
}
\description{
Convert a \code{\link[tibble:tbl_df-class]{tibble}} of extract definition
specifications to a list of \code{\link[=ipums_extract-class]{ipums_extract}} objects
or vice versa.

Use \code{\link[=get_extract_history]{get_extract_history()}} to obtain definitions of recently-submitted
extracts.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
\dontrun{
# Get tibble of recent extracts
tbl_of_last_10_extracts <- get_extract_history("usa", as_table = TRUE)

# Filter down to extracts with "income" in the description
description_mentions_income <- grepl(
  "[Ii]ncome",
  tbl_of_last_10_extracts$description
)

income_extracts <- tbl_of_last_10_extracts[description_mentions_income, ]

# Convert tibble of extracts to list of extracts
income_extracts <- extract_tbl_to_list(income_extracts)

# Now it's easier to operate on those elements as extract objects:
revised_income_extract <- add_to_extract(
  income_extracts[[1]],
  samples = "us2018a"
)

submitted_revised_income_extract <- submit_extract(revised_income_extract)

# Get list of recent extracts
list_of_last_10_extracts <- get_extract_history("usa")

# Print the extract number for extracts that are downloadable:
for (extract in list_of_last_10_extracts) {
  if (is_extract_ready(extract)) {
    print(extract$number)
  }
}

# Convert list of extracts to tibble of extracts to view in a tabular format
extract_list_to_tbl(list_of_last_10_extracts)
}
}
