% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-extract-info.R
\name{extract_tbl_to_list}
\alias{extract_tbl_to_list}
\alias{extract_list_to_tbl}
\title{Convert recent extract definitions from tibble to list format}
\usage{
extract_tbl_to_list(extract_tbl, validate = TRUE)

extract_list_to_tbl(extract_list)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are now defunct. Use \code{\link[=get_extract_history]{get_extract_history()}} to obtain
extract history as a list.
}
\keyword{internal}
