% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{extract_tbl_to_list}
\alias{extract_tbl_to_list}
\title{Convert a tibble of extract definitions to a list}
\usage{
extract_tbl_to_list(extract_tbl, validate = TRUE)
}
\arguments{
\item{extract_tbl}{A \code{\link[tibble]{tbl_df}} (or \code{data.frame})
where each row contains the definition of one extract.}

\item{validate}{Logical (\code{TRUE} or \code{FALSE}) value indicating
whether to check that each row of \code{extract_tbl} contains a valid and
complete extract definition. Defaults to \code{TRUE}}
}
\value{
A list of length equal to the number of rows of \code{extract_tbl}.
}
\description{
Convert a \code{\link[tibble]{tbl_df}} (or \code{data.frame}) of extract
definitions, such as that returned by
\code{\link{get_recent_extracts_info_tbl}}, to a list of \code{ipums_extract}
objects.
}
\seealso{
Other ipums_api: 
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
