% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_extract_info.R
\name{get_extract_history}
\alias{get_extract_history}
\title{Browse definitions of previously-submitted extract requests}
\usage{
get_extract_history(
  collection = NULL,
  how_many = 10,
  table = FALSE,
  sleep = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{collection}{Character string of the IPUMS collection for which to
retrieve recent extract requests.

If \code{NULL}, uses the current default
collection, if it exists. See \code{\link[=set_ipums_default_collection]{set_ipums_default_collection()}}.}

\item{how_many}{The number of extract requests for which to retrieve
information. Defaults to the 10 most recent extracts.}

\item{table}{Logical value indicating whether to return recent extract
information as a \code{\link[tibble:tbl_df-class]{tibble}} (\code{TRUE}) or a list of
\code{\link[=ipums_extract-class]{ipums_extract}} objects (\code{FALSE}). Defaults to
\code{FALSE}.}

\item{sleep}{Logical indicating whether to add a brief delay between
API calls when retrieving extract history.

This is highly unlikely to be needed. However, if you cannot
retrieve your entire extract history without hitting the API rate limit,
you may want to set this to \code{TRUE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[=set_ipums_api_key]{set_ipums_api_key()}}.}
}
\value{
If \code{table = FALSE}, a list of \code{\link[=ipums_extract-class]{ipums_extract}} objects.
If \code{table = TRUE}, a \code{\link[tibble:tbl_df-class]{tibble}} containing the
extract definitions.
}
\description{
Retrieve definitions of an arbitrary number of previously submitted extract
requests for a given IPUMS collection, starting from the most recent
extract request.

To check the status of a particular extract request, use
\code{\link[=get_extract_info]{get_extract_info()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
\dontrun{
# Get information for most recent extract requests
get_extract_history("usa")

# Adjust the number to return and output format
get_extract_history("cps", how_many = 3, table = TRUE)

# Setting `how_many` to `Inf` will retrieve your entire history.
# You may want to set `table = TRUE` to consolidate output depending on
# the size of your extract history.
get_extract_history("nhgis", how_many = Inf, table = TRUE)

# To get the most recent extract (for instance, if you have forgotten its
# extract number), use `get_last_extract_info()`
get_last_extract_info("nhgis")
}
}
\seealso{
\code{\link[=get_extract_info]{get_extract_info()}} to get the current status of a specific extract request.

\code{\link[=add_to_extract]{add_to_extract()}}, \code{\link[=remove_from_extract]{remove_from_extract()}} and \code{\link[=combine_extracts]{combine_extracts()}} to
revise an extract definition.

\code{\link[=extract_tbl_to_list]{extract_tbl_to_list()}} and \code{\link[=extract_list_to_tbl]{extract_list_to_tbl()}} to manipulate information
about recent extract requests.
}
