% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_extract_info}
\alias{get_extract_info}
\alias{get_last_extract_info}
\title{Get information about previous extract requests}
\usage{
get_extract_info(
  extract = NULL,
  how_many = NULL,
  table = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

get_last_extract_info(collection = NULL, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{extract}{One of:
\itemize{
\item An \code{\link[=ipums_extract-class]{ipums_extract}} object
\item An IPUMS collection with API support
\item The data collection and extract number formatted as a string of the
form \code{"collection:number"} or as a vector of the form
\code{c("collection", number)}
\item The extract number for your default IPUMS collection. See
\code{\link[=set_ipums_default_collection]{set_ipums_default_collection()}}
}

Extract numbers do not need to be zero-padded. That is, use \code{1}, not
\code{"0001"}.

For a list of codes used to refer to each collection, see
\code{\link[=ipums_data_collections]{ipums_data_collections()}}.}

\item{how_many}{If the provided \code{extract} is an IPUMS collection, the number
of recent extract requests for which to retrieve information. Defaults to
10.}

\item{table}{If the provided \code{extract} is an IPUMS collection, a logical
indicating whether to return recent extract information as a
\code{\link[tibble:tbl_df-class]{tibble}} (\code{TRUE}) or a list of
\code{\link[=ipums_extract-class]{ipums_extract}} objects (\code{FALSE}). Defaults to
\code{FALSE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable \code{"IPUMS_API_KEY"}.}

\item{collection}{Character string of the IPUMS collection for which to
retrieve the most recent extract request.}
}
\value{
If obtaining information about a single extract, an
\code{\link[=ipums_extract-class]{ipums_extract}} object.

If obtaining information for a collection, a list of
\code{\link[=ipums_extract-class]{ipums_extract}} objects (if \code{table = FALSE}) or
a \code{\link[tibble:tbl_df-class]{tibble}} (if \code{table = TRUE}).
}
\description{
Retrieve the specifications of your previously submitted extract requests.

If a specific extract number is indicated, retrieve the most recent
status of that extract request. If a collection is indicated, retrieve the
information for the most recent extract requests for that collection.

\code{get_last_extract_info()} is a convenience function to retrieve the
information for the most recent extract of a given collection.

For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\examples{
my_extract <- define_extract_usa("Example", "us2013a", "YEAR")

\dontrun{
submitted_extract <- submit_extract(my_extract)

# Get info by supplying an ipums_extract object:
get_extract_info(submitted_extract)

# Get info by supplying the data collection and extract number, as a string:
get_extract_info("usa:1")
# Note that there is no space before or after the colon, and no zero-padding
# of the extract number.

# Get info by supplying the data collection and extract number, as a vector:
get_extract_info(c("usa", "1"))

# If you have a default collection, you can use the extract number alone:
set_ipums_default_collection("usa")
get_extract_info(1)

# The default collection will automatically be used even if no extract
# number is supplied:
get_extract_info(how_many = 3, table = TRUE)
get_last_extract_info()
}
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
