% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_extract_info}
\alias{get_extract_info}
\title{Get information about a submitted extract via the IPUMS API}
\usage{
get_extract_info(extract, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{extract}{Either an object of class \code{ipums_extract}; the data
collection and extract number of an existing extract formatted as a single
string (\code{"collection: number"}) or a length two vector
(\code{c("collection", "number")}); or a \code{data.frame} in which each
row contains an extract definition..}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
An \code{ipums_extract} object.
}
\description{
Get information about a submitted extract via the IPUMS API
}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
