% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_metadata_nhgis.R
\name{get_nhgis_metadata}
\alias{get_nhgis_metadata}
\title{View NHGIS metadata}
\usage{
get_nhgis_metadata(
  type = NULL,
  dataset = NULL,
  ds_table = NULL,
  time_series_table = NULL,
  keywords = NULL,
  match_all = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{type}{If provided, one of \code{"datasets"}, \code{"data_tables"},
\code{"time_series_tables"}, or \code{"shapefiles"} indicating the desired
summary metadata.}

\item{dataset}{Character indicating the name of the dataset for
which to obtain metadata. Required if \code{ds_table} is specified.}

\item{ds_table}{Character indicating the name of the summary
table for which to obtain metadata. The summary table must be present in
the provided \code{dataset}.}

\item{time_series_table}{Character indicating the name of the
time series table for which to obtain metadata.}

\item{keywords}{Optional vector of strings used to subset requested summary
metadata. Only records for those entities whose description field contains
the provided keywords will be included in the output. Keywords are
interpreted as regular expressions and are case-insensitive.}

\item{match_all}{If TRUE, metadata records whose description contains
all of the provided `keywords` will be included in the output. If FALSE,
any metadata records whose description contains at least one of the
provided `keywords` will be included in the output.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
If obtaining summary metadata, a \code{\link[tibble]{tibble}}
  containing metadata for the specified data type. If obtaining metadata
  for a single dataset, summary table, or time series table, a named list
  containing metadata for that data source.
}
\description{
NHGIS metadata is provided for individual datasets, summary tables, and
time series tables as well as in summary form for all available datasets,
time series tables, and shapefiles.

Use the \code{type} argument to obtain summary metadata for all datasets,
time series tables, or shapefiles.

Use the \code{dataset}, \code{ds_table}, or \code{time_series_table}
arguments to get detailed metadata for a specific dataset, summary table,
or time series table, respectively. Only one of \code{type}, \code{dataset},
or \code{time_series_table} may be provided at a time.

For more information on the NHGIS metadata API, click
\href{https://developer.ipums.org/docs/workflows/explore_metadata/nhgis/}{here}
}
\examples{
get_nhgis_metadata("datasets")
get_nhgis_metadata("time_series_tables")
get_nhgis_metadata("shapefiles")

get_nhgis_metadata("data_tables", keywords = c("Sex", "Age", "Race"))

# Metadata for a particular data source
get_nhgis_metadata(dataset = "1980_STF1")
get_nhgis_metadata(dataset = "1980_STF1", ds_table = "NT1A")
get_nhgis_metadata(time_series_table = "CW3")
}
\concept{ipums_api_nhgis}
