% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_extract_info.R
\name{get_recent_extracts_info}
\alias{get_recent_extracts_info}
\alias{get_extract_history}
\alias{get_last_extract_info}
\title{Retrieve information about past extract requests}
\usage{
get_recent_extracts_info(
  collection = NULL,
  how_many = 10,
  table = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

get_extract_history(
  collection = NULL,
  sleep = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

get_last_extract_info(collection = NULL, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{collection}{Character string of the IPUMS collection for which to
retrieve recent extract requests.

If \code{NULL}, uses the current default
collection, if it exists. See \code{\link[=set_ipums_default_collection]{set_ipums_default_collection()}}.}

\item{how_many}{The number of recent extract requests for which to retrieve
information.}

\item{table}{Logical value indicating whether to return recent extract
information as a \code{\link[tibble:tbl_df-class]{tibble}} (\code{TRUE}) or a list of
\code{\link[=ipums_extract-class]{ipums_extract}} objects (\code{FALSE}). Defaults to
\code{FALSE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[=set_ipums_api_key]{set_ipums_api_key()}}.}

\item{sleep}{Logical indicating whether to add a brief delay between
API calls when retrieving extract history.

In general, this should not be needed. However, if you cannot
retrieve your entire extract history without hitting the API rate limit,
you may want to set this to \code{TRUE}.}
}
\value{
If \code{table = FALSE}, a list of \code{\link[=ipums_extract-class]{ipums_extract}} objects.
If \code{table = TRUE}, a \code{\link[tibble:tbl_df-class]{tibble}} containing the
extract definitions.

\code{get_extract_history()} always returns a \code{tibble} of extract definitions.

\code{get_last_extract_info()} returns a single \code{ipums_extract} object.
}
\description{
Browse definitions of previously-submitted extract requests.

\code{get_recent_extracts_info()} retrieves definitions for your most recently
submitted extract requests, while \code{get_extract_history()} retrieves
definitions for all of your extract requests for a given collection.

\code{get_last_extract_info()} is a convenience function to retrieve the most
recent extract for a given collection.

To check the status of a particular extract request, use
\code{\link[=get_extract_info]{get_extract_info()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
\dontrun{
# Get information for most recent extract requests
get_recent_extracts_info("usa")

# Adjust the number to return and output format
get_recent_extracts_info("cps", how_many = 3, table = TRUE)

# Get information on all past extract requests
get_extract_history("nhgis")

# To get the most recent extract (for instance, if you have forgotten its
# extract number), use `get_last_extract_info()`
get_last_extract_info("nhgis")
}
}
\seealso{
\code{\link[=get_extract_info]{get_extract_info()}} to get the current status of a specific extract request.

\code{\link[=add_to_extract]{add_to_extract()}}, \code{\link[=remove_from_extract]{remove_from_extract()}} and \code{\link[=combine_extracts]{combine_extracts()}} to
revise an extract definition.

\code{\link[=extract_tbl_to_list]{extract_tbl_to_list()}} and \code{\link[=extract_list_to_tbl]{extract_list_to_tbl()}} to manipulate information
about recent extract requests.
}
