% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec-extract-info.R
\name{get_recent_extracts_info_list}
\alias{get_recent_extracts_info_list}
\alias{get_recent_extracts_info_tbl}
\title{Get information on recent extracts}
\usage{
get_recent_extracts_info_list(
  collection = NULL,
  how_many = 10,
  api_key = Sys.getenv("IPUMS_API_KEY")
)

get_recent_extracts_info_tbl(
  collection = NULL,
  how_many = 10,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Functionality for accessing recent extract information has been moved to
\code{\link[=get_extract_history]{get_extract_history()}}. Please use that function instead.

Additionally, support for tabular extract history has been deprecated.
To identify past extracts that meet certain criteria, iterate through the
list provided by \code{get_extract_history()}.
}
\keyword{internal}
