% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_parse.r
\name{ip_read_ddi}
\alias{ip_read_ddi}
\title{Read metadata about an IPUMS extract}
\usage{
ip_read_ddi(ddi_file)
}
\description{
Reads the metadata about an IPUMS extract from a DDI file into R.
Includes information about variable and value labels, terms of
usage for the data and positions for the fixed-width file.
}
\details{
@param ddi_file Filepath to DDI xml file
#' @examples
\dontrun{
metadata <- ip_read_ddi("cps_00001.xml")
}
}
\seealso{
ip_read_data
}
