% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_info.r
\name{ip_var_info}
\alias{ip_var_info}
\alias{ip_var_label_long}
\alias{ip_var_label}
\alias{ip_val_labels}
\title{Display IPUMS variable information}
\usage{
ip_var_info(object, vars = NULL)

ip_var_label_long(object, var = NULL)

ip_var_label(object, var = NULL)

ip_val_labels(object, var = NULL)
}
\description{
Get IPUMS metadata information about data loaded into R. Will try to read the
metadata fron the loaded datasets, but it is more reliable to load the DDI
into a separate object and use it instead.
}
\details{
\code{ip_var_info()} loads all available variable information for one or more
variables into a data.frame. If \code{object} is a vector, it will include
the variable label, long variable label and value labels. If \code{object} is
a data.frame, it will include it for all variables (or only those specified
by vars). If it is a DDI, it will also include information used to read the
data from disk, including start/end position in the fixed-width file, implied
decimals and variable type.

\code{ip_var_label_long()} loads the variable description for a single variable.

\code{ip_var_label()} loads the short variable label for a single variable.

\code{ip_value_labels()} loads the value labels for a single variable.

Note that many R functions drop attributes that provide this information.
In order to make sure that they are available, it is best to keep a copy of the
separate from the data your are manipulating using \code{\link{read_ddi}}. Then
you can refer to the IPUMS documentation in this object.
}
