% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_rows.r
\name{ipums_bind_rows}
\alias{ipums_bind_rows}
\title{Bind rows together, but preserve labelled class attributes}
\usage{
ipums_bind_rows(..., .id = NULL)
}
\arguments{
\item{...}{Either data.frames or list of data.frames}

\item{.id}{Data frame identifier, when arguments are named (or are named lists
of data.frames), will make a new column with this name that has the original
names.}
}
\value{
A data.frame
}
\description{
Bind rows together, but preserve labelled class attributes
}
