% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{read_nhgis_codebook}
\alias{read_nhgis_codebook}
\title{Read metadata from an NHGIS codebook (.txt) file}
\usage{
read_nhgis_codebook(
  cb_file,
  file_select = NULL,
  raw = FALSE,
  data_layer = deprecated()
)
}
\arguments{
\item{cb_file}{Path to a codebook (.txt) file, a .zip
archive from an NHGIS extract, or a directory containing the codebook file.}

\item{file_select}{If \code{cb_file} is a .zip archive or directory that contains
multiple codebook files, an expression identifying the file to read.
Accepts a character string specifying the file name, a
\link[=selection_language]{tidyselect selection}, or an index position of the
file. Ignored if \code{cb_file} is the path to a single codebook file.}

\item{raw}{If \code{TRUE}, return a character vector containing the lines
of \code{cb_file} rather than an \code{ipums_ddi} object. Defaults to
\code{FALSE}.}

\item{data_layer}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{file_select}
instead.}
}
\value{
If \code{raw = FALSE}, an \code{ipums_ddi} object with information on the
variables contained in the data for the extract associated with the given
\code{cb_file}.

If \code{raw = TRUE}, a character vector with one element for each
line of the given \code{cb_file}.
}
\description{
Read the variable metadata contained in the .txt codebook file included with
NHGIS extracts into an \link{ipums_ddi} object.

Because NHGIS variable metadata do not
adhere to all the standards of microdata DDI files, some of the \code{ipums_ddi}
fields will not be populated.
}
\examples{
# Example NHGIS extract
nhgis_file <- ipums_example("nhgis0972_csv.zip")
codebook <- read_nhgis_codebook(nhgis_file)

# Summary of variables included in the extract:
codebook$var_info
}
\seealso{
Other ipums_read: 
\code{\link{ipums_list_files}()},
\code{\link{read_ipums_ddi}()},
\code{\link{read_ipums_micro_chunked}()},
\code{\link{read_ipums_micro_yield}()},
\code{\link{read_ipums_micro}()},
\code{\link{read_ipums_sf}()},
\code{\link{read_nhgis}()}
}
\concept{ipums_read}
