% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{read_nhgis_codebook}
\alias{read_nhgis_codebook}
\title{Read metadata from an NHGIS extract codebook file}
\usage{
read_nhgis_codebook(
  cb_file,
  file_select = NULL,
  raw = FALSE,
  data_layer = deprecated()
)
}
\arguments{
\item{cb_file}{Path to the codebook file to be loaded. This can be a .zip
archive as provided by the extract system or \code{\link[=download_extract]{download_extract()}},
a directory containing the codebook, or the codebook .txt file itself.}

\item{file_select}{If \code{cb_file} is a .zip archive or directory that contains
multiple codebook files, an expression identifying the file to load.
Accepts a character string specifying the file name,
\code{\link{dplyr_select_style}} conventions, or an index position of the file.
Ignored if \code{ddi_file} is the path to a single codebook file.}

\item{raw}{If \code{TRUE}, read lines of the provided \code{cb_file} instead of
summarizing variable information in an \code{ipums_ddi} object. Defaults to
\code{FALSE}.}

\item{data_layer}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{file_select}
instead.}
}
\value{
If \code{raw = FALSE}, an \code{ipums_ddi} object with information on the
variables contained in the data for the extract associated with the given
\code{cb_file}.

If \code{raw = TRUE}, a character vector with one element for each
line of the given \code{cb_file}.
}
\description{
NHGIS extracts include a .txt codebook file with metadata about
the contents of the extract. This loads the information contained in this
file into a structured format.
}
\details{
This function includes functionality that previously was contained in
\code{read_ipums_codebook()}, which supported the loading of codebooks
for both NHGIS and IPUMS Terra extracts. Support for IPUMS Terra has been
discontinued. \code{read_ipums_codebook()} has been deprecated and will be
removed in a future release.
}
\examples{
# Example NHGIS extract
nhgis_file <- ipums_example("nhgis0972_csv.zip")
codebook <- read_nhgis_codebook(nhgis_file)

# Summary of variables included in the extract:
codebook$var_info

}
