% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_info.r
\name{ipums_conditions}
\alias{ipums_conditions}
\title{Get IPUMS citation and conditions}
\usage{
ipums_conditions(object)
}
\arguments{
\item{object}{A DDI object (loaded with \code{\link{read_ipums_ddi}}), or a data.frame
with ipums metadata attached.}
}
\description{
Gets information about citation and conditions from objects with
IPUMS metadata, like a DDI or a loaded extract. Because of how some
R functions drop attributes, it is best to load a DDI object separately
from the data and get the conditions and citation from this object.
}
