% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_info.r
\name{ipums_file_info}
\alias{ipums_file_info}
\title{Get IPUMS file information}
\usage{
ipums_file_info(object)
}
\arguments{
\item{object}{A DDI object (loaded with \code{\link{read_ipums_ddi}}), or a data.frame
with ipums metadata attached.}
}
\value{
A list with the \code{ipums_project}, \code{extract_date}, \code{extract_notes},
  \code{conditions}, and \code{citation}.
}
\description{
Get IPUMS metadata information about the data file loaded into R. Will try to read
the metadata from the loaded datasets, but it is more reliable to load the DDI
into a separate object and use it instead.
}
\examples{
ddi <- read_ipums_ddi(ripums_example("cps_00006.xml"))
ipums_file_info(ddi)
}
