% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.r
\name{ipums_list_files}
\alias{ipums_list_files}
\alias{ipums_list_data}
\alias{ipums_list_shape}
\alias{ipums_list_raster}
\title{List files available for analysis in an IPUMS extract zip file}
\usage{
ipums_list_files(file, types = NULL, data_layer = NULL,
  shape_layer = NULL, raster_layer = NULL)

ipums_list_data(file, data_layer = NULL)

ipums_list_shape(file, shape_layer = NULL)

ipums_list_raster(file, raster_layer = NULL)
}
\arguments{
\item{file}{An IPUMS extract zip file}

\item{types}{One or more of "data", "shape", or "raster" indicating
what type of files to look for.}

\item{data_layer}{dplyr \code{\link[dplyr]{select}}-style notation for the data
files to look for}

\item{shape_layer}{dplyr \code{\link[dplyr]{select}}-style notation for the
shape files to look for}

\item{raster_layer}{dplyr \code{\link[dplyr]{select}}-style notation for the
raster files to look for}
}
\description{
Find which files can be loaded from an IPUMS extract zip file.
}
