% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{ipums_view}
\alias{ipums_view}
\title{View a static webpage with variable information from a IPUMS extract}
\usage{
ipums_view(x, out_file = NULL)
}
\arguments{
\item{x}{A DDI or other object with ipums attributes (such as data loaded from an extract)}

\item{out_file}{Optionally specify a location to save HTML file. NULL the default
makes a temporary file.}
}
\description{
Requires that htmltools, shiny and DT are installed.
}
