% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.r
\name{ipums_view}
\alias{ipums_view}
\title{View a static webpage with variable information from a IPUMS extract}
\usage{
ipums_view(x, out_file = NULL)
}
\arguments{
\item{out_file}{Optionally specify a location to save HTML file. NULL the default
makes a temporary file.}

\item{var_info}{Variable information from a IPUMS extract (from \code{\link{ipums_var_info}})}
}
\description{
Requires that htmltools, shiny and DT are installed.
}
