% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_url.r
\name{ipums_website}
\alias{ipums_website}
\title{Launch a browser window to the ipums website}
\usage{
ipums_website(x, var, project = NULL, launch = TRUE)
}
\arguments{
\item{x}{A DDI or empty (if specifying project)}

\item{var}{A single variable name in a character vector}

\item{project}{If not using a DDI (or object with a project attribute)
A name of an IPUMS project, one of:
  "IPUMS-USA", "IPUMS-CPS", "IPUMS-International", "IPUMS-DHS",
  "ATUS-X", "AHTUS-X", "MTUS-X", "NHIS", "Higher Ed", "NHGIS",
  or "IPUMS Terra"}

\item{launch}{If \code{TRUE}, launch the website.}
}
\description{
Takes a DDI (or other object with ipums metadata) and
a variable name, and launches a web browser to the variable's
page on the ipums.org website. You can also specify a project.
}
