% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_extract_ready}
\alias{is_extract_ready}
\title{Check if an extract is ready to download}
\usage{
is_extract_ready(extract, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{extract}{One of:
\itemize{
\item An \code{\link[=ipums_extract-class]{ipums_extract}} object
\item The data collection and extract number formatted as a string of the
form \code{"collection:number"} or as a vector of the form
\code{c("collection", number)}
\item An extract number to be associated with your default IPUMS
collection. See \code{\link[=set_ipums_default_collection]{set_ipums_default_collection()}}
}

Extract numbers do not need to be zero-padded. That is, use \code{1}, not
\code{"0001"}.

For a list of codes used to refer to each collection, see
\code{\link[=ipums_data_collections]{ipums_data_collections()}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[=set_ipums_api_key]{set_ipums_api_key()}}.}
}
\value{
A logical vector of length one.
}
\description{
Get the latest status of an in-progress extract request and determine if
it has finished processing and is ready for download.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\details{
The "status" of a submitted extract is one of \code{"queued"}, \code{"started"},
\code{"produced"}, \code{"canceled"}, \code{"failed"}, or \code{"completed"}.

To be ready to download, an extract must have a \code{"completed"} status.
However, some requests that are \code{"completed"} may still be unavailable for
download, as extracts expire and are removed from IPUMS servers after a set
period of time (72 hours for microdata collections, 2 weeks for IPUMS NHGIS).

Therefore, this function also checks the \code{"download_links"} field of the
extract request to determine if data are available for download. If an
extract has expired (that is, it has completed but its download links are
no longer available), this function will warn that the extract request
must be resubmitted.
}
\examples{
my_extract <- define_extract_usa(
  description = "2013-2014 ACS Data",
  samples = c("us2013a", "us2014a"),
  variables = c("SEX", "AGE", "YEAR")
)

\dontrun{
submitted_extract <- submit_extract(my_extract)

# Check the extract request associated with a given `ipums_extract` object
is_extract_ready(submitted_extract)

# Or by supplying the collection and extract number
is_extract_ready("usa:1")
is_extract_ready(c("usa", "1"))

# If you have a default collection, you can use the extract number alone:
set_ipums_default_collection("usa")
is_extract_ready(1)
}
}
\seealso{
\code{\link[=download_extract]{download_extract()}} to download an extract.

\code{\link[=get_extract_info]{get_extract_info()}} to check the status of an extract request.

\code{\link[=submit_extract]{submit_extract()}} to resubmit an expired extract request.
}
