% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{is_extract_ready}
\alias{is_extract_ready}
\title{Is the extract ready to download?}
\usage{
is_extract_ready(extract, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{extract}{Either an object of class \code{ipums_extract}; the data
collection and extract number of an existing extract formatted as a single
string (\code{"collection: number"}) or a length two vector
(\code{c("collection", "number")}); or a \code{data.frame} in which each
row contains an extract definition..}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
A logical vector. If the first argument is an object of class
  \code{ipums_extract} or the name of a data collection, the function will
  return a length-one logical vector. If the first
  argument is a \code{data.frame}, the function will return a logical vector
  with length equal to the number of rows of the \code{data.frame}.
}
\description{
This function uses the IPUMS API to check whether the given extract is ready
to download, returning TRUE for extracts that are ready and FALSE for those
that are not.
}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{revise_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
