% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_join.r
\name{join_failures}
\alias{join_failures}
\title{Report on observations dropped by a join}
\usage{
join_failures(join_results)
}
\arguments{
\item{join_results}{A dataset that has just been created by a shape join
(like \code{\link{ipums_shape_left_join}})}
}
\value{
returns a list of data.frames, where the first item (shape) is the observations
  dropped from the shape file and the second (data) is the observations dropped from the
  data.
}
\description{
Helper for learning which observations were dropped from a dataset because
they were not joined on.
}
