% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_helpers.r
\name{lbl}
\alias{lbl}
\title{Make a label placeholder object}
\usage{
lbl(...)
}
\arguments{
\item{...}{Either one or two arugments, possibly named .val and .lbl. If a
single unnamed value, represents the label, if 2 unnamed values, the first
is the value and the second is the label.}
}
\description{
Helper to make a placeholder for a label-value pair.
}
\seealso{
Other lbl_helpers: \code{\link{lbl_add}},
  \code{\link{lbl_collapse}}, \code{\link{lbl_na_if}},
  \code{\link{lbl_relabel}}
}
