% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbl_helpers.r
\name{lbl_clean}
\alias{lbl_clean}
\title{Clean unused labels}
\usage{
lbl_clean(x)
}
\arguments{
\item{x}{A \code{\link[haven]{labelled}} vector}
}
\value{
A haven::labelled vector
}
\description{
Remove labels that do not appear in the data.
}
\examples{
x <- haven::labelled(
  c(1, 2, 3, 1, 2, 3, 1, 2, 3),
  c(Q1 = 1, Q2 = 2, Q3 = 3, Q4= 4)
)

lbl_clean(x)

}
\seealso{
Other lbl_helpers: 
\code{\link{lbl_add}()},
\code{\link{lbl_collapse}()},
\code{\link{lbl_define}()},
\code{\link{lbl_na_if}()},
\code{\link{lbl_relabel}()},
\code{\link{lbl}()},
\code{\link{zap_ipums_attributes}()}
}
\concept{lbl_helpers}
