% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nhgis_extract_types}
\alias{nhgis_extract_types}
\title{Identify data types specified in an NHGIS extract}
\usage{
nhgis_extract_types(extract)
}
\arguments{
\item{extract}{An nhgis_extract object}
}
\description{
An NHGIS extract will request at least one of dataset, time series tables or
shapefiles.
}
