% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_read.r
\name{read_ddi}
\alias{read_ddi}
\title{Read metadata about an IPUMS extract}
\usage{
read_ddi(ddi_file)
}
\arguments{
\item{ddi_file}{Filepath to DDI xml file}
}
\description{
Reads the metadata about an IPUMS extract from a DDI file into R.
Includes information about variable and value labels, terms of
usage for the data and positions for the fixed-width file.
}
\examples{
\dontrun{
metadata <- read_ddi("cps_00001.xml")
}
}
\seealso{
Other ipums_read: \code{\link{read_ipums_micro}},
  \code{\link{read_ipums_sf}},
  \code{\link{read_nhgis_codebook}},
  \code{\link{read_nhgis}}, \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
