% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{read_ipums_codebook}
\alias{read_ipums_codebook}
\title{Read metadata from a text codebook in a NHGIS or Terra area-level extract}
\usage{
read_ipums_codebook(cb_file, data_layer = NULL)
}
\arguments{
\item{cb_file}{Filepath to the codebook (either the .zip file directly downloaded
from the webiste, or the path to the unzipped .txt file).}

\item{data_layer}{A regular expression uniquely identifying the data layer to
load. Required for reading from .zip files for extracts with multiple files.}
}
\value{
A \code{ipums_ddi} object with information on the variables included in the
  csv file of a NHGIS extract.
}
\description{
Read text formatted codebooks provided by some IPUMS extract systems such as
NHGIS and Terra Area-level extracts in a format analagous to the DDI's
available for other projects.
}
\examples{
\dontrun{
data <- read_ipums_codebook("nhgis0001_csv.zip")
}
}
\seealso{
Other ipums_metadata: \code{\link{read_ddi}}
}
