% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddi_read.r
\name{read_ipums_ddi}
\alias{read_ipums_ddi}
\title{Read metadata about an IPUMS extract from a DDI (.xml) file}
\usage{
read_ipums_ddi(
  ddi_file,
  file_select = NULL,
  lower_vars = FALSE,
  data_layer = deprecated()
)
}
\arguments{
\item{ddi_file}{Path to the DDI file to be loaded. This can be a .zip
archive, a directory containing the DDI file, or the .xml file itself.}

\item{file_select}{If \code{ddi_file} is a .zip archive or directory that contains
multiple DDI files, an expression identifying the .xml file to load.
Accepts a character string specifying the file name,
\code{\link{dplyr_select_style}} conventions, or an index position of the file.
Ignored if \code{ddi_file} is the path to a single .xml file.}

\item{lower_vars}{Logical indicating whether to convert variable names
to lowercase (default is FALSE, in line with IPUMS conventions)}

\item{data_layer}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{file_select}
instead.}
}
\value{
An \code{ipums_ddi} object with metadata information.
}
\description{
Reads the metadata about an IPUMS extract from a DDI file into R.
Includes information about variable and value labels, terms of
usage for the data and positions for the fixed-width file.
}
\examples{
# Example extract DDI
ddi_file <- ipums_example("cps_00006.xml")
ddi <- read_ipums_ddi(ddi_file)
}
\concept{ipums_metadata}
