% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_read.r
\name{read_ipums_micro}
\alias{read_ipums_micro}
\alias{read_ipums_micro_list}
\title{Read data from an IPUMS microdata extract}
\usage{
read_ipums_micro(
  ddi,
  vars = NULL,
  n_max = Inf,
  data_file = NULL,
  verbose = TRUE,
  var_attrs = c("val_labels", "var_label", "var_desc"),
  lower_vars = FALSE
)

read_ipums_micro_list(
  ddi,
  vars = NULL,
  n_max = Inf,
  data_file = NULL,
  verbose = TRUE,
  var_attrs = c("val_labels", "var_label", "var_desc"),
  lower_vars = FALSE
)
}
\arguments{
\item{ddi}{Either a path to a DDI .xml file downloaded from
\href{https://www.ipums.org/}{IPUMS}, or an
\link{ipums_ddi} object parsed by \code{\link[=read_ipums_ddi]{read_ipums_ddi()}}.}

\item{vars}{Names of variables to include in the output. Accepts a
vector of names or a \link[tidyselect:language]{selection helper}.
If \code{NULL}, includes all variables in the file.

For hierarchical data, the \code{RECTYPE} variable is always included even if
unspecified.}

\item{n_max}{The maximum number of lines to read. For
\code{read_ipums_micro_list()}, this applies before splitting records into
list components.}

\item{data_file}{Path to the data (.gz) file associated with
the provided \code{ddi} file. By default, looks for the data file in the same
directory as the DDI file. If the data file has been moved, specify
its location here.}

\item{verbose}{Logical indicating whether to print progress information
to the console.}

\item{var_attrs}{Variable attributes from the DDI to add to the columns of
the output data. Defaults to all available attributes (\code{"val_labels"},
\code{"var_label"}, and \code{"var_desc"}). See \code{\link[=set_ipums_var_attributes]{set_ipums_var_attributes()}}
for more details.}

\item{lower_vars}{If reading a DDI from a file,
a logical indicating whether to convert variable names to lowercase.
Defaults to \code{FALSE} for consistency with IPUMS conventions.

This argument will be ignored if argument \code{ddi} is
an \link{ipums_ddi} object. Use \code{\link[=read_ipums_ddi]{read_ipums_ddi()}} to convert variable
names to lowercase when reading a DDI file.

If \code{lower_vars = TRUE} and \code{vars} is specified, \code{vars} should reference the
converted column names.}
}
\value{
\code{read_ipums_micro()} returns a single
\code{\link[tibble:tbl_df-class]{tibble}} object.

\code{read_ipums_micro_list()} returns a list of \code{tibble} objects with one
entry for each record type.
}
\description{
Read a microdata dataset downloaded from the IPUMS extract system using
the information in its associated DDI codebook (.xml) file.

\code{read_ipums_micro()} and \code{read_ipums_micro_list()} differ in their handling
of extracts that contain multiple record types. See details.
}
\details{
IPUMS microdata extracts use two associated files: a DDI codebook (.xml)
file and a fixed-width data file. The DDI file contains metadata about the
associated data file which are used to parse its data correctly upon load.

Data are loaded with value label and variable label information
attached to each column, where appropriate. See
\code{\link[haven:labelled]{haven::labelled()}}.
\subsection{Data structures}{

Files from IPUMS projects that contain data for multiple types of records
(e.g. household records and person records) may be either rectangular
or hierarchical.

Rectangular data are transformed such that each row of data
represents only one type of record. For instance, each row will represent
a person record, and all household-level information for that person will
be included in the same row.

Hierarchical data have records of
different types interspersed in a single file. For instance, a household
record will be included in its own row followed by the person records
associated with that household.

Hierarchical data can be read in two different formats:
\itemize{
\item \code{read_ipums_micro()} reads data into a \code{\link[tibble:tbl_df-class]{tibble}}
where each row represents a single record, regardless of record type.
Variables that do not apply to a particular record type will be filled with
\code{NA} in rows of that record type. For instance, a person-specific variable
will be missing in all rows associated with household records.
\item \code{read_ipums_micro_list()} reads data into a list of
\code{tibble} objects, where each list element contains
only one record type. Each list element is named with its corresponding
record type.
}
}
}
\examples{
# Rectangular example file
cps_rect_ddi_file <- ipums_example("cps_00006.xml")

cps <- read_ipums_micro(cps_rect_ddi_file)
# Or read DDI separately to keep the metadata
ddi <- read_ipums_ddi(cps_rect_ddi_file)
cps <- read_ipums_micro(ddi)

# Hierarchical example file
cps_hier_ddi_file <- ipums_example("cps_00010.xml")

# Read in "long" format and you get 1 data frame
cps_long <- read_ipums_micro(cps_hier_ddi_file)
head(cps_long)

# Read in "list" format and you get a list of multiple data frames
cps_list <- read_ipums_micro_list(cps_hier_ddi_file)
head(cps_list$PERSON)
head(cps_list$HOUSEHOLD)

# Or you can use the \code{\%<-\%} operator from zeallot to unpack
c(household, person) \%<-\% read_ipums_micro_list(cps_hier_ddi_file)
head(person)
head(household)
}
\seealso{
Other ipums_read: 
\code{\link{ipums_list_files}()},
\code{\link{read_ipums_ddi}()},
\code{\link{read_ipums_micro_chunked}()},
\code{\link{read_ipums_micro_yield}()},
\code{\link{read_ipums_sf}()},
\code{\link{read_nhgis_codebook}()},
\code{\link{read_nhgis}()}
}
\concept{ipums_read}
