% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_read.r
\name{read_ipums_micro}
\alias{read_ipums_micro}
\title{Read data from an IPUMS extract}
\usage{
read_ipums_micro(ddi, vars = NULL, n_max = -1, data_file = NULL,
  verbose = TRUE)
}
\arguments{
\item{ddi}{Either a filepath to a DDI xml file downloaded from
the website, or a \code{ipums_ddi} object parsed by \code{\link{read_ddi}}}

\item{vars}{A vector of variable names using \code{\link[dplyr]{select}}-style
convetions. For hierarchical data, the rectype id variable will be added even
if it is not specified.}

\item{n_max}{The maximum number of records to load.}

\item{data_file}{Specify a directory to look for the data file.
If left empty, it will look in the same directory as the DDI file.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\description{
Reads a dataset downloaded from the IPUMS extract system.
}
\details{
For IPUMS projects with microdata, it relies on a downloaded
DDI codebook and a fixed-width file. Loads the data with
value labels (using \code{\link[haven]{labelled}} format)
and variable labels.
}
\examples{
\dontrun{
data <- read_micro("cps_00001.xml")
}
}
\seealso{
Other ipums_read: \code{\link{read_ddi}},
  \code{\link{read_nhgis_codebook}},
  \code{\link{read_nhgis}}, \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
