% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micro_read_yield.r
\docType{data}
\name{read_ipums_micro_yield}
\alias{read_ipums_micro_yield}
\alias{read_ipums_micro_list_yield}
\alias{IpumsLongYield}
\alias{IpumsListYield}
\title{Read data from an IPUMS extract (in yields)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
read_ipums_micro_yield(ddi, vars = NULL, data_file = NULL,
  verbose = TRUE, rectype_convert = NULL, var_attrs = c("val_labels",
  "var_label", "var_desc"), lower_vars = FALSE)

read_ipums_micro_list_yield(ddi, vars = NULL, data_file = NULL,
  verbose = TRUE, rectype_convert = NULL, var_attrs = c("val_labels",
  "var_label", "var_desc"), lower_vars = FALSE)

IpumsLongYield

IpumsListYield
}
\arguments{
\item{ddi}{Either a filepath to a DDI xml file downloaded from
the website, or a \code{ipums_ddi} object parsed by \code{\link{read_ipums_ddi}}}

\item{vars}{Names of variables to load. Accepts a character vector of names, or
\code{\link{dplyr_select_style}} conventions. For hierarchical data, the
rectype id variable will be added even if it is not specified.}

\item{data_file}{Specify a directory to look for the data file.
If left empty, it will look in the same directory as the DDI file.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}

\item{rectype_convert}{(Usually determined by project) A named vector
indicating a conversion from the rectype in data to DDI. Not usually
needed to be specified by the user.}

\item{var_attrs}{Variable attributes to add from the DDI, defaults to
adding all (val_labels, var_label and var_desc). See
\code{\link{set_ipums_var_attributes}} for more details.}

\item{lower_vars}{If reading a DDI from a file, a logical indicating
whether to convert variable names to lowercase (default is FALSE due
to tradition)}
}
\value{
A HipYield R6 object (See 'Details' for more information)
}
\description{
Reads a dataset downloaded from the IPUMS extract system, but does
so by returning an object that can read a group of lines at a time.
This is a more flexible way to read data in chunks than
the functions like \code{\link{read_ipums_micro_chunked}}, allowing
you to do things like reading parts of multiple files at the same time
and resetting from the beginning more easily than with the chunked
functions.
}
\details{
These functions return an IpumsYield R6 object which have the following
methods:
\itemize{
\item \code{yield(n = 10000)} A function to read the next 'yield' from the data,
  returns a `tbl_df` (or list of `tbl_df` for `hipread_list_yield()`)
  with up to n rows (it will return NULL if no rows are left, or all
  available ones if less than n are available).
\item \code{reset()} A function to reset the data so that the next yield will
  read data from the start.
\item \code{is_done()} A function that returns whether the file has been completely
  read yet or not.
\item \code{cur_pos} A property that contains the next row number that will be
   read (1-indexed).
}
}
\examples{
# An example using "long" data
long_yield <- read_ipums_micro_yield(ipums_example("cps_00006.xml"))
# Get first 10 rows
long_yield$yield(10)
# Get 20 more rows now
long_yield$yield(20)
# See what row we're on now
long_yield$cur_pos
# Reset to beginning
long_yield$reset()
# Read the whole thing in chunks and count Minnesotans
total_mn <- 0
while (!long_yield$is_done()) {
  cur_data <- long_yield$yield(1000)
  total_mn <- total_mn + sum(as_factor(cur_data$STATEFIP) == "Minnesota")
}
total_mn

# Can also read hierarchical data as list:
list_yield <- read_ipums_micro_list_yield(ipums_example("cps_00006.xml"))
list_yield$yield(10)

}
\seealso{
Other ipums_read: \code{\link{read_ipums_micro_chunked}},
  \code{\link{read_ipums_micro}},
  \code{\link{read_ipums_sf}}, \code{\link{read_nhgis}},
  \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
\concept{ipums_read}
\keyword{datasets}
