% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_read.r
\name{read_ipums_sf}
\alias{read_ipums_sf}
\title{Read boundary files from an IPUMS extract}
\usage{
read_ipums_sf(shape_file, shape_layer = NULL, bind_multiple = TRUE)
}
\arguments{
\item{shape_file}{Filepath to one or more .shp files or a .zip file from an IPUMS extract}

\item{shape_layer}{A regular expression selecting the layers you wish to load (`NULL` the default
loads all)}

\item{bind_multiple}{If \code{TRUE}, will combine multiple shape files found into
a single object.}
}
\description{
Reads the boundary files form an IPUMS extract into R.
}
\examples{
\dontrun{
boundaries <- read_ipums_sf("nhgis_00001.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ddi}},
  \code{\link{read_ipums_micro}},
  \code{\link{read_nhgis_codebook}},
  \code{\link{read_nhgis}}, \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
