% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_read.r
\name{read_ipums_sp}
\alias{read_ipums_sp}
\title{Read boundary files from an IPUMS extract as Spatial Objects (from sp package)}
\usage{
read_ipums_sp(shape_file, shape_layer = NULL, bind_multiple = TRUE)
}
\arguments{
\item{shape_file}{Filepath to one or more .shp files or a .zip file from an IPUMS extract}

\item{shape_layer}{For .zip extracts with multiple datasets, the name of the
 shape files to load. Accepts a character vector specifying the file name, or
\code{\link{dplyr_select_style}} conventions. Can load multiple shape files,
  which will be combined.}

\item{bind_multiple}{If \code{TRUE}, will combine multiple shape files found into
a single object.}
}
\description{
Reads the boundary files form an IPUMS extract into R as Spatial Objects (from sp package).
}
\examples{
\dontrun{
boundaries <- read_ipums_sp("nhgis_00001.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ipums_micro}},
  \code{\link{read_ipums_sf}}, \code{\link{read_nhgis}},
  \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
