% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgis_read.r
\name{read_nhgis}
\alias{read_nhgis}
\title{Read tabular data from an NHGIS extract}
\usage{
read_nhgis(
  data_file,
  file_select = NULL,
  var_attrs = c("val_labels", "var_label", "var_desc"),
  remove_extra_header = TRUE,
  do_file = NULL,
  file_type = NULL,
  na = NULL,
  col_names = TRUE,
  locale = NULL,
  verbose = TRUE,
  progress = NULL,
  show_col_types = NULL,
  ...,
  data_layer = deprecated()
)
}
\arguments{
\item{data_file}{Path to a data file, a .zip archive from an NHGIS
extract, or a directory containing the data file.}

\item{file_select}{If \code{data_file} is a .zip archive or directory that
contains multiple files, an expression identifying the file to load.
Accepts a character vector specifying the
file name, a \link[=selection_language]{tidyselect selection}, or an index
position. This must uniquely identify a file.}

\item{var_attrs}{Variable attributes to add from the codebook (.txt) file
included in the extract. Defaults to
all available attributes.

If the codebook is not found, data will be loaded with no variable
attributes.

See \code{\link[=set_ipums_var_attributes]{set_ipums_var_attributes()}} for more details.}

\item{remove_extra_header}{If \code{TRUE}, remove the additional descriptive
header row that is included in some NHGIS .csv files. Otherwise, the
additional header will appear in the first row of the output data frame.

The header contains similar information to that
included in the \code{"label"} attribute of each data column (if \code{var_attrs}
includes \code{"var_label"}), and is therefore not usually needed.}

\item{do_file}{For fixed-width files, path to the .do file associated with
the provided \code{data_file}. The .do file contains the specifications that
indicate how to parse the data file.

By default, looks in the same directory as \code{data_file} for a .do
file with the same name. If \code{FALSE} or if the .do file cannot be found,
the .dat file will be parsed by the values provided to \code{col_positions}
in \code{\link[readr:read_fwf]{read_fwf()}}. See details.}

\item{file_type}{One of \code{"csv"} (for csv files) or \code{"dat"} (for fixed-width
files) indicating the type of file to search for in the path provided to
\code{data_file}. If \code{NULL}, determines the file type automatically based on
the files found in \code{data_file}. Only needed if \code{data_file} contains both
.csv and .dat files.}

\item{na}{Character vector of strings to interpret as missing values.
If \code{NULL}, defaults to \code{c("", "NA")} for csv files and \code{c(".", "", "NA")}
for fixed-width files. See \code{\link[readr:read_delim]{read_csv()}}.}

\item{col_names}{If reading a .csv file, either \code{TRUE}, \code{FALSE} or a
character vector of column names.

If \code{TRUE}, use the default column names included in NHGIS .csv files. If
\code{FALSE} or a character vector, replace the default column names with the
provided names or default placeholders. Note that any columns that have
altered names will not be attached to the requested \code{var_attrs}.

Note that unlike \code{\link[readr:read_delim]{readr::read_csv()}}, the first row
of the input (which contains NHGIS default headers) will always be removed
from the data.}

\item{locale}{Controls defaults that vary from place to place. If \code{NULL},
uses a locale designed to provide appropriate defaults for NHGIS files;
altering the locale may cause problems with data parsing.

If needed, you can use \code{\link[readr:locale]{readr::locale()}} to
specify a different locale to control things like default time zone,
decimal mark, big mark, and day/month names. If you do so, specify
\code{encoding = "latin1"} to ensure files are encoded properly.}

\item{verbose}{Logical controlling whether to display output when loading
data. If \code{TRUE}, displays IPUMS conditions, a progress bar, and
guessed column types, unless \code{progress} or \code{show_col_types} is specified.}

\item{progress}{Logical indicating whether to display a progress bar when
loading data. By default, uses the value of \code{verbose}, unless
the \code{readr.show_progress} option is set to \code{FALSE}.}

\item{show_col_types}{Logical indicating whether to display guessed column
types. By default, uses the value of \code{verbose}, unless
the \code{readr.show_col_types} option is set to \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[readr:read_delim]{read_csv()}} or
\code{\link[readr:read_fwf]{read_fwf()}}.}

\item{data_layer}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please
use \code{file_select} instead.}
}
\value{
A \code{\link[tibble:tbl_df-class]{tibble}} containing the data found in
\code{data_file}
}
\description{
Read a csv or fixed-width (.dat) file downloaded from the NHGIS extract
system.

To read spatial data from an NHGIS extract, use \code{\link[=read_ipums_sf]{read_ipums_sf()}}.
}
\details{
The .do file that is included when downloading an NHGIS fixed-width
extract contains important metadata about the data file's column
positions and implicit decimals. \code{read_nhgis()} uses this information to
parse and recode the fixed-width data appropriately. Therefore,
while possible, it is not recommended to read fixed-width files without a
corresponding .do file.

If you no longer have access to the .do file, consider resubmitting the
extract that produced the data.
}
\examples{
# Example files
csv_file <- ipums_example("nhgis0972_csv.zip")
fw_file <- ipums_example("nhgis0730_fixed.zip")

# Provide the .zip archive directly to load the data inside:
read_nhgis(csv_file)

# For extracts that contain multiple files, use `file_select` to specify
# a single file to load.

# This accepts a tidyselect expression
read_nhgis(fw_file, file_select = matches("ds239"), verbose = FALSE)

# Or an index position
read_nhgis(fw_file, file_select = 2, verbose = FALSE)

# NHGIS fixed-width files use the information contained in the extract's
# .do file to parse the data correctly. If it does not exist, the parsing
# will be incorrect. If you have moved the .do file, be sure to provide
# its new file path to the `do_file` argument.
bad_parse <- read_nhgis(
  fw_file,
  file_select = 2,
  do_file = FALSE,
  verbose = FALSE
)

bad_parse$X1[1:10]

# `readr::read_csv()` arguments are accepted, though most defaults should
# be appropriate.
read_nhgis(csv_file, col_types = readr::cols(.default = "c"))
}
\seealso{
\code{\link[=read_ipums_sf]{read_ipums_sf()}} to read spatial data from an IPUMS extract.

\code{\link[=read_nhgis_codebook]{read_nhgis_codebook()}} to read metadata about an IPUMS NHGIS extract.

\code{\link[=ipums_list_files]{ipums_list_files()}} to list files in an IPUMS extract.
}
