% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgis_read.r
\name{read_nhgis}
\alias{read_nhgis}
\title{Read data from an NHGIS extract}
\usage{
read_nhgis(data_file, shape_file = NULL, data_layer = NULL,
  shape_layer = data_layer, verbose = TRUE)
}
\arguments{
\item{data_file}{Filepath to the data (either the .zip file directly downloaded
from the webiste, or the path to the unzipped .csv file).}

\item{shape_file}{(Optional) filepath to the shape files (either the .zip
file directly downloaded from the webiste, or the path to the unzipped
files).}

\item{data_layer}{A regular expression uniquely identifying the data layer to
load. Required for reading from .zip files for extracts with multiple files.}

\item{shape_layer}{(Defaults to using the same value as data_layer) A regular
expression uniquely identifying the shape layer to load. Required for
reading from .zip files for extracts with multiple files.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\value{
Either a \code{tbl_df} with only the tabular data, or if a \code{shape_file} is
  specified, a \code{\link[sf]{sf}} object with the tabular data and polygons.
}
\description{
Reads a dataset downloaded from the NHGIS extract system. Relies on csv files
(with or without the extra header row).
}
\examples{
\dontrun{
data <- read_nhgis("nhgis0001_csv.zip", "nhgis0001_shp.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ipums_micro}},
  \code{\link{read_ipums_sf}},
  \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
