% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhgis_read.r
\name{read_nhgis_codebook}
\alias{read_nhgis_codebook}
\title{Read metadata from a text codebook in a NHGIS extract}
\usage{
read_nhgis_codebook(cb_file, data_layer = NULL)
}
\arguments{
\item{cb_file}{Filepath to the codebook (either the .zip file directly downloaded
from the webiste, or the path to the unzipped .txt file).}

\item{data_layer}{A regular expression uniquely identifying the data layer to
load. Required for reading from .zip files for extracts with multiple files.}
}
\value{
A \code{ip_ddi} object with information on the variables included in the
  csv file of a NHGIS extract.
}
\description{
Reads a codebook downloaded from the NHGIS extract system and formats it
analagously to metadata read from a DDI.
}
\examples{
\dontrun{
data <- read_nhgis_codebook("nhgis0001_csv.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ddi}},
  \code{\link{read_ipums_micro}}, \code{\link{read_nhgis}},
  \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
