% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_read.r
\name{read_terra_area}
\alias{read_terra_area}
\title{Read data from an IPUMS Terra area extract}
\usage{
read_terra_area(data_file, ddi_file = NULL, shape_file = NULL,
  data_layer = NULL, shape_layer = data_layer, verbose = TRUE)
}
\arguments{
\item{data_file}{Path to the data file, which can either be the .zip file directly
downloaded from the IPUMS Terra website, or to the csv unzipped from the download.}

\item{ddi_file}{(Optional) If the download is unzipped, path to the .xml file which
provides usage and ciation information for extract.}

\item{shape_file}{(Optional) If the download is unzipped, path to the .zip or .shp file
representing the the shape file. If only the data table is needed, can be set to FALSE
to indicate not to load the shape file.}

\item{data_layer}{If data_file is an extract with multiple csvs, a regular expression indicating
which .csv data layer to load.}

\item{shape_layer}{If data_file is an extract with multiple shape files, a regular expression
indicating which shape layer to load. (Defaults to using the same as the data_layer)}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\value{
If a shape file, is found a \code{\link[sf]{sf}} object with the data and geography
  information, otherwise just a data.frame.
}
\description{
Reads a area-level dataset downloaded from the IPUMS Terra extract system.
}
\examples{
\dontrun{
data <- read_terra_area("2553_bundle.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ddi}},
  \code{\link{read_ipums_micro}},
  \code{\link{read_ipums_sf}},
  \code{\link{read_nhgis_codebook}},
  \code{\link{read_nhgis}}, \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
