% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_read.r
\name{read_terra_area}
\alias{read_terra_area}
\title{Read data from an IPUMS Terra area extract}
\usage{
read_terra_area(data_file, shape_file = NULL, data_layer = NULL,
  shape_layer = data_layer, ddi_file = NULL, cb_file = NULL,
  verbose = TRUE)
}
\arguments{
\item{data_file}{Path to the data file, which can either be the .zip file directly
downloaded from the IPUMS Terra website, or to the csv unzipped from the download.}

\item{shape_file}{(Optional) If the download is unzipped, path to the .zip or .shp file
representing the the shape file. If only the data table is needed, can be set to FALSE
to indicate not to load the shape file.}

\item{data_layer}{For .zip extracts with multiple datasets, the name of the
 data to load. Accepts a character vector specifying the file name, or
\code{\link{dplyr_select_style}} conventions. Data layer must uniquely identify
a dataset.}

\item{shape_layer}{(Defaults to using the same value as data_layer) Specification
of which shape files to load using the same semantics as \code{data_layer}. Can
load multiple shape files, which will be combined.}

\item{ddi_file}{(Optional) If the download is unzipped, path to the .xml file which
provides usage and ciation information for extract.}

\item{cb_file}{(Optional) If the download is unzipped, path to the .txt file which
provides usage and ciation information for extract.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\value{
If a shape file, is found a \code{\link[sf]{sf}} object with the data and geography
  information, otherwise just a data.frame.
}
\description{
Reads a area-level dataset downloaded from the IPUMS Terra extract system.
}
\examples{
\dontrun{
data <- read_terra_area("2553_bundle.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ipums_micro}},
  \code{\link{read_ipums_sf}}, \code{\link{read_ipums_sp}},
  \code{\link{read_nhgis}}, \code{\link{read_terra_micro}},
  \code{\link{read_terra_raster}}
}
