% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_read.r
\name{read_terra_micro}
\alias{read_terra_micro}
\title{Read data from an IPUMS Terra microdata extract}
\usage{
read_terra_micro(data_file, ddi_file = NULL, shape_file = NULL,
  data_layer = NULL, shape_layer = NULL, verbose = TRUE)
}
\arguments{
\item{data_file}{Path to the data file, which can either be the .zip file directly
downloaded from the IPUMS Terra website, or to the csv unzipped from the download.}

\item{ddi_file}{(Optional) If the download is unzipped, path to the .xml file which
provides usage and ciation information for extract.}

\item{shape_file}{(Optional) If the download is unzipped, path to the .zip or .shp file
representing the the shape file. If only the data table is needed, can be set to FALSE
to indicate not to load the shape file.}

\item{data_layer}{If data_file is an extract with multiple csvs, a regular expression indicating
which .csv data layer to load.}

\item{shape_layer}{If data_file is an extract with multiple shape files, a regular expression
indicating which shape layer to load.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\value{
If a shape file is found, a list containing a data.frame with the microdata,
  and a \code{\link[sf]{sf}} with the geography. Otherwise, just a data.frame with
  the microdata.
}
\description{
Reads a microdata dataset downloaded from the IPUMS Terra extract system.
}
\examples{
\dontrun{
data <- read_terra_micro("2553_bundle.zip")
}
}
\seealso{
Other ipums_read: \code{\link{read_ddi}},
  \code{\link{read_ipums_micro}}, \code{\link{read_nhgis}},
  \code{\link{read_terra_area}},
  \code{\link{read_terra_raster}}
}
