% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra_read.r
\name{read_terra_raster}
\alias{read_terra_raster}
\title{Read data from an IPUMS Terra raster extract}
\usage{
read_terra_raster(data_file, data_layer = NULL, verbose = TRUE)
}
\arguments{
\item{data_file}{Filepath to the data (either the .zip file directly downloaded
from the webiste, or the path to the unzipped .tiff file).}

\item{data_layer}{A regular expression uniquely identifying the data layer to
load. Required for reading from .zip files for extracts with multiple files.}

\item{verbose}{Logical, indicating whether to print progress information
to console.}
}
\value{
A \code{\link[raster]{raster}} object
}
\description{
Reads a raster dataset downloaded from the IPUMS Terra extract system.
}
\examples{
\dontrun{
data <- read_terra_raster("2552_bundle.zip", "LCDECIDOPZM2013.tiff")
}
}
\seealso{
Other ipums_read: \code{\link{read_data}},
  \code{\link{read_ddi}}, \code{\link{read_nhgis}},
  \code{\link{read_terra_area}},
  \code{\link{read_terra_micro}}
}
