% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_factor}
\alias{zap_labels}
\alias{is.labelled}
\alias{\%<-\%}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{everything}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{contains}}, \code{\link[dplyr:reexports]{ends_with}}, \code{\link[dplyr:reexports]{everything}}, \code{\link[dplyr:reexports]{matches}}, \code{\link[dplyr:reexports]{num_range}}, \code{\link[dplyr:reexports]{one_of}}, \code{\link[dplyr:reexports]{starts_with}}, \code{\link[dplyr:reexports]{starts_with}}}

  \item{haven}{\code{\link[haven]{as_factor}}, \code{\link[haven:labelled]{is.labelled}}, \code{\link[haven]{zap_labels}}}

  \item{zeallot}{\code{\link[zeallot:operator]{\%<-\%}}}
}}

