% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_factor}
\alias{reexports}
\alias{zap_labels}
\alias{reexports}
\alias{is.labelled}
\alias{reexports}
\alias{\%<-\%}
\alias{reexports}
\alias{starts_with}
\alias{reexports}
\alias{ends_with}
\alias{reexports}
\alias{contains}
\alias{reexports}
\alias{matches}
\alias{reexports}
\alias{num_range}
\alias{reexports}
\alias{starts_with}
\alias{reexports}
\alias{one_of}
\alias{reexports}
\alias{everything}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{starts_with}}, \code{\link[dplyr]{ends_with}}, \code{\link[dplyr]{contains}}, \code{\link[dplyr]{matches}}, \code{\link[dplyr]{num_range}}, \code{\link[dplyr]{starts_with}}, \code{\link[dplyr]{one_of}}, \code{\link[dplyr]{everything}}}

  \item{haven}{\code{\link[haven]{as_factor}}, \code{\link[haven]{zap_labels}}, \code{\link[haven]{is.labelled}}}

  \item{zeallot}{\code{\link[zeallot]{\%<-\%}}}
}}

