% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{remove_from_extract.cps_extract}
\alias{remove_from_extract.cps_extract}
\title{Remove values from an existing IPUMS CPS extract definition}
\usage{
\method{remove_from_extract}{cps_extract}(extract, samples = NULL, variables = NULL, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{samples}{Character vector of samples to remove from the extract.}

\item{variables}{Character vector of variables to remove from the extract.}

\item{...}{Ignored}
}
\value{
A modified \code{cps_extract} object
}
\description{
Remove existing values from an IPUMS CPS extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS CPS extract definition, see
\code{\link[=add_to_extract.cps_extract]{add_to_extract()}}. When replacing values,
it is best to first add new values using
\code{add_to_extract()} before removing values with
\code{remove_from_extract()}. This limits the possibility of producing a
temporarily invalid extract specification.

For an overview of ipumsr API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
\examples{
cps_extract <- define_extract_cps(
  description = "CPS example",
  samples = c("cps2019_03s", "cps2020_03s"),
  variables = "YEAR"
)

revised_cps_extract <- remove_from_extract(
  cps_extract,
  samples = "cps2020_03s"
)

revised_cps_extract
}
