% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{remove_from_extract.nhgis_extract}
\alias{remove_from_extract.nhgis_extract}
\title{Remove values from an existing NHGIS extract definition}
\usage{
\method{remove_from_extract}{nhgis_extract}(
  extract,
  datasets = NULL,
  data_tables = NULL,
  time_series_tables = NULL,
  geog_levels = NULL,
  years = NULL,
  breakdown_values = NULL,
  geographic_extents = NULL,
  shapefiles = NULL,
  ...
)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{datasets}{Character vector of
\href{https://www.nhgis.org/overview-nhgis-datasets}{datasets} to remove from
the extract definition. All \code{data_tables}, \code{geog_levels}, \code{years}, and
\code{breakdown_values} associated with the specified
\code{datasets} will also be removed.}

\item{data_tables}{Summary tables to remove from the \code{datasets} in the
extract definition.

If a vector, values are removed from all \code{datasets} in the extract
definition. If a named list, values are removed from the \code{datasets}
indicated by name in the list. See details.}

\item{time_series_tables}{Character vector of
\href{https://www.nhgis.org/time-series-tables}{time series tables}
to remove from the extract definition. All \code{geog_levels}
associated  with the specified \code{time_series_tables} will also be removed.}

\item{geog_levels}{Geographic levels to remove from the \code{datasets} and/or
\code{time_series_tables} in the extract definition

If a vector, values are removed from all \code{datasets} and
\code{time_series_tables} in the extract
definition. If a named list, values are removed from the \code{datasets} and/or
\code{time_series_tables} indicated by name in the list. See details.}

\item{years}{Years to remove from the \code{datasets} in the extract definition.

Uses the same logic as \code{data_tables} when determining which \code{datasets} to
modify. See details.}

\item{breakdown_values}{\href{https://www.nhgis.org/frequently-asked-questions-faq#breakdowns}{breakdown values}
to remove from the \code{datasets} in the extract definition.

Uses the same logic as \code{data_tables} when determining which datasets to
modify. See details.}

\item{geographic_extents}{Geographic extents to remove from the extract
definition.}

\item{shapefiles}{Character vector of
\href{https://www.nhgis.org/gis-files}{shapefiles} to remove from the extract
definition.}

\item{...}{Ignored}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Remove existing values from an IPUMS NHGIS extract definition. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS NHGIS extract definition, see
\code{\link[=add_to_extract.nhgis_extract]{add_to_extract()}}. When replacing values,
it is best to first add new values using
\code{add_to_extract()} before removing values with
\code{remove_from_extract()}. This limits the possibility of producing a
temporarily invalid extract specification.

In general, removing from an extract follows the same syntax conventions as
used in \code{\link[=define_extract_nhgis]{define_extract_nhgis()}}. See details for
more information on how values passed to dataset and time series table
subfields are interpreted.
}
\details{
NHGIS extract definitions may contain multiple \code{datasets} or
\code{time_series_tables.} Each dataset or time series table is associated with
several subfields that apply only to that particular dataset or time series
table.

The following arguments are removed from the extract definition's \code{datasets}:
\itemize{
\item \code{data_tables}
\item \code{years}
\item \code{breakdown_values}
}

The following arguments are removed from the extract definition's \code{datasets}
\emph{and} \code{time_series_tables}:
\itemize{
\item \code{geog_levels}
}

To remove the values of these arguments from \emph{all} of the \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a vector.

To remove the values of these arguments to \emph{specific} \code{datasets} and/or
\code{time_series_tables} in the extract definition, pass the values as a named
list. In this case, names should correspond to the \code{datasets} and/or
\code{time_series_tables} and their associated values will only be removed from
the indicated dataset or time series table.

Note that subfields are modified \emph{after} the removal of any \code{datasets}
and \code{time_series_tables} specified in \code{datasets} or
\code{time_series_tables} arguments.

Thus, while while it is possible to match
elements to \code{datasets} and/or
\code{time_series_tables} by index by providing an unnamed list, we recommend
using names to explicitly specify the values that should be removed from each
of the request's \code{datasets} and \code{time_series_tables}.

Any extract fields that are rendered irrelevant after modifying the extract
will be automatically removed. (For instance, if all \code{time_series_tables}
are removed from an extract, \code{tst_layout} will also be
removed.) Thus, it is not necessary to explicitly remove these values.

Note that it is possible to produce invalid extracts using
\code{remove_from_extract()} (for instance, an extract that includes a
time series table without associated geographic levels). This can occur if
you intend to replace the existing values for a required extract field.
If the goal is not to remove values but to replace them,
it is best to first use \code{add_to_extract()},
then \code{remove_from_extract()}, as this will avoid the possibility
of temporarily producing an invalid extract.

If the supplied extract definition comes from
a previously submitted extract request, this function will reset the
definition to an unsubmitted state.
}
\examples{
extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  data_tables = c("NP1", "NP2", "NP3"),
  geog_levels = "county"
)

# Remove a table from existing datasets:
remove_from_extract(extract, data_tables = "NP3")

extract2 <- define_extract_nhgis(
  time_series_tables = c("CW3", "CW5"),
  geog_levels = c("state", "county")
)

# Remove an entire time series table
remove_from_extract(
  extract2,
  time_series_tables = "CW3"
)

# Use a named list to remove values from specific datasets or time
# series tables:
remove_from_extract(
  extract2,
  geog_levels = list(CW5 = "state", CW3 = "county")
)

# Replace values by using `add_to_extract()` first to avoid invalid
# extract definitions
revised <- add_to_extract(extract2, geog_levels = list(CW3 = "tract"))
remove_from_extract(revised, geog_levels = list(CW3 = c("county", "state")))
}
\keyword{internal}
