% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{remove_from_extract.nhgis_extract}
\alias{remove_from_extract.nhgis_extract}
\title{Remove values from an existing NHGIS extract}
\usage{
\method{remove_from_extract}{nhgis_extract}(
  extract,
  datasets = NULL,
  ds_tables = NULL,
  ds_geog_levels = NULL,
  ds_years = NULL,
  ds_breakdown_values = NULL,
  geographic_extents = NULL,
  time_series_tables = NULL,
  tst_geog_levels = NULL,
  shapefiles = NULL,
  ...
)
}
\arguments{
\item{extract}{A \code{nhgis_extract} object. This can be a user-defined
extract (see \code{\link{define_extract_nhgis}()}) or an extract object
returned from another \code{ipumsr} API function.}

\item{datasets}{Character vector of datasets to remove from the extract.
All dataset subfields associated with these datasets will also be removed.}

\item{ds_tables}{Character vector or list of summary tables to remove from
datasets in the extract. This is a subfield of \code{datasets}: values
provided to this argument will be applied to all the datasets that exist
in the extract after removing the datasets specificed in \code{datasets}.
See details for available syntax options.}

\item{ds_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") to remove from datasets in the extract.
This is a subfield of \code{datasets}: values provided to this argument
will be applied to all the datasets that exist in the extract after
removing the datasets specified in \code{datasets}. See details for
available syntax options.}

\item{ds_years}{Character or integer vector or list of years to remove from
datasets in the extract. This is a subfield of \code{datasets}: values
provided to this argument will be applied to all the datasets that exist in
the extract after removing the datasets specified in \code{datasets}. See
details for available syntax options.}

\item{ds_breakdown_values}{Character vector or list of selected breakdown
values to remove from datasets in the extract. This is a subfield of
\code{datasets}: values provided to this argument will be applied to all
the datasets that exist in the extract after removing the datasets
specified in \code{datasets}. See details for available syntax options.}

\item{geographic_extents}{Character vector of geographic instances to use as
extents for all datasets included in the extract. Use \code{"*"}
to select all available extents. Required when any dataset in the extract
includes a \code{ds_geog_levels} value that requires extent selection.
See \code{\link{get_nhgis_metadata}()} to determine whether a requested
\code{ds_geog_level} requires extent selection. Currently, NHGIS supports
extent selection only for blocks and block groups.}

\item{time_series_tables}{Character vector of time series tables to remove
from the extract. All time series table subfields associated with these
time series tables will also be removed.}

\item{tst_geog_levels}{Character vector or list of geographic levels (for
example, "county" or "state") to remove from time series tables in the
extract. This is a subfield of \code{time_series_tables}: values
provided to this argument will be applied to all the time series tables
that exist in the extract after removing the time series tables specified
in \code{time_series_tables}. See details for available syntax options.}

\item{shapefiles}{Character vector of shapefiles to remove from the extract,
if any. For more information on NHGIS shapefiles, click here.}

\item{...}{Ignored}
}
\value{
A modified \code{nhgis_extract} object
}
\description{
Remove existing values present in extract fields of an NHGIS extract.
This can include removing entire datasets and/or time series tables (along
with their associated subfields) or removing subfield values for existing
datasets and/or time series tables.

To add new values or replace existing values in an NHGIS extract, see
\code{\link[=remove_from_extract.nhgis_extract]{remove_from_extract}()}.
When replacing values, it is recommended to first add new values using
\code{add_to_extract} before removing unwanted values with
\code{remove_from_extract} to avoid the possibility of producing an invalid
extract specification.

In general, removing from an extract follows the same syntax conventions
as used in \code{\link{define_extract_nhgis}()}. See details for more
information on how values passed to dataset and/or time series table
subfields are interpreted.
}
\details{
NHGIS extracts may contain multiple datasets or time series tables. Each
dataset or time series table is associated with several subfields that apply
only to that particular dataset or time series table. Dataset subfields are
prefixed with \code{ds_}, while time series table subfields are prefixed with
\code{tst_}.

There are three ways the values passed to these arguments can be provided:

\itemize{
  \item{If values are passed as a \strong{vector}, they will be
  applied across all datasets or time series tables present in the extract}
  \item{If values are passed as an \strong{unnamed list}, they will be
  matched to datasets or time series tables by index. That
  is, the first element will be associated with the first dataset in the
  extract, the second element with the second dataset in
  the extract, etc.}
  \item{If values are passed as a \strong{named list}, they
  will be matched to datasets or time series tables by name. Names should
  correspond to datasets or time series tables that exist in the extract
  definition (unrecognized names will be ignored).}
}

Importantly, subfields are modified after the removal of any datasets
and time series tables specified in either the \code{datasets} or
\code{time_series_tables} arguments. This can cause confusion if providing
unnamed lists to subfield arguments (i.e. removing values from datasets
or time series tables by index). It is safest to use the named list syntax
for subfields to avoid this ambiguity.

Any extract fields that are rendered irrelevant after modifying the extract
will be automatically removed. (For instance, if all time
series tables are removed from an extract, \code{tst_layout} will also be
removed.) Thus, it is not necessary to explicitly remove these values. To
replace the existing values for these fields, see \code{add_to_extract}.

Note that it is possible to produce invalid extracts using
\code{remove_from_extract} (for instance, an extract that includes a
time series table without associated geographic levels). This can occur if
you intend to replace the existing values for a required extract field.
If your goal is not simply to add or remove values, but to replace values in
an extract, it is recommended that you first use \code{add_to_extract} and
then use \code{remove_from_extract}, as this will avoid the possibility
of temporarily producing an invalid extract. Alternatively, you can set
\code{validate = FALSE} in \code{remove_from_extract} to prevent extract
validation while you make the replacement.
}
\examples{

library(dplyr)

extract <- define_extract_nhgis(
  datasets = "1990_STF1",
  ds_tables = c("NP1", "NP2", "NP3"),
  ds_geog_levels = "county",
  time_series_tables = c("CW3", "CW5"),
  tst_geog_levels = c("state", "county")
)

# Modify existing datasets in extract:
remove_from_extract(
  extract,
  ds_tables = "NP3"
)

# Remove an entire dataset
remove_from_extract(
  extract,
  datasets = "1990_STF1"
)

# Use a list to remove subfield values from specific datasets or time
# series tables. Named lists map by name, unnamed lists map by index.
remove_from_extract(
  extract,
  tst_geog_levels = list(CW5 = "state", CW3 = "county")
)
}
