% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_define_extract.R
\name{remove_from_extract}
\alias{remove_from_extract}
\title{Remove values from an existing IPUMS extract definition}
\usage{
remove_from_extract(extract, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{...}{Additional arguments specifying the extract fields and values to
remove from the extract definition.

All arguments available in a collection's \link{define_extract} function can be
passed to \code{add_to_extract()}.}
}
\value{
An object of the same class as \code{extract} containing the modified
extract definition
}
\description{
Remove values for specific fields in an existing \code{ipums_extract}
object. This function is an S3 generic whose behavior will depend on the
subclass (i.e. collection) of the extract being modified.
\itemize{
\item To remove from an \strong{IPUMS USA} extract definition, click
\link[=remove_from_extract.usa_extract]{here}
\item To remove from an \strong{IPUMS CPS} extract definition, click
\link[=remove_from_extract.cps_extract]{here}
\item To remove from an \strong{IPUMS NHGIS} extract definition, click
\link[=remove_from_extract.nhgis_extract]{here}
}

In general, for a given collection, the arguments to
\code{remove_from_extract()} are identical to those used when defining an
extract for that collection. For more about defining an extract, click
\link[=define_extract]{here}.

To add new values to an extract, see \code{\link[=add_to_extract]{add_to_extract()}}.

Learn more about the IPUMS API in \code{vignette("ipums-api")}.
}
\examples{
usa_extract <- define_extract_usa(
  description = "USA example",
  samples = c("us2013a", "us2014a"),
  variables = c("AGE", "SEX", "YEAR")
)

remove_from_extract(
  usa_extract,
  samples = "us2014a",
  variables = c("AGE", "SEX")
)

nhgis_extract <- define_extract_nhgis(
  datasets = new_dataset("1990_STF1", c("NP1", "NP2", "NP3"), "county"),
  time_series_tables = new_tst("A00", geog_levels = "county")
)

# Remove an existing dataset
remove_from_extract(nhgis_extract, datasets = "1990_STF1")
}
\seealso{
\code{\link[=add_to_extract]{add_to_extract()}} or \code{\link[=combine_extracts]{combine_extracts()}} to add values to an extract
definition.

\code{\link[=submit_extract]{submit_extract()}} and \code{\link[=download_extract]{download_extract()}} to submit and process an
extract request.

\code{\link[=define_extract_usa]{define_extract_usa()}}, \code{\link[=define_extract_cps]{define_extract_cps()}}, or \code{\link[=define_extract_nhgis]{define_extract_nhgis()}}
to create a new extract definition.
}
