% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{remove_from_extract}
\alias{remove_from_extract}
\title{Remove values from an existing IPUMS extract}
\usage{
remove_from_extract(
  extract,
  samples = NULL,
  variables = NULL,
  validate = TRUE,
  ...
)
}
\arguments{
\item{extract}{An extract object created with \code{\link{define_extract_micro}} or
returned from another ipumsr API function.}

\item{samples}{Character vector of samples to remove from the extract,
if any.}

\item{variables}{Character vector of variables to remove from the extract,
if any.}

\item{validate}{Logical value indicating whether to check the modified
extract structure for validity. Defaults to \code{TRUE}.}

\item{...}{Ignored}
}
\value{
A modified \code{ipums_extract} object
}
\description{
Remove existing values from extract fields of an \code{ipums_extract} object.

To add new values to an extract, see
\code{\link{add_to_extract}}.
}
\examples{
my_extract <- define_extract_micro(
  collection = "usa",
  description = "Example",
  samples = c("us2013a", "us2014a"),
  variables = "YEAR"
)

remove_from_extract(
  my_extract,
  samples = "us2014a"
)
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
