% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_revise_extract.R
\name{remove_from_extract.usa_extract}
\alias{remove_from_extract.usa_extract}
\title{Remove values from an existing USA extract}
\usage{
\method{remove_from_extract}{usa_extract}(
  extract,
  samples = NULL,
  variables = NULL,
  validate = TRUE,
  ...
)
}
\arguments{
\item{samples}{Character vector of samples to remove from the extract,
if any.}

\item{variables}{Character vector of variables to remove from the extract,
if any.}

\item{validate}{Logical value indicating whether to check the modified
extract structure for validity. Defaults to \code{TRUE}}
}
\value{
A modified \code{usa_extract} object
}
\description{
Remove existing values from extract fields of a USA extract.

To add new values to a USA extract, click
\link[=add_to_extract.usa_extract]{here}.
}
\examples{
my_extract <- define_extract_micro(
  collection = "usa",
  description = "Example",
  samples = c("us2013a" "us2014a"),
  variables = "YEAR"
)

remove_from_extract(
  my_extract,
  samples = "us2014a"
)
}
