% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{remove_from_extract.usa_extract}
\alias{remove_from_extract.usa_extract}
\title{Remove values from an existing IPUMS USA extract}
\usage{
\method{remove_from_extract}{usa_extract}(extract, samples = NULL, variables = NULL, ...)
}
\arguments{
\item{extract}{An \code{\link[=ipums_extract-class]{ipums_extract}} object.}

\item{samples}{Character vector of samples to remove from the extract.}

\item{variables}{Character vector of variables to remove from the extract.}

\item{...}{Ignored}
}
\value{
A modified \code{usa_extract} object
}
\description{
Remove existing values from an IPUMS USA extract. All
fields are optional, and if omitted, will be unchanged.

To add new values to an IPUMS USA extract, see
\code{\link[=add_to_extract.usa_extract]{add_to_extract()}}.

For an overview of ipumsr microdata API functionality, see
\code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
If the supplied extract definition comes from
a previously submitted extract, this function will reset the definition to an
unsubmitted state.
}
\examples{
usa_extract <- define_extract_usa(
  description = "USA example",
  samples = c("us2013a", "us2014a"),
  variables = "YEAR"
)

revised_usa_extract <- remove_from_extract(
  usa_extract,
  samples = "us2014a"
)

revised_usa_extract

}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract.cps_extract}()},
\code{\link{add_to_extract.usa_extract}()},
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{define_extract}},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract.cps_extract}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_api_key}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
