% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{revise_extract}
\alias{revise_extract}
\title{Revise an extract definition}
\usage{
revise_extract(
  extract,
  description = NULL,
  samples_to_add = NULL,
  samples_to_remove = NULL,
  vars_to_add = NULL,
  vars_to_remove = NULL,
  data_format = NULL,
  data_structure = NULL,
  rectangular_on = NULL
)
}
\arguments{
\item{extract}{An object of class \code{ipums_extract}.}

\item{description}{The modified extract description. If NULL (the default),
leave the description unchanged.}

\item{samples_to_add}{Samples to add to the extract definition, as a
character vector. If NULL (the default), no samples will be added.}

\item{samples_to_remove}{Samples to remove from the extract definition, as a
character vector. If NULL (the default), no samples will be removed.}

\item{vars_to_add}{Names of variables to add to the extract definition, as
a character vector. If NULL (the default), no variables will be added.}

\item{vars_to_remove}{Names of variables to remove from the extract
definition, as a character vector. If NULL (the default), no variables will
be removed.}

\item{data_format}{The desired data file format for the modified extract
definition. If NULL (the default), leave the data format unchanged.}

\item{data_structure}{The desired data structure ("rectangular" or
"hierarchical") for the modified extract definition. If NULL (the default),
leave the data structure unchanged.}

\item{rectangular_on}{If the modified extract will have a rectangular
data structure, on what record type should it be rectangularized? If NULL,
(the default), leave the \code{rectangular_on} field unchanged.}
}
\value{
An object of class \code{ipums_extract} containing the modified
  extract definition.
}
\description{
Revise an extract definition. If the supplied extract definition comes from
a previously submitted extract, this function will reset the definition to an
unsubmitted state.
}
\seealso{
Other ipums_api: 
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{is_extract_ready}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
