% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{ripums_example}
\alias{ripums_example}
\title{Get path to ripumstest examples}
\usage{
ripums_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
Get access to example extracts from the ripumstest `inst/extdata`
directory. This data has been put in a separate file because they
are too large for CRAN.
}
\details{
The 'ripumstest' package can be installed using the command:
\code{devtools::install_github('mnpopcenter/ripumstest')}
}
\examples{
ripums_example()
ripums_example("cps_00006.xml")
}
