% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{ripums_example}
\alias{ripums_example}
\title{Get path to ripums example datasets}
\usage{
ripums_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\value{
The filepath to an example file, or if path is empty, a vector of all
  available files.
}
\description{
Get access to example extracts.
}
\examples{
ripums_example() # Lists all available examples
ripums_example("cps_00006.xml") # Gives filepath for a cps DDI
}
