% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{ripums_example}
\alias{ripums_example}
\title{Get path to ripums example datasets}
\usage{
ripums_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\value{
The filepath to an example file, or if path is empty, a vector of all
  available files.
}
\description{
Get access to example extracts. Some extracts (such as the full
nhgis shape file and vignette examples) are too big for inclusion
on CRAN and so have been included in a spearate package called
'ripumsexamples'.
}
\details{
The 'ripumsexamples' package can be installed using the command:
\code{devtools::install_github('mnpopcenter/ripums', subdir = "ripumsexamples")}
}
\examples{
ripums_example() # Lists all available examples
ripums_example("cps_00006.xml") # Gives filepath for a cps DDI
}
