% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{save_extract_as_json}
\alias{save_extract_as_json}
\title{Save an extract definition to disk as JSON}
\usage{
save_extract_as_json(extract, file)
}
\arguments{
\item{extract}{An extract object created with \code{\link{define_extract_micro}} or
returned from another ipumsr API function.}

\item{file}{File path at which to write the JSON-formatted extract
definition.}
}
\value{
The file path where the extract definition was written, invisibly.
}
\description{
Save an extract definition to a JSON-formatted file. For an overview of
ipumsr API functionality, see \code{vignette("ipums-api", package = "ipumsr")}.
}
\details{
Note that this function only saves out the properties of an extract
  that are required to submit a new extract request, namely, the description,
  data structure, data format, samples, and variables.
}
\examples{
my_extract <- define_extract_micro("usa", "Example", "us2013a", "YEAR")

extract_json_path <- file.path(tempdir(), "usa_extract.json")
save_extract_as_json(my_extract, file = extract_json_path)

copy_of_my_extract <- define_extract_from_json(extract_json_path, "usa")

identical(my_extract, copy_of_my_extract)

}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_micro}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_last_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract_micro}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
