% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_ipums_api_key}
\alias{set_ipums_api_key}
\title{Set your IPUMS API key}
\usage{
set_ipums_api_key(api_key, save = FALSE, overwrite = FALSE, unset = FALSE)
}
\arguments{
\item{api_key}{API key associated with your user account.}

\item{save}{If \code{TRUE}, save the key for use in future
sessions by adding it to the \code{.Renviron} file in your home directory.
Defaults to \code{FALSE}.}

\item{overwrite}{If \code{TRUE}, overwrite any existing value of
\code{IPUMS_API_KEY} in the \code{.Renviron} file with the provided \code{api_key}.
Defaults to \code{FALSE}.}

\item{unset}{if \code{TRUE}, remove the existing value of \code{IPUMS_API_KEY}
from the environment and the \code{.Renviron} file in your home directory.}
}
\value{
The value of \code{api_key}, invisibly.
}
\description{
Set your IPUMS API key as the value associated with the \code{IPUMS_API_KEY}
environment variable.

The key can be stored for the duration of your session or for future
sessions. If saved for future sessions, it is added to the \code{.Renviron}
file in your home directory. If you choose to save your key to \code{.Renviron},
this function will create a backup copy of the file before modifying.

This function is modeled after the \code{census_api_key()} function
from \href{https://walker-data.com/tidycensus/}{tidycensus}.
}
\seealso{
Other ipums_api: 
\code{\link{add_to_extract.cps_extract}()},
\code{\link{add_to_extract.usa_extract}()},
\code{\link{add_to_extract}()},
\code{\link{define_extract_cps}()},
\code{\link{define_extract_from_json}()},
\code{\link{define_extract_nhgis}()},
\code{\link{define_extract_usa}()},
\code{\link{define_extract}},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{remove_from_extract.cps_extract}()},
\code{\link{remove_from_extract.usa_extract}()},
\code{\link{remove_from_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{set_ipums_default_collection}()},
\code{\link{submit_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
