% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{submit_extract}
\alias{submit_extract}
\title{Submit an extract request via the IPUMS API}
\usage{
submit_extract(extract, api_key = Sys.getenv("IPUMS_API_KEY"))
}
\arguments{
\item{extract}{An extract object created by \code{\link{define_extract}}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
An object of class \code{ipums_extract} containing the extract
  definition and newly-assigned extract number of the submitted extract.
}
\description{
Submit an extract request via the IPUMS API
}
\seealso{
Other ipums_api: 
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{wait_for_extract}()}
}
\concept{ipums_api}
