% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{wait_for_extract}
\alias{wait_for_extract}
\title{Wait for extract to finish}
\usage{
wait_for_extract(
  extract,
  initial_delay_seconds = 0,
  max_delay_seconds = 300,
  timeout_seconds = 10800,
  verbose = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{extract}{Either an object of class \code{ipums_extract}; the data
collection and extract number of an existing extract formatted as a single
string (\code{"collection: number"}) or a length two vector
(\code{c("collection", "number")}); or a \code{data.frame} in which each
row contains an extract definition..}

\item{initial_delay_seconds}{How many seconds to wait before first status
check.}

\item{max_delay_seconds}{Maximum seconds to wait between status checks. The
function doubles the wait time after each check, but will cap the wait
wait time at this maximum value (300 seconds, or 5 minutes, by default).}

\item{timeout_seconds}{Maximum total number of seconds to continue waiting
for the extract before throwing an error. Defaults to 10,800 seconds (three
hours).}

\item{verbose}{If \code{TRUE}, the default, messages will be printed at the
beginning of each wait interval with the current wait time, each time the
status of the extract is checked, and when the extract is ready to
download. Setting this argument to \code{FALSE} will silence these
messages.}

\item{api_key}{API key associated with your user account. Defaults to the
value of environment variable "IPUMS_API_KEY".}
}
\value{
An object of class \code{ipums_extract} containing the extract
  definition and the URLs from which to download extract files.
}
\description{
Wait for an extract to finish by periodically checking its status via the
IPUMS API and returning when the extract is ready to download.
}
\seealso{
Other ipums_api: 
\code{\link{define_extract_from_json}()},
\code{\link{define_extract}()},
\code{\link{download_extract}()},
\code{\link{extract_list_to_tbl}()},
\code{\link{extract_tbl_to_list}()},
\code{\link{get_extract_info}()},
\code{\link{get_recent_extracts_info}},
\code{\link{ipums_data_collections}()},
\code{\link{is_extract_ready}()},
\code{\link{revise_extract}()},
\code{\link{save_extract_as_json}()},
\code{\link{submit_extract}()}
}
\concept{ipums_api}
